### 2025-04-02 #7980 @editorialbot
[PRE REVIEW]: MultiHazard: Copula-based Joint Probability Analysis in R
**Submitting author:** @rjaneUCF (<a href="http://orcid.org/0000-0003-4194-852X">Robert Jane</a>)
**Repository:** https://github.com/rjaneUCF/MultiHazard
**Branch with paper.md** (empty if default branch):
**Version:** v1.13
