+++ joss_submission/paper.bib
@@ -27,7 +27,7 @@ journal = {Water Resources Research},
 volume = {53},
 number = {6},
 pages = {5166-5183},
-doi = {https://doi.org/10.1002/2016WR020242},
+doi = {10.1002/2016WR020242},
 year = {2017}
 }
 
@@ -55,7 +55,7 @@ journal = {Water Resources Research},
 volume = {53},
 number = {4},
 pages = {2833-2849},
-doi = {https://doi.org/10.1002/2016WR019426},
+doi = {10.1002/2016WR019426},
 year = {2017}
 }
 
@@ -79,7 +79,7 @@ journal = {Journal of the Royal Statistical Society: Series B (Statistical Metho
 volume = {66},
 number = {3},
 pages = {497-546},
-doi = {https://doi.org/10.1111/j.1467-9868.2004.02050.x},
+doi = {10.1111/j.1467-9868.2004.02050.x},
 year = {2004}
 }
 
@@ -95,7 +95,7 @@ title = {A typology of compound weather and climate events},
 journal = {Nat Rev Earth Environ}, 
 volume = {1}, 
 pages = {333–347}, 
-doi = {https://doi.org/10.1038/s43017-020-0060-z},
+doi = {10.1038/s43017-020-0060-z},
 year = {2020}
 }
 
@@ -115,7 +115,8 @@ title = {ReturnCurves: Estimation of Return Curves},
 author = {André, L., and Murphy-Barltrop, C.},
 year = {2024},
 note = {R package version 1.0},
-url = {https://CRAN.R-project.org/package=ReturnCurves}
+url = {https://CRAN.R-project.org/package=ReturnCurves},
+doi = {10.32614/cran.package.returncurves}
 }
 
 @article{Barltrop:2023,
@@ -125,7 +126,7 @@ journal = {Environmetrics},
 volume = {34},
 number = {5},
 pages = {e2797},
-doi = {https://doi.org/10.1002/env.2797},
+doi = {10.1002/env.2797},
 year = {2023}
 }
 
@@ -143,7 +144,7 @@ year = {2013}
 @article{Murphy:2024,
 author = {Murphy, C., and Tawn, J. A., and Varty Z.},
 title = {Automated Threshold Selection and Associated Inference Uncertainty for Univariate Extremes},
-journal =Technometrics},
+journal = {Technometrics},
 volume = {0},
 pages = {1–10}, 
 doi = {10.1080/00401706.2024.2421744},
@@ -156,7 +157,7 @@ title={Improving estimation for asymptotically independent bivariate extremes vi
 journal={Extremes},
 volume={27},
 number={4},
-pages={643--671},
-year={2024},
-publisher={Springer}
+pages={643-671},
+doi = {10.1007/s10687-024-00490-4},
+year={2024}
 }
