+++ R/Cooley19.R
@@ -24,31 +24,35 @@
 #' Cooley19(Data=na.omit(S20.Detrend.df[,3:4]),Migpd=s.Migpd,
 #' p.base=0.01,p.proj=0.001,PLOT=TRUE,x_lim_max_T=500,y_lim_max_T=500)
 Cooley19<-function(Data,Migpd,p.base=0.01,p.proj=0.001,u=0.95,PLOT=FALSE,x_lim_min_T=NA, x_lim_max_T=NA,y_lim_min_T=NA,y_lim_max_T=NA,x_lim_min=NA,x_lim_max=NA,y_lim_min=NA,y_lim_max=NA){
-  
+
   #Check the asymptotic properities
   Chi<-chi(Data,nq=1000)
   CHI<-round(Chi$chi[which(abs(Chi$quantile-u)==min(abs(Chi$quantile-u))),2],2)
   n.bar<-(Chi$chibar[which(abs(Chi$quantile-u)==min(abs(Chi$quantile-u)))]+1)/2
-  
+
   #Fit Kernel cumulative survival function to Data
   Fhat <- kcde(Data,gridsize=1000,tail.flag="upper.tail")
-  
+
   #Extract contour
   I.base<-contourLines(x=Fhat$eval.points[[1]], y=Fhat$eval.points[[2]],z=Fhat$estimate, levels=p.base)[[1]]
   I.base <- data.frame(I.base$x,I.base$y)
   colnames(I.base)<-c("x","y")
-  
+
   #Transform observations and contour to frechet (_F) scale
+  #print(EmpFun(x=Data[,1],r=Data[,1], mod=Migpd$models[[1]]))
+
   x_F<- -1/(log(EmpFun(x=Data[,1],r=Data[,1], mod=Migpd$models[[1]])))
   y_F <- -1/(log(EmpFun(x=Data[,2],r=Data[,2], mod=Migpd$models[[2]])))
-  z<-which(x_F==-Inf | y_F==-Inf)
-  x_F<-x_F[-z]
-  y_F<-y_F[-z]
+  #print(which(x_F==-Inf | y_F==-Inf))
+  #z<-which(x_F==-Inf | y_F==-Inf)
+  #x_F<-x_F[-z]
+  #y_F<-y_F[-z]
   I.base.x_F <- -1/(log(EmpFun(x=Data[,1], r=I.base$x, mod=Migpd$models[[1]])))
   I.base.y_F<-  -1/(log(EmpFun(x=Data[,2], r=I.base$y, mod=Migpd$models[[2]])))
   I.base_F <- data.frame(I.base.x_F,I.base.y_F)
   colnames(I.base_F)<-c("x","y")
-  
+  #print(summary(x_F))
+  #print(summary(y_F))
   #Projected isoline
   if(CHI>0){
     I.proj.x_F <- (p.base/p.proj)*I.base.x_F
@@ -61,39 +65,39 @@ Cooley19<-function(Data,Migpd,p.base=0.01,p.proj=0.001,u=0.95,PLOT=FALSE,x_lim_m
   }
   I.proj_F <- data.frame(I.proj.x_F,I.proj.y_F)
   colnames(I.proj_F)<-c("x","y")
-  
+
   #Transforming back to original scale
   I.proj.x_U<- exp(-1/I.proj.x_F)
   I.proj.y_U<- exp(-1/I.proj.y_F)
-  
+
   I.proj.x<-as.numeric(quantile(Data[,1],I.proj.x_U))
   I.proj.x[which(I.proj.x_U>Migpd$models[[1]]$mqu)]<-u2gpd(u=I.proj.x_U[which(I.proj.x_U>Migpd$models[[1]]$mqu)],
                                                            p=Migpd$models[[1]]$rate,
                                                            th=Migpd$models[[1]]$threshold,
[…]
