+++ NAMESPACE
@@ -39,6 +39,7 @@ export(NOAA_SLR)
 export(OsWL_Intensity)
 export(return_curve_diag)
 export(return_curve_est)
+export(SLR_Scenarios)
 export(Standard_Copula_Fit)
 export(Standard_Copula_Sel)
 export(Standard_Copula_Sim)
