+++ R/return_curve_diag.R
@@ -193,7 +193,7 @@ lines(ang_ind, lb_x_1, lty = 'dashed', col = 'blue',lwd=2)
 lines(ang_ind,rep(p,length(ang_ind)),type="l",lwd=3,col=2)
 legend("topleft",legend = c("True Probability","Median Estimate","95% Confidence Intervals"),lty=c('solid','solid','dashed'),col=c(2,1,"blue"),cex=1.2,lwd=2,bg="white")
 
-plot(ang_ind,med_y_1,xlab="Angle Index",ylab = "Probability",ylim=range(med_y_1,lb_y_1,ub_y_1,p),type='n',sub=paste('Decl.',colnames(data[2])),main="HT isoline",cex.lab=1.3, cex.axis=1.4,cex.main=1.7)
+plot(ang_ind,med_y_1,xlab="Angle Index",ylab = "Probability",ylim=range(med_y_1,lb_y_1,ub_y_1,p),type='n',sub=paste('Decl.',colnames(data[3])),main="HT isoline",cex.lab=1.3, cex.axis=1.4,cex.main=1.7)
 polygon(c(rev(ang_ind), ang_ind), c(rev( lb_y_1), ub_y_1), col = 'grey80', border = NA)
 lines(ang_ind,med_y_1,type="l",col=1,lwd=2)
 lines(ang_ind, ub_y_1, lty = 'dashed', col = 'blue',lwd=2)
@@ -201,7 +201,7 @@ lines(ang_ind, lb_y_1, lty = 'dashed', col = 'blue',lwd=2)
 lines(ang_ind,rep(p,length(ang_ind)),type="l",lwd=3,col=2)
 legend("topleft",legend = c("True Probability","Median Estimate","95% Confidence Intervals"),lty=c('solid','solid','dashed'),col=c(2,1,"blue"),cex=1.2,lwd=2,bg="white")
 
-plot(ang_ind,med_x_2,xlab="Angle Index",ylab = "Probability",ylim=range(med_x_2,lb_x_2,ub_x_2,p),type='n',sub=paste('Decl.',colnames(data[3])),main="WT isoline",cex.lab=1.3, cex.axis=1.4,cex.main=1.7)
+plot(ang_ind,med_x_2,xlab="Angle Index",ylab = "Probability",ylim=range(med_x_2,lb_x_2,ub_x_2,p),type='n',sub=paste('Decl.',colnames(data[2])),main="WT isoline",cex.lab=1.3, cex.axis=1.4,cex.main=1.7)
 polygon(c(rev(ang_ind), ang_ind), c(rev( lb_x_2), ub_x_2), col = 'grey80', border = NA)
 lines(ang_ind,med_x_2,type="l",col=1,lwd=2)
 lines(ang_ind, ub_x_2, lty = 'dashed', col = 'blue',lwd=2)
