### 2025-05-20 #7980 @lrnv
@davidAtance, @tnagler, @LingfengLuo0510, @adibender 👋 ! Would any of yo be willing to review this submission for JOSS ? Please tell me if you do not know what it means I can provide details.
