### 2025-06-05 #8350 @editorialbot
[REVIEW]: MultiHazard: Copula-based Joint Probability Analysis in R
**Submitting author:** @rjaneUCF (<a href="http://orcid.org/0000-0003-4194-852X">Robert Jane</a>)
**Repository:** https://github.com/rjaneUCF/MultiHazard
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.0
