### 2025-06-12 #8350 @mingzehuang
Hi @rjaneUCF, I found a typo in line 53 of your submitted paper "independent annual-non exceedance probabilities associated with each _**poopulation**_ as outlined". Could you take a look at it? Thanks!
