### 2025-06-12 #8350 @mingzehuang
Hi @rjaneUCF, I found that Figure 1 in the submitted paper has no clear label for sub-figures (a), (b), (c), (d). Also, there were multiple typos in the caption of Figure 1 (e.g., "envelopen", "mimum"). Would you mind checking them as well? Thanks!
