+++ tests/testthat/test-Imputation.R
@@ -1,7 +1,7 @@
 #Setting up test data
 
 setup_test_data <-function(){
- GW_S20<-data.frame(G_3356,G_3355[65:12336,])
+ GW_S20<-data.frame(G_3356,G_3355[65:12336,-1])
  colnames(GW_S20)<-c("Date","G3356","G3355")
  return(GW_S20)
 }
@@ -10,55 +10,55 @@ setup_test_data <-function(){
 test_that("Linear method works", {
 
   GW_S20 <- setup_test_data()
-  
+
   result <- Imputation(Data=GW_S20,Variable="G3356",
                        x_lab="Groundwater level (ft NGVD 29)",
                        y_lab="Groundwater level (ft NGVD 29)")
-  
-  # Check return type 
+
+  # Check return type
   expect_type(result, "list")
-  
+
   # Check return names
   expect_named(result, c("Data", "Model"))
-  
+
   # Check return length
   expect_equal(nrow(result$Data), nrow(GW_S20))
-  
+
   # Check that imputation changes the data
   expect_false(identical(result$Data$ValuesFilled, GW_S20$G3356))
-  
+
   # Check new column exists
-  expect_true("ValuesFilled" %in% names(result$Data)) 
-  
+  expect_true("ValuesFilled" %in% names(result$Data))
+
   # Verify imputed values are numeric
   expect_type(result$Data$ValuesFilled, "double")
-  
-  # Check that imputation reduces number of NAs 
+
+  # Check that imputation reduces number of NAs
   expect_lt(sum(is.na(result$Data$ValuesFilled)), sum(is.na(GW_S20$G3356)))
 })
 
 # Test model quality
 test_that("Linear model is reasonable", {
   GW_S20 <- setup_test_data()
-  
+
   result <- Imputation(Data=GW_S20, Variable="G3356",
                        x_lab="GW Level", y_lab="GW Level")
-  
+
   # Check that model summary exists and has expected structure
   expect_s3_class(result$Model, "summary.lm")
   expect_true("coefficients" %in% names(result$Model))
   expect_true("r.squared" %in% names(result$Model))
-  
+
   # Basic sanity check - R-squared should be between 0 and 1
   expect_gte(result$Model$r.squared, 0)
   expect_lte(result$Model$r.squared, 1)
 })
 
 
-#Test reproducibility 
+#Test reproducibility
 test_that("Function is deterministic", {
   GW_S20 <- setup_test_data()
-  
+
   # Window method should be reproducible
   result1 <- Imputation(Data=GW_S20,Variable="G3356",
                         x_lab="Groundwater level (ft NGVD 29)",
@@ -67,19 +67,19 @@ test_that("Function is deterministic", {
                         x_lab="Groundwater level (ft NGVD 29)",
                         y_lab="Groundwater level (ft NGVD 29)")
   expect_identical(result1$Data$ValuesFilled, result2$Data$ValuesFilled)
-  
+
 })
 
 # Test that only missing values are imputed
 test_that("Only missing values are imputed", {
   GW_S20 <- setup_test_data()
-  
+
   result <- Imputation(Data=GW_S20, Variable="G3356",
                        x_lab="GW Level", y_lab="GW Level")
-  
+
   # Non-missing values should remain unchanged
   non_missing_idx <- !is.na(GW_S20$G3356)
-  expect_equal(result$Data$ValuesFilled[non_missing_idx], 
+  expect_equal(result$Data$ValuesFilled[non_missing_idx],
                GW_S20$G3356[non_missing_idx])
 })
 
@@ -99,10 +99,10 @@ test_that("Handles edge cases", {
     var1 = c(1, NA, 3),
     var2 = c(2, 4, 6)
[…]
