+++ tests/testthat/test-Annual_Max.R
@@ -2,7 +2,10 @@
 
 test_that("Annual maximum functions basic functionality", {
 
-  result <- Annual_Max(Data_Detrend = S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(2,5)])
+  test_data = S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(2,4)]
+  test_data[,2] = as.Date(test_data[,2])
+
+  result <- Annual_Max(Data_Detrend = test_data)
 
   # Check return type
   expect_type(result, "list")
@@ -25,8 +28,11 @@ test_that("Annual maximum functions basic functionality", {
 #Test function with different percent complete values
 test_that("Annual maximum function with different Complete_Prop values", {
 
-  result1 <- Annual_Max(Data_Detrend = S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(2,5)], Complete_Prop = 0.9)
-  result2 <- Annual_Max(Data_Detrend = S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(2,5)], Complete_Prop = 0.5)
+  test_data = S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(2,4)]
+  test_data[,2] = as.Date(test_data[,2])
+
+  result1 <- Annual_Max(Data_Detrend = test_data, Complete_Prop = 0.9)
+  result2 <- Annual_Max(Data_Detrend = test_data, Complete_Prop = 0.5)
 
   expect_true(length(result1$Event) <= length(result2$Event))
   expect_true(length(result1$AM) <= length(result2$AM))
@@ -36,11 +42,12 @@ test_that("Annual maximum function with different Complete_Prop values", {
 # Test that invalid inputs gives errors
 test_that("Invalid inputs produce errors", {
 
-  expect_error(Annual_Max(Data_Detrend = S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(2,5)],
-                          Complete_Prop = -1))
-  expect_error(Annual_Max(Data_Detrend = S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(2,5)],
-                          Complete_Prop = 1.15))
-  expect_error(Annual_Max(Data_Detrend = S20_T_MAX_Daily_Completed_Detrend_Declustered[,5]))
+  test_data = S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(2,4)]
+  test_data[,2] = as.Date(test_data[,2])
+
+  expect_error(Annual_Max(Data_Detrend = test_data, Complete_Prop = -1))
+  expect_error(Annual_Max(Data_Detrend = test_data, Complete_Prop = 1.15))
+  expect_error(Annual_Max(Data_Detrend = test_data[,2]))
   expect_error(Annual_Max(Data_Detrend = letters[1:10]))
 })
 
@@ -48,9 +55,12 @@ test_that("Invalid inputs produce errors", {
 
 test_that("Function is deterministic", {
 
+  test_data = S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(2,4)]
+  test_data[,2] = as.Date(test_data[,2])
+
   #Find annual max
-  result1 <- Annual_Max(Data_Detrend = S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(2,5)])
-  result2 <- Annual_Max(Data_Detrend = S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(2,5)])
+  result1 <- Annual_Max(Data_Detrend = test_data)
+  result2 <- Annual_Max(Data_Detrend = test_data)
 
   #See whether outputs are the same
   expect_identical(result1$Event, result2$Event)
