+++ tests/testthat/test-Annual_Max.R
@@ -3,8 +3,8 @@
 test_that("Annual maximum functions basic functionality", {
 
   test_data = data.frame(S20_T_MAX_Daily_Completed_Detrend_Declustered$Date,S20_T_MAX_Daily_Completed_Detrend_Declustered$Detrend)
-  test_data[,1] = as.Date(test_data[,1], format = "%m/%d/%Y")
-   print(summary(test_data))
+  test_data[,1] <- as.Date(as.character(test_data[,1]), format = "%m/%d/%Y")
+
   result <- Annual_Max(Data_Detrend = test_data)
 
   # Check return type
@@ -29,7 +29,7 @@ test_that("Annual maximum functions basic functionality", {
 test_that("Annual maximum function with different Complete_Prop values", {
 
   test_data = data.frame(S20_T_MAX_Daily_Completed_Detrend_Declustered$Date,S20_T_MAX_Daily_Completed_Detrend_Declustered$Detrend)
-  test_data[,1] = as.Date(test_data[,1], format = "%m/%d/%Y")
+  test_data[,1] <- as.Date(as.character(test_data[,1]), format = "%m/%d/%Y")
 
   result1 <- Annual_Max(Data_Detrend = test_data, Complete_Prop = 0.9)
   result2 <- Annual_Max(Data_Detrend = test_data, Complete_Prop = 0.5)
@@ -43,7 +43,7 @@ test_that("Annual maximum function with different Complete_Prop values", {
 test_that("Invalid inputs produce errors", {
 
   test_data = data.frame(S20_T_MAX_Daily_Completed_Detrend_Declustered$Date,S20_T_MAX_Daily_Completed_Detrend_Declustered$Detrend)
-  test_data[,1] = as.Date(test_data[,1], format = "%m/%d/%Y")
+  test_data[,1] <- as.Date(as.character(test_data[,1]), format = "%m/%d/%Y")
 
   expect_error(Annual_Max(Data_Detrend = test_data, Complete_Prop = -1))
   expect_error(Annual_Max(Data_Detrend = test_data, Complete_Prop = 1.15))
@@ -56,7 +56,7 @@ test_that("Invalid inputs produce errors", {
 test_that("Function is deterministic", {
 
   test_data = data.frame(S20_T_MAX_Daily_Completed_Detrend_Declustered$Date,S20_T_MAX_Daily_Completed_Detrend_Declustered$Detrend)
-  test_data[,1] = as.Date(test_data[,1], format = "%m/%d/%Y")
+  test_data[,1] <- as.Date(as.character(test_data[,1]), format = "%m/%d/%Y")
 
   #Find annual max
   result1 <- Annual_Max(Data_Detrend = test_data)
