+++ tests/testthat/test-Annual_Max.R
@@ -4,7 +4,7 @@ test_that("Annual maximum functions basic functionality", {
 
   test_data = data.frame(S20_T_MAX_Daily_Completed_Detrend_Declustered$Date,S20_T_MAX_Daily_Completed_Detrend_Declustered$Detrend)
   print(summary(S20_T_MAX_Daily_Completed_Detrend_Declustered$Date))
-  test_data[,1] <- as.Date(test_data[,1], format = "%Y/%m/%d")
+  test_data[,1] <- as.Date(test_data[,1], format = "%Y-%m-%d")
   print(summary(test_data))
   print(class(test_data[,1]))
 
