+++ R/Copula_Threshold_2D_Lag.R
@@ -19,6 +19,9 @@
 #' @param Lower Numeric vector specifying the element number of the \code{u2} argument for which the copula family name label to appear below the corresponding point on the Kendall's tau coefficient vs uhold plot, when conditioning on the variable in column 2. Default is \code{NA}.
 #' @param GAP Numeric vector of length one specifying the distance above or below the copula family name label appears the corresponding point on the Kendall's tau coefficient vs uhold plot. Default is \code{0.05}.
 #' @param Legend Logic vector of length one specifying whether a legend should be plotted. Default is \code{TRUE}.
+#' @param Cex_Legend Numeric vector of length one specifying the font size of the legend. Default is \code{1}.
+#' @param Cex_Axis Numeric vector of length one specifying the font size of the axes. Default is \code{1}.
+#' @param Cex_Axis_Original Numeric vector of length one specifying the font size of the values of the quantiles on the original (data) scale (i.e. second x-axis). Default is \code{1}.
 #' @return List comprising: \itemize{
 #' \item \code{Kendalls_Tau1}
 #' Kendall's tau of a sample
@@ -31,14 +34,63 @@
 #' }
 #' when the dataset is conditioned on the variable in column 1.
 #' Analogous vector \code{Kendalls_Tau2}, \code{p_value_Var2}, \code{N_Var2} and \code{Copula_Family_Var2} for the specified thresholds when the dataset is conditioned on the variable in column 2.
+#' If \code{PLOT=TRUE} then a plot of the Kendall's tau correlation coefficient versus quantile threshold is also returned.
+#' Filled circles denote statistically significant correlation at a \code{5\%} significance level. Numbers inside the circles correspond to the sample size while the best fitting copula family is printed above.
+#' Numbers below x-axis are the values of the corresponding quantiles on the original (data) scale.
 #' @seealso \code{\link{Dataframe_Combine}}
 #' @export
 #' @examples
-#' Copula_uhold_2D(Data_Detrend=S20.Detrend.df[,-c(1,4)],
-#'                     Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
-#'                     y_lim_min=-0.075, y_lim_max =0.25,
-#'                     Upper=c(6,8), Lower=c(6,8),GAP=0.1)
-Copula_Threshold_2D_Lag<-function(Data_Detrend,Data_Declust,u1=seq(0.9,0.99,0.01),u2=seq(0.9,0.99,0.01),PLOT=TRUE,Lag_Backward_Var1=1,Lag_Forward_Var1=1,Lag_Backward_Var2=1,Lag_Forward_Var2=1,x_lim_min=NA,x_lim_max=NA,y_lim_min=-1,y_lim_max=1,Upper=NA,Lower=NA,GAP=0.05,Legend=TRUE){
+#' Copula_Threshold_2D_Lag(Data_Detrend=S20.Detrend.df[,-c(1,4)],
+#'                         Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
+#'                         y_lim_min=-0.075, y_lim_max =0.25,
+#'                         Upper=c(6,8), Lower=c(6,8),GAP=0.1)
[…]
