+++ tests/testthat/test-Copula_Threshold_2D_Lag.R
@@ -22,7 +22,7 @@ test_that("Copula_Threshold_2D_Lag method works", {
   expect_equal(length(result$N_Var2), length(seq(0.9,0.99,0.01)))
   expect_true(all(result$Copula_Family_Var1 %in% c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 13, 14, 16, 17, 18, 19, 20, 23, 24, 26, 27, 28, 29, 30, 33, 34, 36, 37, 38, 39, 40, 104, 114, 124, 134, 204, 214, 224, 234)))
   expect_true(all(result$Copula_Family_Var2 %in% c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 13, 14, 16, 17, 18, 19, 20, 23, 24, 26, 27, 28, 29, 30, 33, 34, 36, 37, 38, 39, 40, 104, 114, 124, 134, 204, 214, 224, 234)))
-  expect_true(all(result$Kendalls_Tau1 >-1 & result$Kendalls_Tau1<=1))
+  expect_true(all(result$Kendalls_Tau1 >=-1 & result$Kendalls_Tau1<=1))
   expect_true(all(result$Kendalls_Tau2 >=-1 & result$Kendalls_Tau2 <=1))
   expect_true(all(result$p_value_Var1 >0 & result$p_value_Var1 <=1))
   expect_true(all(result$p_value_Var2 >0 & result$p_value_Var2 <=1))
@@ -74,7 +74,7 @@ test_that("Invalid inputs produce errors", {
   
   expect_error(
     Copula_Threshold_2D_Lag(Data_Detrend=S20.Detrend.df[,-c(1,4)],
-                            Data_Declust=S20.Detrend.Declustered.df[,4],
+                            Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
                             y_lim_min=-0.075, y_lim_max =0.25,
                             u1= seq(0.5,0.99,0.01),u2=seq(0.5,0.99,0.01),
                             Upper=c(6,8), Lower=c(6,8),
@@ -84,7 +84,7 @@ test_that("Invalid inputs produce errors", {
     "Lag_Backward_Var1 must be numeric" )
   
   expect_error(Copula_Threshold_2D_Lag(Data_Detrend=S20.Detrend.df[,-c(1,4)],
-                                       Data_Declust=S20.Detrend.Declustered.df[,4],
+                                       Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
                                        y_lim_min=-0.075, y_lim_max =0.25,
                                        u1= seq(0.5,0.99,0.01),u2=seq(0.5,0.99,0.01),
                                        Upper=c(6,8), Lower=c(6,8),
@@ -93,7 +93,7 @@ test_that("Invalid inputs produce errors", {
                                        GAP=0.1, PLOT=FALSE), "Lag_Forward_Var1 must be numeric")
   
   expect_error(Copula_Threshold_2D_Lag(Data_Detrend=S20.Detrend.df[,-c(1,4)],
-                                       Data_Declust=S20.Detrend.Declustered.df[,4],
+                                       Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
                                        y_lim_min=-0.075, y_lim_max =0.25,
                                        u1= seq(0.5,0.99,0.01),u2=seq(0.5,0.99,0.01),
                                        Upper=c(6,8), Lower=c(6,8),
@@ -102,7 +102,7 @@ test_that("Invalid inputs produce errors", {
                                        GAP=0.1, PLOT=FALSE), "Lag_Backward_Var2 must be numeric")
   
   expect_error(Copula_Threshold_2D_Lag(Data_Detrend=S20.Detrend.df[,-c(1,4)],
[…]
