+++ R/Decluster_S_SW.R
@@ -22,16 +22,92 @@
 #'        S13_Rainfall_Totals_Declust$Totals[S13_Rainfall_Totals_Declust$EventID],
 #'        col=2,pch=16)
 Decluster_S_SW<-function(Data, Window_Width_Sum, Window_Width) {
-  
+
+  # Input validation
+  if (missing(Data)) {
+    stop("Data parameter is required")
+  }
+
+  if (missing(Window_Width)) {
+    stop("Window_Width parameter is required")
+  }
+
+  # Check if Data is a data frame
+  if (!is.data.frame(Data)) {
+    stop("Data must be a data frame")
+  }
+
+  # Check if Data has exactly 2 columns
+  if (ncol(Data) != 2) {
+    stop("Data must have exactly 2 columns (Date and numeric values)")
+  }
+
+  # Check first column is Date-like
+  if (!inherits(Data[,1], c("Date", "POSIXct", "POSIXt"))) {
+    # Try to convert to Date if it's character
+    if (is.character(Data[,1])) {
+      tryCatch({
+        Data[,1] <- as.Date(Data[,1])
+      }, error = function(e) {
+        stop("First column must be a Date object or convertible to Date")
+      })
+    } else {
+      stop("First column must be a Date object")
+    }
+  }
+
+  # Check for NA values in Date column
+  if (any(is.na(Data[,1]))) {
+    stop("First column (Date) contains NA values")
+  }
+
+  # Check if dates are in ascending order
+  if (any(diff(Data[,1]) <= 0)) {
+    stop("Dates in first column must be in ascending order with no duplicates")
+  }
+
+  # Check if second column contains numeric data
+  if (!is.numeric(Data[,2])) {
+    stop("Second column of Data must be numeric")
+  }
+
+  # Check if Window_Width is numeric
+  if (!is.numeric(Window_Width)) {
+    stop("Window_Width must be numeric")
+  }
+
+  # Check if Window_Width is positive
+  if (Window_Width <= 0) {
+    stop("Window_Width must be positive")
+  }
+
+  if (length(Window_Width) != 1 || Window_Width != round(Window_Width)) {
+    stop("Window_Width must be a single integer value")
+  }
+
+  # Check if Window_Width is numeric
+  if (!is.numeric(Window_Width_Sum)) {
+    stop("Window_Width_Sum must be numeric")
+  }
+
+  # Check if Window_Width is positive
+  if (Window_Width_Sum <= 0) {
+    stop("Window_Width_Sum must be positive")
+  }
+
+  if (length(Window_Width_Sum) != 1 || Window_Width_Sum != round(Window_Width_Sum)) {
+    stop("Window_Width_Sum must be a single integer value")
+  }
+
   #Index of NA elements will be added to z
   z<-0
-  
+
   #Assigning any NA values to a very small number relative to the data
   if (length(which(is.na(Data[,2]) == T)) > 0) {
     z <- which(is.na(Data[,2]) == T)
     Data[z,2] <- min(Data[,2], na.rm = T) - 1000
   }
-  
+
   #Summing observations using a centered window
   if(Window_Width_Sum>1){
     Sum<-c(rep(NA,round((Window_Width_Sum - Window_Width_Sum %% 2)/2,0)),c(cumsum(Data[,2])[-c(1:(Window_Width_Sum -1))]) -
@@ -41,15 +117,15 @@ Decluster_S_SW<-function(Data, Window_Width_Sum, Window_Width) {
   if(Window_Width_Sum==1){
     Sum<-Data[,2]
   }
-  
+
[…]
