+++ tests/testthat/test-Decluster_S_SW.R
@@ -9,12 +9,12 @@ test_that("Decluster_SW functions basic functionality", {
   expect_type(result, "list")
   
   #Check names of output
-  expect_named(result, c("Detrend", "Totals", "Declustered", "EventID", "Event_start", "Event_End"))
+  expect_named(result, c("Detrend", "Totals", "Declustered", "EventID", "Event_start", "Event_end"))
   
   # Check length of outputs
-  expect_equal(length(result$Detrend), nrow(test_data))
-  expect_equal(length(result$Totals),  nrow(test_data))
-  expect_equal(length(result$Declustered),  nrow(test_data))
+  expect_equal(length(result$Detrend), nrow(S13_Rainfall))
+  expect_equal(length(result$Totals),  nrow(S13_Rainfall))
+  expect_equal(length(result$Declustered),  nrow(S13_Rainfall))
   
 })
 
@@ -24,18 +24,18 @@ test_that("Declustering actually works", {
   result <- Decluster_S_SW(Data = S13_Rainfall, Window_Width_Sum=24, Window_Width = 7*24)
   
   # Should have more NAs in declustered than original
-  original_nas <- sum(is.na(test_data[,2]))
+  original_nas <- sum(is.na(S13_Rainfall[,2]))
   declustered_nas <- sum(is.na(result$Declustered))
   expect_true(declustered_nas >= original_nas)
   
   # Filter out boundary events that can't be tested
-  testable_events <- result$EventID[result$EventID > 1 & result$EventID < nrow(test_data)]
+  testable_events <- result$EventID[result$EventID > 1 & result$EventID < nrow(S13_Rainfall)]
   
   if (length(testable_events) > 0) {
     # Test that events are >= neighbors (allows for equal values)
-    left_neighbors <- test_data[testable_events - 1, 2]
-    right_neighbors <- test_data[testable_events + 1, 2]
-    event_values <- test_data[testable_events, 2]
+    left_neighbors <- S13_Rainfall[testable_events - 1, 2]
+    right_neighbors <- S13_Rainfall[testable_events + 1, 2]
+    event_values <- S13_Rainfall[testable_events, 2]
     
     # Remove any comparisons involving NA values
     valid_comparisons <- !is.na(event_values) & !is.na(left_neighbors) & !is.na(right_neighbors)
@@ -98,6 +98,7 @@ test_that("Invalid inputs produce errors", {
   expect_error(Decluster_S_SW(Data = S13_Rainfall, Window_Width_Sum=24, Window_Width= -2),
                "Window_Width must be positive")
   
+  S13_Rainfall = S13_Rainfall
   S13_Rainfall_modified[,2] <- rep("2",nrow(S13_Rainfall))                       
   expect_error(Decluster_S_SW(Data = S13_Rainfall_modified, Window_Width_Sum=24, Window_Width = 7*24),
                "Second column of Data must be numeric")
