+++ R/Diag_Non_Con_Sel.R
@@ -3,11 +3,12 @@
 #' Plots demonstrating the goodness of fit of a selected (not truncated) non-extreme marginal distribution to a dataset.
 #'
 #' @param Data Numeric vector containing realizations of the variable of interest.
-#' @param x_lab Numeric vector of length one specifyingLabel on the x-axis of histogram and cummulative distribution plot.
+#' @param Omit Character vector specifying any distributions that are not to be tested. Default \code{"NA"}, all distributions are fit.
+#' @param x_lab Numeric vector of length one specifying. Label on the x-axis of histogram and cumulative distribution plot.
 #' @param y_lim_min Numeric vector of length one specifying the lower y-axis limit of the histogram.
 #' @param y_lim_max Numeric vector of length one specifying the upper y-axis limit of the histogram.
-#' @param Selected Charactor vector of length one specifying the chosen distribution, options are the Gaussian \code{"Gaus"} and logistic \code{"Logis"}.
-#' @return Panel consisting of three plots. Upper plot: Plots depicting the AIC of the two fitted distributions. Middle plot: Probabilty Density Functions (PDFs) of the \code{selected} distribtions superimposed on a histgram of the data. Lower plot: Cummulative distribution function (CDFs) of the \code{selected} distribution overlaid on a plot of the empirical CDF.
+#' @param Selected Character vector of length one specifying the chosen distribution. Options are: Gaussian \code{"Gaus"}, Gumbel \code{"Gum"}, Laplace \code{"Lapl"}, logistic \code{"Logis"}, and reverse Gumbel \code{"RGum"}.
+#' @return Panel consisting of three plots. Upper plot: Plots depicting the AIC of the two fitted distributions. Middle plot: Probability Density Functions (PDFs) of the \code{selected} distributions superimposed on a histogram of the data. Lower plot: Cumulative distribution function (CDFs) of the \code{selected} distribution overlaid on a plot of the empirical CDF.
 #' @seealso \code{\link{Diag_Non_Con}}
 #' @export
 #' @examples
@@ -17,68 +18,252 @@
 #' Diag_Non_Con(Data=S20.Rainfall$Data$OsWL,x_lab="O-sWL (ft NGVD 29)",
 #'              y_lim_min=0,y_lim_max=1.5)
 #' Diag_Non_Con_Sel(Data=S20.Rainfall$Data$OsWL,x_lab="O-sWL (ft NGVD 29)",
-#'                  y_lim_min=0,y_lim_max=1.5,Selected="Twe")
-Diag_Non_Con_Sel<-function(Data,x_lab = "Data",y_lim_min = 0,y_lim_max = 1,Selected){
+#'                  y_lim_min=0,y_lim_max=1.5,Selected="Logis")
+Diag_Non_Con_Sel<-function(Data, Omit=NA, x_lab = "Data",y_lim_min = 0,y_lim_max = 1,Selected){
+
+  # Check if Data is provided
+  if (missing(Data)) {
+    stop("Data parameter is required")
+  }
+
+  # Check if Data is numeric
[…]

[…]
