+++ R/Diag_Non_Con.R
@@ -129,21 +129,22 @@ Diag_Non_Con<-function(Data,Omit=NA,x_lab,y_lim_min=0,y_lim_max=1){
   if(any(Test==2)){
     fit <- tryCatch(gamlss(Data  ~ 1, family=GU),
                     error = function(e) "error")
-    Omit.2[1] = ifelse(fit=="error" | exp(fit$sigma.coefficients) < 0,"Gum",NA)
+    Omit.2[1] = ifelse(fit=="error" | (!is.character(fit) && exp(fit$sigma.coefficients) < 0),"Gum",NA)
   }
 
+  # Fixed error handling for Laplace distribution
   if(any(Test==3)){
     fit <- tryCatch(fitdistr(Data, dlaplace, start=list(location=mean(Data),scale=sd(Data)/sqrt(2))),
                     error = function(e) "error")
     Omit.2[2] = ifelse(fit=="error","Lapl",NA)
   }
 
+  # Fixed error handling for Reverse Gumbel distribution
   if(any(Test==5)){
     fit <- tryCatch(gamlss(Data ~ 1,family=RG),
                     error = function(e) "error")
-    Omit.2[3] = ifelse(fit=="error" | exp(fit$sigma.coefficients) < 0,"RGum",NA)
+    Omit.2[3] = ifelse(fit=="error" | (!is.character(fit) && exp(fit$sigma.coefficients) < 0),"RGum",NA)
   }
-
 
   #Distributions to test
   if(any(is.na(Omit.2[1])==F)){
