### 2025-07-09 1fb412d
Delete tests/testthat/test-Decluster_S_SW.R
