### 2025-07-09 b977a0c
Delete tests/testthat/Diag_Non_Con.R
