+++ tests/testthat/Diag_Non_Con.R
@@ -1,258 +1,0 @@
-#' Goodness of fit of non-extreme marginal distributions
-#'
-#' Fits two (unbounded) non-extreme marginal distributions to a dataset and returns three plots demonstrating their relative goodness of fit.
-#' The distributions are the Gaussian \code{"Gaus"}, Gumbel \code{"Gum"}, Laplace \code{"Lapl"}, Logistic \code{"Logis"} and the reverse Gumbel \code{"RGum"}.
-
-#' @param Data Numeric vector containing realizations of the variable of interest.
-#' @param Omit Character vector specifying any distributions that are not to be tested. Default \code{"NA"}, all distributions are fit.
-#' @param x_lab Character vector of length one specifying the label on the x-axis of histogram and cumulative distribution plot.
-#' @param y_lim_min Numeric vector of length one specifying the lower y-axis limit of the histogram. Default is \code{0}.
-#' @param y_lim_max Numeric vector of length one specifying the upper y-axis limit of the histogram. Default is \code{1}.
-#' @return Dataframe \code{$AIC} giving the AIC associated with each distribution and the name of the best fitting distribution \code{$Best_fit}. Panel consisting of three plots. Upper plot: Plot depicting the AIC of the two fitted distributions. Middle plot: Probability Density Functions (PDFs) of the fitted distributions superimposed on a histogram of the data. Lower plot: Cumulative Distribution Functions (CDFs) of the fitted distributions overlaid on a plot of the empirical CDF.
-#' @seealso \code{\link{Copula_Threshold_2D}}
-#' @export
-#' @examples
-#' S20.Rainfall<-Con_Sampling_2D(Data_Detrend=S20.Detrend.df[,-c(1,4)],
-#'                               Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
-#'                               Con_Variable="Rainfall",Thres=0.97)
-#' Diag_Non_Con(Data=S20.Rainfall$Data$OsWL,x_lab="O-sWL (ft NGVD 29)",
-#'              y_lim_min=0,y_lim_max=1.5)
-Diag_Non_Con<-function(Data,Omit=NA,x_lab,y_lim_min=0,y_lim_max=1){
-
-  # Check if Data is provided
-  if (missing(Data)) {
-    stop("Data parameter is required")
-  }
-
-  # Check if Data is numeric
-  if (!is.numeric(Data)) {
-    stop("Data must be numeric, got: ", class(Data)[1])
-  }
-
-  # Check if Data is vector or can be converted to vector
-  if (is.matrix(Data) || is.data.frame(Data)) {
-    if (ncol(Data) > 1) {
-      warning("Data has multiple columns. Using first column only.")
-      Data <- Data[, 1]
-    } else {
-      Data <- as.vector(Data)
-    }
-  }
-
-  # Check for empty data
-  if (length(Data) == 0) {
-    stop("Data is empty")
-  }
-
-  # Check for all NA values
-  if (all(is.na(Data))) {
-    stop("Data contains only NA values")
-  }
[…]
