+++ tests/testthat/Diag_Non_Con_Trunc.R
@@ -1,334 +1,0 @@
-#' Goodness of fit of non-extreme marginal distributions
-#'
-#' Fits ten (truncated) non-extreme marginal distributions to a dataset and returns three plots demonstrating their relative goodness of fit.
-#' The distributions are the Birnbaum-Saunders \code{"BS"}, exponential \code{"Exp"}, two-parameter gamma \code{"Gam(2)"}, three-parameter gamma \code{"Gam(3)"}, mixed two-parameter gamma \code{"GamMix(2)"}, mixed three-parameter gamma \code{"GamMix(3)"}, lognormal \code{"LNorm"}, truncated normal \code{"TNorm"}, Tweedie \code{"Twe"} and the Weibull \code{"Weib"}.
-#' @param Data Numeric vector containing realizations of the variable of interest.
-#' @param Omit Character vector specifying any distributions that are not to be tested. Default \code{"NA"}, all distributions are fit.
-#' @param x_lab Character vector of length one specifying the label on the x-axis of histogram and cumulative distribution plot.
-#' @param y_lim_min Numeric vector of length one specifying the lower y-axis limit of the histogram. Default is \code{0}.
-#' @param y_lim_max Numeric vector of length one specifying the upper y-axis limit of the histogram. Default is \code{1}.
-#' @return Dataframe \code{$AIC} giving the AIC associated with each distribution and the name of the best fitting distribution \code{$Best_fit}. Panel consisting of three plots. Upper plot: Plot depicting the AIC of the ten fitted distributions. Middle plot: Probability Density Functions (PDFs) of the fitted distributions superimposed on a histogram of the data. Lower plot: Cumulative Distribution Functions (CDFs) of the fitted distributions overlaid on a plot of the empirical CDF.
-#' @seealso \code{\link{Copula_Threshold_2D}}
-#' @export
-#' @examples
-#' S20.OsWL<-Con_Sampling_2D(Data_Detrend=S20.Detrend.df[,-c(1,4)],
-#'                          Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
-#'                          Con_Variable="OsWL",Thres=0.97)
-#' S20.OsWL$Data$Rainfall <- S20.OsWL$Data$Rainfall + runif(length(S20.OsWL$Data$Rainfall),0.001,0.01)
-#' Diag_Non_Con_Trunc(Data=S20.OsWL$Data$Rainfall,x_lab="Rainfall (Inches)",
-#'                    y_lim_min=0,y_lim_max=2)
-Diag_Non_Con_Trunc<-function(Data,Omit=NA,x_lab="Data",y_lim_min=0,y_lim_max=1){
-
-  # Check if Data is provided
-  if (missing(Data)) {
-    stop("Data parameter is required")
-  }
-
-  # Check if Data is numeric
-  if (!is.numeric(Data)) {
-    stop("Data must be numeric, got: ", class(Data)[1])
-  }
-
[…]
