+++ tests/testthat/test-Diag_Non_Con_Sel.R
@@ -8,48 +8,48 @@ S20.Rainfall<-Con_Sampling_2D(Data_Detrend=S20.Detrend.df[,-c(1,4)],
 
 # Test basic functionality for both distributions
 test_that("Basic functionality works for all distributions", {
-  
+
   # Test Gaussian
-  expect_silent(Diag_Non_Con_Sel(Data = S20.Rainfall$Data$OsWL, 
-                                 x_lab = "O-sWL (ft NGVD 29)", 
-                                 y_lim_min = 0, y_lim_max = 1.5, 
+  expect_silent(Diag_Non_Con_Sel(Data = S20.Rainfall$Data$OsWL,
+                                 x_lab = "O-sWL (ft NGVD 29)",
+                                 y_lim_min = 0, y_lim_max = 1.5,
                                  Selected = "Gaus"))
-  
-  # Test Gumbel  
-  expect_silent(Diag_Non_Con_Sel(Data = S20.Rainfall$Data$OsWL, 
-                                 x_lab = "O-sWL (ft NGVD 29)", 
-                                 y_lim_min = 0, y_lim_max = 1.5, 
+
+  # Test Gumbel
+  expect_silent(Diag_Non_Con_Sel(Data = S20.Rainfall$Data$OsWL,
+                                 x_lab = "O-sWL (ft NGVD 29)",
+                                 y_lim_min = 0, y_lim_max = 1.5,
                                  Selected = "Gum"))
-  
+
   # Test Laplace
-  expect_silent(Diag_Non_Con_Sel(Data = S20.Rainfall$Data$OsWL, 
-                                 x_lab = "O-sWL (ft NGVD 29)", 
-                                 y_lim_min = 0, y_lim_max = 1.5, 
+  expect_silent(Diag_Non_Con_Sel(Data = S20.Rainfall$Data$OsWL,
+                                 x_lab = "O-sWL (ft NGVD 29)",
+                                 y_lim_min = 0, y_lim_max = 1.5,
                                  Selected = "Lapl"))
-  # Test Logistic  
-  expect_silent(Diag_Non_Con_Sel(Data = S20.Rainfall$Data$OsWL, 
-                                 x_lab = "O-sWL (ft NGVD 29)", 
-                                 y_lim_min = 0, y_lim_max = 1.5, 
+  # Test Logistic
+  expect_silent(Diag_Non_Con_Sel(Data = S20.Rainfall$Data$OsWL,
+                                 x_lab = "O-sWL (ft NGVD 29)",
+                                 y_lim_min = 0, y_lim_max = 1.5,
                                  Selected = "Logis"))
-  
+
   # Test Revere Gumbel
-  expect_silent(Diag_Non_Con_Sel(Data = S20.Rainfall$Data$OsWL, 
-                                 x_lab = "O-sWL (ft NGVD 29)", 
-                                 y_lim_min = 0, y_lim_max = 1.5, 
+  expect_silent(Diag_Non_Con_Sel(Data = S20.Rainfall$Data$OsWL,
+                                 x_lab = "O-sWL (ft NGVD 29)",
+                                 y_lim_min = 0, y_lim_max = 1.5,
                                  Selected = "RGum"))
 })
 
 # Test NA handling
 test_that("Handles NA values", {
-  
-  data_with_na <- c(S20.Rainfall$Data$OsWL, NA, NA, NA)
-  
-  expect_warning(Diag_Non_Con_Sel(Data = data_with_na,x_lab="O-sWL (ft NGVD 29)",
[…]
