+++ DESCRIPTION
@@ -53,6 +53,7 @@ Imports:
     dplyr,
     fitdistrplus,
     gamlss,
+    gamlss.dist,
     gamlss.mx,
     GeneralizedHyperbolic,
     ks,
