### 2025-07-09 51d8b00
Delete tests/testthat/Diag_Non_Con.R
