+++ tests/testthat/Diag_Non_Con_Sel.R
@@ -1,275 +1,0 @@
-#' Demonstrate the goodness of fit of the selected non-extreme marginal distribution
-#'
-#' Plots demonstrating the goodness of fit of a selected (not truncated) non-extreme marginal distribution to a dataset.
-#'
-#' @param Data Numeric vector containing realizations of the variable of interest.
-#' @param Omit Character vector specifying any distributions that are not to be tested. Default \code{"NA"}, all distributions are fit.
-#' @param x_lab Numeric vector of length one specifying. Label on the x-axis of histogram and cumulative distribution plot.
-#' @param y_lim_min Numeric vector of length one specifying the lower y-axis limit of the histogram.
-#' @param y_lim_max Numeric vector of length one specifying the upper y-axis limit of the histogram.
-#' @param Selected Character vector of length one specifying the chosen distribution. Options are: Gaussian \code{"Gaus"}, Gumbel \code{"Gum"}, Laplace \code{"Lapl"}, logistic \code{"Logis"}, and reverse Gumbel \code{"RGum"}.
-#' @return Panel consisting of three plots. Upper plot: Plots depicting the AIC of the two fitted distributions. Middle plot: Probability Density Functions (PDFs) of the \code{selected} distributions superimposed on a histogram of the data. Lower plot: Cumulative distribution function (CDFs) of the \code{selected} distribution overlaid on a plot of the empirical CDF.
-#' @seealso \code{\link{Diag_Non_Con}}
-#' @export
-#' @examples
-#' S20.Rainfall<-Con_Sampling_2D(Data_Detrend=S20.Detrend.df[,-c(1,4)],
-#'                               Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
-#'                               Con_Variable="Rainfall",Thres=0.97)
-#' Diag_Non_Con(Data=S20.Rainfall$Data$OsWL,x_lab="O-sWL (ft NGVD 29)",
-#'              y_lim_min=0,y_lim_max=1.5)
-#' Diag_Non_Con_Sel(Data=S20.Rainfall$Data$OsWL,x_lab="O-sWL (ft NGVD 29)",
-#'                  y_lim_min=0,y_lim_max=1.5,Selected="Logis")
-Diag_Non_Con_Sel<-function(Data, Omit=NA, x_lab = "Data",y_lim_min = 0,y_lim_max = 1,Selected){
-
-  # Check if Data is provided
-  if (missing(Data)) {
-    stop("Data parameter is required")
-  }
-
-  # Check if Data is numeric
-  if (!is.numeric(Data)) {
-    stop("Data must be numeric, got: ", class(Data)[1])
-  }
-
-  # Check if Data is vector or can be converted to vector
-  if (is.matrix(Data) || is.data.frame(Data)) {
-    if (ncol(Data) > 1) {
-      warning("Data has multiple columns. Using first column only.")
-      Data <- Data[, 1]
-    } else {
-      Data <- as.vector(Data)
-    }
-  }
-
-  # Check for empty data
-  if (length(Data) == 0) {
-    stop("Data is empty")
-  }
-
-  # Check for all NA values
-  if (all(is.na(Data))) {
[…]
