### 2025-07-09 f9d5326
Delete tests/testthat/test-Diag_Non_Con.R
