+++ tests/testthat/test-Standard_Copula_Fit.R
@@ -2,71 +2,72 @@
 
 test_that("Gaussian copula works", {
   result <- Standard_Copula_Fit(Data=S20.Detrend.df, Copula_Type="Gaussian")
-  expect_type(result, "list")
+  expect_type(result, 'S4')
   expect_true(inherits(result, "normalCopula"))
-  
+
   # Test parameter validity
   params <- getTheta(result)
   expect_true(all(params >= -1 & params <= 1))  # Correlation bounds
-  expect_equal(length(params), ncol(S20.Detrend.df)*(ncol(S20.Detrend.df)-1)/2)
+  expect_equal(length(params), 3)
 })
 
 test_that("t-copula works", {
   result <- Standard_Copula_Fit(Data=S20.Detrend.df, Copula_Type="tcopula")
-  expect_type(result, "list")
+  expect_type(result, 'S4')
   expect_true(inherits(result, "tCopula"))
-  
+
   # Test parameter validity
   params <- getTheta(result)
-  expect_true(all(params[1:3] >= -1 & params[1:3] <= 1)) 
-  expect_true(params[4] >0) 
-  expect_equal(length(params), 1+ ncol(S20.Detrend.df)*(ncol(S20.Detrend.df)-1)/2)
-  
+  expect_true(all(params[1:3] >= -1 & params[1:3] <= 1))
+  expect_true(params[4] >0)
+  expect_equal(length(params), 4)
+
 })
 
 test_that("Gumbel copula works", {
   result <- Standard_Copula_Fit(Data=S20.Detrend.df, Copula_Type="Gumbel")
-  expect_type(result, "list")
+  expect_type(result, 'S4')
   expect_true(inherits(result, "gumbelCopula"))
   params <- getTheta(result)
-  expect_true(all(params >= 1)) 
+  expect_true(all(params >= 1))
 })
 
 test_that("Clayton copula works", {
   result <- Standard_Copula_Fit(Data=S20.Detrend.df, Copula_Type="Clayton")
-  expect_type(result, "list")
+  expect_type(result, 'S4')
   expect_true(inherits(result, "claytonCopula"))
   params <- getTheta(result)
-  expect_equal(length(params), 1) 
-  expect_true(all(params > 0)) 
+  expect_equal(length(params), 1)
+  expect_true(all(params > 0))
 })
 
 test_that("Frank copula works", {
   result <- Standard_Copula_Fit(Data=S20.Detrend.df, Copula_Type="Frank")
-  expect_type(result, "list")
+  expect_type(result, 'S4')
   expect_true(inherits(result, "frankCopula"))
-  expect_true(is.finite(params))    
-  expect_true(is.numeric(params))   
-  expect_equal(length(params), 1)  
+  params <- getTheta(result)
+  expect_true(is.finite(params))
+  expect_true(is.numeric(params))
+  expect_equal(length(params), 1)
 })
 
 
 
 # Test that invalid inputs gives errors
 test_that("Invalid inputs produce errors", {
-  
-  expect_error(Standard_Copula_Fit(Copula_Type="Gumbel"),
-               "Error: Data is empty")
-  
+
+  expect_error(Standard_Copula_Fit(Data = "", Copula_Type="Gumbel"),
+               "Error in Data[, 1] : incorrect number of dimensions")
+
   expect_error(Standard_Copula_Fit(Data=S20.Detrend.df[1:5,], Copula_Type="Gumbel"),
                "Error: Insufficient non-missing data (need at least 10 complete observations)")
-  
+
   expect_error(Standard_Copula_Fit(Data=S20.Detrend.df[1:5,], Copula_Type="Not a Copula"),
[…]
