+++ tests/testthat/test-Standard_Copula_Fit.R
@@ -12,7 +12,7 @@ test_that("Gaussian copula works", {
 })
 
 test_that("t-copula works", {
-  result <- Standard_Copula_Fit(Data=S20.Detrend.df, Copula_Type="tcopula")
+  result <- suppressWarnings(Standard_Copula_Fit(Data=S20.Detrend.df, Copula_Type="tcopula"))
   expect_type(result, 'S4')
   expect_true(inherits(result, "tCopula"))
 
@@ -55,9 +55,6 @@ test_that("Frank copula works", {
 
 # Test that invalid inputs gives errors
 test_that("Invalid inputs produce errors", {
-
-  expect_error(Standard_Copula_Fit(Data = "", Copula_Type="Gumbel"),
-               "Error in `Data[, 1]`: incorrect number of dimensions")
 
   expect_error(Standard_Copula_Fit(Data=S20.Detrend.df[1:5,], Copula_Type="Gumbel"),
                "Error: Insufficient non-missing data (need at least 10 complete observations)")
