+++ R/Standard_Copula_Sel.R
@@ -8,23 +8,33 @@
 #' @export
 #' @seealso \code{\link{Dataframe_Combine}} \code{\link{Standard_Copula_Fit}}
 #' @examples
-#' Standard_Copula_Sel(Data_Detrend=S20.Detrend.df)
+#' Standard_Copula_Sel(Data=S20.Detrend.df)
 Standard_Copula_Sel<-function(Data){
-  
+
+  if(missing(Data) || is.null(Data)) {
+    stop("Error: Data is required")
+  }
+  if(!is.data.frame(Data) && !is.matrix(Data)) {
+    stop("Error: Data must be a data.frame or matrix")
+  }
+  if(nrow(Data) < 10) {
+    stop("Error: Data must have at least 10 rows")
+  }
+
   if(class(Data[,1])=="Date" | class(Data[,1])=="factor"){
     Data <- Data[,-1]
   }
-  
+
   #Fitting Gaussian
   norm.cop <- normalCopula(rep(0.6,ncol(Data)*(ncol(Data)-1)/2), dim=(ncol(Data)), dispstr="un")
   Gaussian_Copula_Fit <- fitCopula(norm.cop,data=pobs(na.omit(Data[,1:ncol(Data)])))
   AIC_Gaussian <- ncol(Data)*2 - logLik(Gaussian_Copula_Fit)[1]
-  
+
   #Fitting student-t
   t.cop<-tCopula(rep(0.6,ncol(Data)*(ncol(Data)-1)/2), dim=ncol(Data), dispstr="un",df=5)
   t_Copula_Fit <- fitCopula(t.cop,data=pobs(na.omit(Data[,1:ncol(Data)])))
   AIC_tCopula <- (ncol(Data)+1)*2 - logLik(t_Copula_Fit)[1]
-  
+
   #Fitting Gumbel
   Fit_Gumbel <- function (Data) {
     out <- tryCatch(1*2 - logLik(fitCopula(gumbelCopula(dim=ncol(Data)),pobs(na.omit(Data[,1:ncol(Data)]))))[1]
@@ -32,7 +42,7 @@ Standard_Copula_Sel<-function(Data){
     return(out)
   }
   AIC_Gumbel<-Fit_Gumbel(Data)
-  
+
   #Fitting Clayton
   Fit_Clayton <- function (Data) {
     out <- tryCatch(1*2 - logLik(fitCopula(claytonCopula(dim=ncol(Data)),pobs(na.omit(Data[,1:ncol(Data)]))))[1]
@@ -40,7 +50,7 @@ Standard_Copula_Sel<-function(Data){
     return(out)
   }
   AIC_Clayton<-Fit_Clayton(Data)
-  
+
   #Fitting Frank
   Fit_Frank <- function (Data) {
     out <- tryCatch(1*2 - logLik(fitCopula(frankCopula(dim=ncol(Data)),pobs(na.omit(Data[,1:ncol(Data)]))))[1]
@@ -48,7 +58,7 @@ Standard_Copula_Sel<-function(Data){
     return(out)
   }
   AIC_Frank<-Fit_Frank(Data)
-  
+
   res<-data.frame(c("Gaussian","t-cop","Gumbel","Clayton","Frank"),c(AIC_Gaussian,AIC_tCopula,AIC_Gumbel,AIC_Clayton,AIC_Frank))
   colnames(res)<-c("Copula","AIC")
   return(res)
