+++ R/Diag_Non_Con_Sel.R
@@ -127,6 +127,12 @@ Diag_Non_Con_Sel<-function(Data, Omit=NA, x_lab = "Data",y_lim_min = 0,y_lim_max
   GU <- get("GU", envir = asNamespace("gamlss.dist"))
   RG <- get("RG", envir = asNamespace("gamlss.dist"))
 
+  # Load density and cumulative functions safely
+  dRG <- get("dRG", envir = asNamespace("gamlss.dist"))
+  pRG <- get("pRG", envir = asNamespace("gamlss.dist"))
+  dGU <- get("dGU", envir = asNamespace("gamlss.dist"))
+  pGU <- get("pGU", envir = asNamespace("gamlss.dist"))
+
   #Distributions to test
   Dist<-c("Gaus","Gum","Lapl","Logis","RGum")
   Dist.2 <- Dist
