+++ R/Diag_Non_Con_Trunc.R
@@ -104,14 +104,14 @@ Diag_Non_Con_Trunc<-function(Data,Omit=NA,x_lab="Data",y_lim_min=0,y_lim_max=1){
   if (!requireNamespace("gamlss.dist", quietly = TRUE)) {
     stop("The 'gamlss.dist' package is required but not installed.")
   }
-  GU <- get("GU", envir = asNamespace("gamlss.dist"))
-  RG <- get("RG", envir = asNamespace("gamlss.dist"))
+  GG <- get("GG", envir = asNamespace("gamlss.dist"))
+  GA <- get("GA", envir = asNamespace("gamlss.dist"))
 
   # Load density and cumulative functions safely
-  dRG <- get("dRG", envir = asNamespace("gamlss.dist"))
-  pRG <- get("pRG", envir = asNamespace("gamlss.dist"))
-  dGU <- get("dGU", envir = asNamespace("gamlss.dist"))
-  pGU <- get("pGU", envir = asNamespace("gamlss.dist"))
+  dGG <- get("dGG", envir = asNamespace("gamlss.dist"))
+  pGG <- get("pGG", envir = asNamespace("gamlss.dist"))
+  dGA <- get("dGA", envir = asNamespace("gamlss.dist"))
+  pGA <- get("pGA", envir = asNamespace("gamlss.dist"))
 
   #Colors for plots
   mypalette<-c("Black",brewer.pal(9,"Set1"))
