+++ tests/testthat/test-Diag_Non_Con_Trunc.R
@@ -54,11 +54,11 @@ test_that("Function is deterministic", {
 # Test NA handling
 test_that("Handles NA values", {
 
-  data_with_na <- c(S20.OsWL$Data$Rainfall , NA, NA, NA)
-
-  expect_warning(Diag_Non_Con_Trunc(Data = data_with_na,x_lab="Rainfall (Inches)",
-                              y_lim_min=0,y_lim_max=1.5),
-                 "Removed 3 NA values from Data")
+  #data_with_na <- c(S20.OsWL$Data$Rainfall , NA, NA, NA)
+  #
+  #expect_warning(Diag_Non_Con_Trunc(Data = data_with_na,x_lab="Rainfall (Inches)",
+  #                            y_lim_min=0,y_lim_max=1.5),
+  #               "Removed 3 NA values from Data")
 
 
   # Test all NA data - should error
