+++ R/Diag_Non_Con_Trunc_Sel.R
@@ -124,14 +124,14 @@ Diag_Non_Con_Trunc_Sel<-function(Data,Selected,Omit=NA,x_lab="Data",y_lim_min=0,
   if (!requireNamespace("gamlss.dist", quietly = TRUE)) {
     stop("The 'gamlss.dist' package is required but not installed.")
   }
-  GU <- get("GU", envir = asNamespace("gamlss.dist"))
-  RG <- get("RG", envir = asNamespace("gamlss.dist"))
+  GU <- get("GG", envir = asNamespace("gamlss.dist"))
+  RG <- get("GA", envir = asNamespace("gamlss.dist"))
 
   # Load density and cumulative functions safely
-  dRG <- get("dRG", envir = asNamespace("gamlss.dist"))
-  pRG <- get("pRG", envir = asNamespace("gamlss.dist"))
-  dGU <- get("dGU", envir = asNamespace("gamlss.dist"))
-  pGU <- get("pGU", envir = asNamespace("gamlss.dist"))
+  dGG <- get("dGG", envir = asNamespace("gamlss.dist"))
+  pGG <- get("pGG", envir = asNamespace("gamlss.dist"))
+  dGA <- get("dGA", envir = asNamespace("gamlss.dist"))
+  pGA <- get("pGA", envir = asNamespace("gamlss.dist"))
 
   #Colors for plots
   mypalette<-c("Black",brewer.pal(9,"Set1"))
