+++ R/SLR_Scenarios.R
@@ -30,17 +30,44 @@
 #' SLR_Scenarios(SeaLevelRise=0.8, Scenario="Other", Unit = "m", Year=2022,
 #'               Location="Fort Myers", New_Scenario=SeaLevelRise.2022_input)
 SLR_Scenarios<-function(SeaLevelRise, Scenario="Compact", Unit = "m", Year=2022, Location="Key West", New_Scenario=NA){
-  
+
+  #Error messages to catch invalid inputs
+  if (missing(SeaLevelRise) || !is.numeric(SeaLevelRise)) {
+    stop("Error: SeaLevelRise missing or not numeric")
+  }
+
+  if (!Scenario %in% c("Compact","NOAA2017","NOAA2022")  && !is.na(Scenario)) {
+    stop("Error: Invalid Scenario")
+  }
+
+  if (!Unit %in% c("m", "Inches")) {
+    stop("Error: Unit input must be m or Inches")
+  }
+
+  if (missing(Year) || !is.numeric(Year) || Year<2020) {
+    stop("Error: Invalid Year input")
+  }
+
+  if (!is.character(Location)) {
+    stop("Error: Location name invalid")
+  }
+
+  if (!is.na(New_Scenario)[1]) {
+    if (!is.data.frame(New_Scenario) || ncol(New_Scenario) < 2 || !all(sapply(New_Scenario, is.numeric))) {
+      stop("Error: New_Scenario invalid")
+    }
+  }
+
   #Colors
   mypalette<-brewer.pal(7,"Set1")
-  
+
   #Layout of graphical output
   layout(matrix(c(1,1,2), 3, 1, byrow = TRUE))
   par(mar=c(4.2,6,1,1))
-  
+
   #Plotting the SLR scenarios
   y_axis<-ifelse(Unit=="m",paste('Projected sea level rise at', Location, '(m)'),past('Projected sea level rise at', Location,'(ft)'))
-  
+
   if(Scenario=="Compact"){
     #Composing the AR5 scenario
     SeaLevelRise.AR5<-data.frame("Year"=1:4,"Medium"=1:4)
@@ -62,7 +89,7 @@ SLR_Scenarios<-function(SeaLevelRise, Scenario="Compact", Unit = "m", Year=2022,
     spline.low<-spline(NOAAetal2017$Year,ifelse(Unit=="m",1,0.00328084)*NOAAetal2017$Low, xout=seq(Year,2100,0.25))
     spline.low$y<-spline.low$y-spline.low$y[1]
   }
-  
+
   if(Scenario=="NOAA2022"){
     NOAAetal2022.Year<-seq(2020,2100,10)
     x<-ifelse(Location=="Miami Beach",1,2)
@@ -73,7 +100,7 @@ SLR_Scenarios<-function(SeaLevelRise, Scenario="Compact", Unit = "m", Year=2022,
     spline.low<-spline(NOAAetal2022.Year,ifelse(Unit=="m",1,0.00328084)*NOAAetal2022[c(17,2)[x],6:14]/1000, xout=seq(Year,2100,0.25))
     spline.low$y<-spline.low$y-spline.low$y[1]
   }
-  
+
   if(Scenario=="Other"){
     if((ncol(New_Scenario)-1)==5){
       mypalette<-c("Dark Red",brewer.pal(7,"Set1")[1:3],"Dark Green")
@@ -92,7 +119,7 @@ SLR_Scenarios<-function(SeaLevelRise, Scenario="Compact", Unit = "m", Year=2022,
[…]
