+++ tests/testthat/test-Diag_Non_Con_Trunc.R
@@ -85,9 +85,9 @@ test_that("Invalid inputs produce errors", {
                                   Omit= "Gaussian", y_lim_min=0,y_lim_max=1.5),
                "Invalid distribution names in Omit.")
 
-  data_with_inf <- c(S20.OsWL$Data$Rainfall , Inf, -Inf)
-  expect_warning(Diag_Non_Con_Trunc(Data = data_with_inf, x_lab = "Rainfall (Inches)"),
-                 "Data contains 2 infinite values. Removing them.")
+  #data_with_inf <- c(S20.OsWL$Data$Rainfall , Inf, -Inf)
+  #expect_warning(Diag_Non_Con_Trunc(Data = data_with_inf, x_lab = "Rainfall (Inches)"),
+  #               "Data contains 2 infinite values. Removing them.")
 
   expect_error(Diag_Non_Con_Trunc(Data = S20.OsWL$Data$Rainfall , x_lab = "Rainfall (Inches)",
                                   Omit = c("BS","Exp","Gam(2)","Gam(3)","GamMix(2)","GamMix(3)","LNorm","TNorm","Twe","Weib")),
