+++ tests/testthat/test-GPD_Fit.R
@@ -7,12 +7,12 @@ test_that("GPD_Fit basic functionality works", {
   expect_type(result, 'list')
   expect_named(result,  c("Threshold","Rate","sigma","xi","sigma.SE","xi.SE"))
 
-  expect_type(result$Threshold,  numeric)
-  expect_type(result$Rate,  numeric)
-  expect_type(result$sigma,  numeric)
-  expect_type(result$xi,  numeric)
-  expect_type(result$sigma.se,  numeric)
-  expect_type(result$xi.se,  numeric)
+  expect_type(result$Threshold,  'numeric')
+  expect_type(result$Rate,  'numeric')
+  expect_type(result$sigma,  'numeric')
+  expect_type(result$xi,  'numeric')
+  expect_type(result$sigma.se,  'numeric')
+  expect_type(result$xi.se,  'numeric')
 
   expect_gte(result$Rate,  0)
 })
