+++ R/Diag_Non_Con_Trunc_Sel.R
@@ -3,7 +3,7 @@
 #' Plots demonstrating the goodness of fit of a selected (truncated) non-extreme marginal distribution to a dataset.
 #'
 #' @param Data Numeric vector containing realizations of the variable of interest.
-#' @param Selected Character vector of length one specifying the chosen distribution, options are the Birnbaum-Saunders \code{"BS"}, exponential \code{"Exp"}, two-parameter gamma \code{"Gam(2)"}, three-parameter gamma \code{"Gam(3)"}, mixed two-parameter gamma \code{"GamMix(2)"}, mixed three-parameter gamma \code{"GamMix(3)"}, lognormal \code{"LogN"}, Tweedie \code{"Twe"} and Weibull \code{"Weib"}.
+#' @param Selected Character vector of length one specifying the chosen distribution, options are the Birnbaum-Saunders \code{"BS"}, exponential \code{"Exp"}, two-parameter gamma \code{"Gam(2)"}, three-parameter gamma \code{"Gam(3)"}, mixed two-parameter gamma \code{"GamMix(2)"}, mixed three-parameter gamma \code{"GamMix(3)"}, lognormal \code{"LNorm"}, Tweedie \code{"Twe"} and Weibull \code{"Weib"}.
 #' @param Omit Character vector specifying any distributions that are not to be tested. Default \code{"NA"}, all distributions are fit.
 #' @param x_lab Character vector of length one specifying the label on the x-axis of histogram and cummulative distribution plot.
 #' @param y_lim_min Numeric vector of length one specifying the lower y-axis limit of the histogram. Default is \code{0}.
@@ -141,10 +141,10 @@ Diag_Non_Con_Trunc_Sel<-function(Data,Selected,Omit=NA,x_lab="Data",y_lim_min=0,
   mypalette<-c("Black",brewer.pal(9,"Set1"))
 
   #Distributions to test
-  Dist<-c("BS","Exp","Gam(2)","Gam(3)","GamMix(2)","GamMix(3)","LNorm","TNorm","Twe","Weib")
-  Test<-1:10
+  Dist<- c("BS","Exp","Gam(2)","Gam(3)","GamMix(2)","GamMix(3)","LNorm","TNorm","Twe","Weib")
+  Test<- Dist
   if(is.na(Omit)==F){
-    Test<-Test[-which(Dist %in% Omit)]
+    Test<-Test[-which(Test %in% Omit)]
   }
 
   #AIC result objects
@@ -160,57 +160,52 @@ Diag_Non_Con_Trunc_Sel<-function(Data,Selected,Omit=NA,x_lab="Data",y_lim_min=0,
   AIC.Weib<-NA
 
   #AIC
-  if(any(Test==1)){
+  if(any(Test=="BS")){
     bdata2 <- data.frame(shape = exp(-0.5), scale = exp(0.5))
     bdata2 <- transform(bdata2, y = Data)
     fit <- vglm(y ~ 1, bisa, data = bdata2, trace = FALSE)
     AIC.BS<-2*length(coef(fit))-2*logLik(fit)
   }
 
-  if(any(Test==2)){
+  if(any(Test=="Exp")){
     fit<-fitdistr(Data,"exponential")
[…]
