### 2025-07-12 67e49a7
Delete tests/testthat/test-Diag_Non_Con_Trunc_Sel.R
