+++ R/return_curve_est.R
@@ -394,7 +394,7 @@ for(i in 1:n_grad){
 
 colnames(median) <- c(names(data)[2],names(data)[3])
 colnames(median2) <- c(names(data)[2],names(data)[3])
-colnames(upper_bound) <- c(names(data)[2],names(data)[3])
+colnames(lower_bound) <- c(names(data)[2],names(data)[3])
 colnames(lower_bound2) <- c(names(data)[2],names(data)[3])
 colnames(upper_bound) <- c(names(data)[2],names(data)[3])
 colnames(upper_bound2) <- c(names(data)[2],names(data)[3])
