+++ R/GPD_Threshold_Solari.R
@@ -54,20 +54,96 @@
 #' GPD_Threshold_Solari(Event=Rainfall_Declust_SW$Declustered,
 #'                      Data=22.Detrend.df[,2])
 GPD_Threshold_Solari<-function(Event,Data,RPs=c(10,50,100,500,1000),RPs_PLOT=c(2,3,4),Min_Quantile=0.95,Alpha=0.1,mu=365.25,N_Sim=10){
-  
+
+    # Check for missing required parameters
+    if (is.null(Event)) {
+      stop("Data is missing.")
+    }
+
+    # Check data types
+    if (!is.numeric(Event) || length(Event) == 0) {
+      stop("Event must be a numeric vector.")
+    }
+
+    if (!is.numeric(Data) || length(Data) == 0) {
+      stop("Data must be a numeric vector.")
+    }
+
+    if (!is.numeric(RPs) || length(RPs) == 0) {
+      stop("RPs must be a numeric vector.")
+    }
+
+    if (!is.numeric(RPs_PLOT)) {
+      stop("RPs_PLOT must be a numeric vector.")
+    }
+
+    # Check for all NA values before na.omit
+    if (all(is.na(Event))) {
+      stop("Event contains only NA values.")
+    }
+
+    if (all(is.na(Data))) {
+      stop("Data contains only NA values.")
+    }
+
+    # Check for infinite values
+    if (any(is.infinite(Event))) {
+      warning("Event contains infinite values which will be removed.")
+      Event <- Event[!is.infinite(Event)]
+    }
+
+    if (any(is.infinite(Data))) {
+      warning("Data contains infinite values which will be removed.")
+      Data <- Data[!is.infinite(Data)]
+    }
+
+    # Check Min_Quantile parameter
+    if (!is.numeric(Min_Quantile) || length(Min_Quantile) != 1 || Min_Quantile <= 0 || Min_Quantile >= 1) {
+      stop("Min_Quantile must be a single numeric value must be between 0 and 1 (exclusive).")
+    }
+
+    if (!is.numeric(Alpha) || length(Alpha) != 1 || Alpha <= 0 || Alpha >= 1) {
+      stop("Alpha must be between 0 and 1 (exclusive).")
+    }
+
+    # Check mu parameter
+    if (!is.numeric(mu) || length(mu) != 1 || mu <= 0) {
+      stop("mu must be a single positive (greater than 0) numeric value.")
+    }
+
+    # Check N_Sim parameter
+    if (!is.numeric(N_Sim) || length(N_Sim) != 1 || N_Sim != round(N_Sim) || N_Sim <= 0) {
+      stop("N_Sim must be a single positive (integer) numeric value.")
+    }
+
+    # Check RPs parameter
+    if (any(RPs <= 0)) {
+      stop("All values in RPs must be positive.")
+    }
+
+    # Check RPs_PLOT parameter
+    if (any(RPs_PLOT <= 0) || any(RPs_PLOT != round(RPs_PLOT))) {
+      stop("All values in RPs_PLOT must be positive integers.")
+    }
+
+    if (any(RPs_PLOT > length(RPs))) {
+      stop("RPs_PLOT indices cannot exceed the length of 'RPs'.")
+    }
+
+  # Check for required exter
[…]
