+++ R/HT04.R
@@ -37,11 +37,84 @@
 #'       upper.panel=NULL,pch=16)
 HT04<-function(data_Detrend_Dependence_df,data_Detrend_Declustered_df,u_Dependence,Migpd,mu=365.25,N=100,Margins="gumbel",V=10,Maxit=10000){
 
-  if(class(data_Detrend_Dependence_df[,1])[1]=="Date" | class(data_Detrend_Dependence_df[,1])[1]=="factor" | class(data_Detrend_Dependence_df[,1])[1]=="POSIXct" | class(data_Detrend_Dependence_df[,1])[1]=="character"){
-    data_Detrend_Dependence_df<-data_Detrend_Dependence_df[,-1]
+  # Input Validation
+
+  # Check if data frames exist and are valid
+  if(missing(data_Detrend_Dependence_df) || is.null(data_Detrend_Dependence_df) || !is.data.frame(data_Detrend_Dependence_df)){
+    stop("data_Detrend_Dependence_df must be a data frame.")
   }
-  if(class(data_Detrend_Declustered_df[,1])[1]=="Date" | class(data_Detrend_Declustered_df[,1])[1]=="factor" | class(data_Detrend_Declustered_df[,1])[1]=="POSIXct" |  class(data_Detrend_Declustered_df[,1])[1]=="character"){
-    data_Detrend_Declustered_df<-data_Detrend_Declustered_df[,-1]
+
+  if(missing(data_Detrend_Declustered_df) || is.null(data_Detrend_Declustered_df) || !is.data.frame(data_Detrend_Declustered_df)){
+    stop("data_Detrend_Declustered_df must be a data frame.")
+  }
+
+  # Check if data frames have matching dimensions after removing date/factor columns
+  temp_dep <- data_Detrend_Dependence_df
+  temp_declust <- data_Detrend_Declustered_df
+
+  if(class(temp_dep[,1])[1] %in% c("Date", "factor", "POSIXct", "character")){
+    temp_dep <- temp_dep[,-1]
+  }
+  if(class(temp_declust[,1])[1] %in% c("Date", "factor", "POSIXct", "character")){
+    temp_declust <- temp_declust[,-1]
+  }
+
+  if(ncol(temp_dep) != ncol(temp_declust)){
+    stop("Data frames must have the same number of numeric columns after removing date/factor columns.")
+  }
+
+  if(!all(colnames(temp_dep) == colnames(temp_declust))){
+    stop("Column names must match between data frames after removing date/factor columns.")
+  }
+
+  # Validate u_Dependence
+  if(missing(u_Dependence) || is.null(u_Dependence) || !is.numeric(u_Dependence) || length(u_Dependence) != 1 || u_Dependence <= 0 || u_Dependence >= 1){
+    stop("u_Dependence must be a single numeric value between 0 and 1 (exclusive).")
+  }
+
+  # Validate Migpd
+  if(missing(Migpd) || is.null(Migpd) || !is.list(Migpd)){
+    stop("Migpd must be a list object (migpd class).")
+  }
+
+  if(is.null(Migpd$models)){
+    stop("Migpd$models is NULL. Please ensure Migpd object contains fitted models.")
+  }
+
+  if(length(Migpd$models) != ncol(temp_dep)){
+    stop("Number of models in Migpd$models (", length(Migpd$models),
+         ") must match number of data columns (", ncol(temp_dep), ").")
+  }
+
+  # Validate numeric parameters
+  if(!is.numeric(mu) || length(mu) != 1 || mu <= 0){
[…]
