+++ R/return_curve_diag.R
@@ -63,148 +63,227 @@
 return_curve_diag = function(data,q,rp,mu,n_sim,n_grad,n_boot,boot_method, boot_replace, block_length, boot_prop, decl_method_x, decl_method_y, window_length_x, window_length_y, u_x=NA, u_y=NA, sep_crit_x=NA, sep_crit_y=NA, boot_method_all="block", boot_replace_all=NA, block_length_all=14, boot_prop_all=0.8,alpha=0.1, x_lab=NA, y_lab=NA,x_lim_min=min(data_df[,2],na.rm=T),x_lim_max=max(data_df[,2],na.rm=T)+0.3*diff(range(data[,2],na.rm=T)),y_lim_min=min(data[,3],na.rm=T),y_lim_max=max(data[,3],na.rm=T)+0.3*diff(range(data[,2],na.rm=T))){
 
   if (missing(data) || is.null(data)) {
-    stop("Error: data missing")
+    stop("data is missing.")
   }
 
   if (!is.data.frame(data)) {
-    stop("Error: data must be a data.frame")
+    stop("data must be a data.frame.")
   }
 
   #Date column validation
   if (!inherits(data[,1], c("Date", "POSIXct", "POSIXt"))) {
-    stop("Error: First column of data must be Date/POSIXct format")
+    stop("First column of data must be Date/POSIXct format.")
   }
 
 
   #Quantile paramter ust be numeric and between 0 and 1
-  if (missing(q) || !is.numeric(q) || length(q) != 1) {
-    stop("Error: q must be a single numeric value")
+  if (missing(q)) {
+    stop("q is missing.")
+  }
+
+  if (!is.numeric(q) ){
+    stop("q must be a numeric value.")
+  }
+
+  if (length(q) != 1){
+    stop("q must be a single numeric value.")
   }
 
   if (q <= 0 || q >= 1) {
-    stop("Error: q must be between 0 and 1 (exclusive)")
+    stop("q must be between 0 and 1 (exclusive)")
   }
 
   #Return period parameter must be a single numeric value
-  if (missing(rp) || !is.numeric(rp) || length(rp) != 1 || rp < 1) {
-    stop("Error: rp must be a single positive numeric value")
+  if (missing(rp)) {
+    stop("rp is missing.")
+  }
+
+  if(!is.numeric(rp)){
+    stop("rp must be a a numeric value.")
+  }
+
+  if (length(rp) != 1 || rp < 1){
+    stop("rp must be a positive numeric value.")
   }
 
   #Rate parameter must be a single positive numeric value
-  if (missing(mu) || !is.numeric(mu) || length(mu) != 1 || mu <= 0) {
-    stop("Error: mu (rate parameter) must be a single positive numeric value")
+  if (missing(mu)) {
+    stop("mu (rate parameter) is missing.")
   }
+
+  if (!is.numeric(mu)){
+    stop("mu (rate parameter) must be a numeric value")
+  }
+
+  if (mu <= 0){
+    stop("mu (rate parameter) must be a positive numeric value.")
+  }
+
 
   #Simulation parameters
-  if (!is.numeric(n_sim) || length(n_sim) != 1 || n_sim <= 0) {
-    stop("Error: n_sim must be a positive integer")
+  if (!is.numeric(n_sim)) {
+    stop("n_sim must be a numeric value.")
   }
 
-  if (!is.numeric(n_grad) || length(n_grad) != 1 || n_grad <= 0) {
-    stop("Error: n_grad must be a positive integer")
+  if (length(n_sim) != 1 ) {
+    stop("n_sim must be an integer.")
   }
 
[…]
