+++ tests/testthat/test-return_curve_diag.R
@@ -44,3 +44,37 @@ test_that("return_curve_est_diag works", {
   expect_equal(colnames(S22.Detrend.df.extended)[-1], colnames(result$ub_ht04))
   expect_equal(colnames(S22.Detrend.df.extended)[-1], colnames(result$lb_ht04))
 })
+
+
+test_that("return_curve_est_diag invalid inputs", {
+
+  expect_error(
+      return_curve_diag(data=S22.Detrend.df.extended[,1:3],
+                               q=0.985,rp=1,mu=365.25,n_sim=100,
+                               n_grad=50,n_boot=100,boot_method="monthly",
+                               boot_replace=NA, block_length=NA, boot_prop=0.8,
+                               decl_method_x="runs", decl_method_y="runs",
+                               window_length_x=NA,window_length_y=NA,
+                               u_x=1.5, u_y=0.95,
+                               sep_crit_x=36, sep_crit_y=36,
+                               alpha=0.1,
+                               boot_method_all="block", boot_replace_all=NA,
+                               block_length_all=14),
+             "u_x threshold must be between 0 and 1.")
+
+  expect_error(
+    return_curve_diag(data=S22.Detrend.df.extended[,1:3],
+                      q=0.985,mu=365.25,n_sim=100,
+                      n_grad=50,n_boot=100,boot_method="monthly",
+                      boot_replace=NA, block_length=NA, boot_prop=0.8,
+                      decl_method_x="runs", decl_method_y="runs",
+                      window_length_x=NA,window_length_y=NA,
+                      u_x=0.95, u_y=0.95,
+                      sep_crit_x=36, sep_crit_y=36,
+                      alpha=0.1,
+                      boot_method_all="block", boot_replace_all=NA,
+                      block_length_all=14),
+               " rp is missing.")
+
+})
+
