+++ tests/testthat/test-HT04.R
@@ -13,23 +13,23 @@ test_that("HT04 works", {
 
 
   #Checking length of outputs
-  expect_equal(nrow(result$Model),50)
-  expect_equal(nrow(result$Prop),1)
-  expect_equal(nrow(result$z),round((1-0.995)*365.25*1000,0))
+  expect_equal(length(result$Model),3)
+  expect_equal(length(result$Prop),3)
+  expect_equal(length(result$z),3)
   expect_equal(nrow(result$u.sim),round((1-0.995)*365.25*1000,0))
   expect_equal(nrow(result$x.sim),round((1-0.995)*365.25*1000,0))
 
   #Checking column names of outputs
   expect_equal(paste(colnames(S20.Detrend.df[,2:4]),".trans",sep=""), colnames(result$u.sim))
   expect_equal(colnames(S20.Detrend.df[,2:4]), colnames(result$x.sim))
-  expect_equal(colnames(S20.Detrend.df[,2:4]), colnames(result$z))
+  expect_equal(colnames(S20.Detrend.df[,2:4]), names(result$z))
 })
 
 test_that("Invalid inputs", {
 
   expect_error(
     HT04(data_Detrend_Dependence_df=S20.Detrend.df,
-         data_Detrend_Declustered_df=S20.Detrend.Declustered.df[,1:2],
+         data_Detrend_Declustered_df=S20.Detrend.Declustered.df[,1:3],
          u_Dependence=0.995,Migpd=S20.Migpd,mu=365.25,N=1000),
     "Data frames must have the same number of numeric columns after removing date/factor columns.")
 
