+++ R/Design_Event_2D.R
@@ -23,7 +23,7 @@
 #' @param mu Numeric vector of length one specifying the (average) occurrence frequency of events in \code{Data}. Default is \code{365.25}, daily data.
 #' @param GPD_Bayes Logical; indicating whether to use a Bayesian approach to estimate GPD parameters. This involves applying a penalty to the likelihood to aid in the stability of the optimization procedure. Default is \code{FALSE}.
 #' @param RP Numeric vector specifying the return periods of interest.
-#' @param Decimal_Palace Numeric vector specifying the number of decimal places to which to specify the isoline. Default is \code{2}.
+#' @param Decimal_Place Numeric vector specifying the number of decimal places to which to specify the isoline. Default is \code{2}.
 #' @param Interval Numeric vector specifying the number of equally spaced points comprising the combined isoline.
 #' @param x_lab Character vector specifying the x-axis label.
 #' @param y_lab Character vector specifying the y-axis label.
@@ -65,6 +65,140 @@
 #'Design.Event$`100`$Isoline
 Design_Event_2D<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Thres2=NA, Copula_Family1, Copula_Family2, Marginal_Dist1, Marginal_Dist2, Marginal_Dist1_Par=NA, Marginal_Dist2_Par=NA, Con1="Rainfall",Con2="OsWL", GPD1=NA, GPD2=NA, Rate_Con1=NA, Rate_Con2=NA, Tab1= NA, Tab2 = NA, mu=365.25, GPD_Bayes=FALSE, Decimal_Place=2, RP, Interval=10000, End=F, Resolution="Low", x_lab="Rainfall (mm)",y_lab="O-sWL (mNGVD 29)",x_lim_min = NA,x_lim_max = NA,y_lim_min = NA,y_lim_max = NA,Isoline_Probs="Sample", N=10^6,N_Ensemble=0,Sim_Max=10,Plot_Quantile_Isoline=FALSE,Isoline_Type="Combined"){
 
+  #Validation of inputs
+  if (!is.data.frame(Data) || ncol(Data) != 2) {
+    stop("Data must be a data frame with exactly 2 columns.")
+  }
+
+  if (!is.data.frame(Data_Con1) || !is.data.frame(Data_Con2)) {
+    stop("Data_Con1 and Data_Con2 must be data frames.")
+  }
+
+  # Threshold validation
+  if (sum(c(!is.na(u1), !is.na(Thres1), !is.null(GPD1), !is.null(Tab1))) != 1) {
+    stop("Exactly one of u1, Thres1, GPD1, or Tab1 must be provided.")
+  }
+
+  if (sum(c(!is.na(u2), !is.na(Thres2), !is.null(GPD2), !is.null(Tab2))) != 1) {
+    stop("Exactly one of u2, Thres2, GPD2, or Tab2 must be provided.")
+  }
+
+  if (!is.na(u1) && (u1 <= 0 || u1 >= 1)) {
+    stop("Threshold u1 must be between 0 and 1 (exclusive).")
+  }
+
+  if (!is.na(u2) && (u2 <= 0 || u2 >= 1)) {
+    stop("Threshold u2 must be between 0 and 1 (exclusive).")
+  }
+
+  # Copula family validation
+  valid_copulas <- c(1:10, 13, 14, 16, 17, 18, 19, 20, 23, 24, 26, 27, 28, 29, 30, 33, 34, 36, 37, 38, 39, 40)
+  if (!Copula_Family1 %in% valid_copulas) {
+    stop("Invalid Copula_Family1.")
+  }
+
[…]
