+++ R/return_curve_est.R
@@ -61,152 +61,176 @@
 return_curve_est = function(data,q,rp,mu,n_sim,n_grad,n_boot,boot_method, boot_replace, block_length, boot_prop, decl_method_x, decl_method_y, window_length_x, window_length_y, u_x=NA, u_y=NA, sep_crit_x=NA, sep_crit_y=NA, alpha=0.1, most_likely=F, n_interp=1000, n=10^6, n_ensemble=0, x_lab=colnames(data)[2], y_lab=colnames(data)[3],x_lim_min=min(data[,2],na.rm=T),x_lim_max=max(data[,2],na.rm=T)+0.3*diff(range(data[,2],na.rm=T)),y_lim_min=min(data[,3],na.rm=T),y_lim_max=max(data[,3],na.rm=T)+0.3*diff(range(data[,3],na.rm=T)),plot=T){
 
   if (missing(data) || is.null(data)) {
-    stop("Error: data missing")
+    stop("data is missing.")
   }
 
   if (!is.data.frame(data)) {
-    stop("Error: data must be a data.frame")
+    stop("data must be a data.frame.")
   }
 
    #Date column validation
   if (!inherits(data[,1], c("Date", "POSIXct", "POSIXt"))) {
-    stop("Error: First column of data must be Date/POSIXct format")
+    stop("First column of data must be Date/POSIXct format.")
   }
 
 
-  #Quantile paramter ust be numeric and between 0 and 1
-  if (missing(q) || !is.numeric(q) || length(q) != 1) {
-    stop("Error: q must be a single numeric value")
+  #Quantile parameter must be numeric and between 0 and 1
+  if (missing(q)) {
+    stop("q is missing.")
+  }
+
+  if (!is.numeric(q) || length(q) != 1) {
+    stop("q must be a single numeric value.")
   }
 
   if (q <= 0 || q >= 1) {
-    stop("Error: q must be between 0 and 1 (exclusive)")
+    stop("q must be between 0 and 1 (exclusive).")
   }
 
   #Return period parameter must be a single numeric value
-  if (missing(rp) || !is.numeric(rp) || length(rp) != 1 || rp <= 1) {
-    stop("Error: rp must be a single positive numeric value")
+  if (missing(rp)) {
+    stop("rp is missing.")
+  }
+
+  if (!is.numeric(rp) || length(rp) != 1 || rp < 1) {
+    stop("rp must be a single positive numeric value.")
   }
 
   #Rate parameter must be a single positive numeric value
-  if (missing(mu) || !is.numeric(mu) || length(mu) != 1 || mu <= 0) {
-    stop("Error: mu (rate parameter) must be a single positive numeric value")
+  if (missing(mu)) {
+    stop("mu (rate parameter) is missing.")
+  }
+
+  if (!is.numeric(mu) || length(mu) != 1 || mu <= 0) {
+    stop("mu (rate parameter) must be a single positive numeric value.")
   }
 
   #Simulation parameters
   if (!is.numeric(n_sim) || length(n_sim) != 1 || n_sim <= 0) {
-    stop("Error: n_sim must be a positive integer")
+    stop("n_sim must be a positive integer.")
   }
 
   if (!is.numeric(n_grad) || length(n_grad) != 1 || n_grad <= 0) {
-    stop("Error: n_grad must be a positive integer")
+    stop("n_grad must be a positive integer.")
   }
 
   if (!is.numeric(n_boot) || length(n_boot) != 1 || n_boot <= 0 || n_boot %% 2 != 0) {
-    stop("Error: n_boot must be a positive even integer")
+    stop("n_boot must be a positive even integer.")
   }
 
 
   # Bootstrap method validation
[…]
