+++ tests/testthat/test-HT04.R
@@ -38,7 +38,6 @@ test_that("Invalid inputs", {
   expect_error(
     HT04(data_Detrend_Dependence_df=S20.Detrend.df,
          data_Detrend_Declustered_df=S20.Detrend.Declustered.df,
-         u_Dependence=0.995,Migpd=S20.Migpd,mu=365.25,N=1000),
-    "Number of models in Migpd$models (", length(Migpd$models),
-    ") must match number of data columns (", ncol(temp_dep), ").")
+         u_Dependence=0.995,Migpd=S20.Migpd.modified,mu=365.25,N=1000),
+    "Number of models in Migpd$models (2) must match number of data columns (3).")
 })
