### 2025-07-17 5e1e841
Update return_curve_diag.R
