+++ man/return_curve_diag.Rd
@@ -31,8 +31,8 @@ return_curve_diag(
   alpha = 0.1,
   x_lab = NA,
   y_lab = NA,
-  x_lim_min = min(data_df[, 2], na.rm = T),
-  x_lim_max = max(data_df[, 2], na.rm = T) + 0.3 * diff(range(data[, 2], na.rm = T)),
+  x_lim_min = min(data[, 2], na.rm = T),
+  x_lim_max = max(data[, 2], na.rm = T) + 0.3 * diff(range(data[, 2], na.rm = T)),
   y_lim_min = min(data[, 3], na.rm = T),
   y_lim_max = max(data[, 3], na.rm = T) + 0.3 * diff(range(data[, 2], na.rm = T))
 )
@@ -54,7 +54,7 @@ return_curve_diag(
 
 \item{boot_method}{Character vector of length one specifying the bootstrap method. Options are \code{"basic"} (default), \code{"block"} or \code{"monthly"}.}
 
-\item{boot_replace}{Character vector of length one specifying whether simple bootstrapping is carried out with \code{"T"} or without \code{"F"} replacement. Only required if \code{boot_method = "basic"}. Default is \code{NA}.}
+\item{boot_replace}{Logical vector of length one specifying whether simple bootstrapping is carried out with \code{T} or without \code{F} replacement. Only required if \code{boot_method = "basic"}. Default is \code{NA}.}
 
 \item{block_length}{Numeric vector of length one specifying block length. Only required if \code{boot_method = "block"}. Default is \code{NA}.}
 
