+++ man/Migpd_Fit.Rd
@@ -43,15 +43,13 @@ Fit multiple independent generalized Pareto models to each column of a data fram
 }
 \examples{
 #With date as first column
-S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df, mqu =c(0.99,0.99,0.99))
-#Without date as first column
-S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df[,-1], mqu =c(0.99,0.99,0.99))
+S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df, Data_Full=S22.Detrend.df, mqu =c(0.99,0.99,0.99))
 #Same GPDs fit as above but thresholds given on the original scale
 S22.Rainfall.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$Rainfall),0.99)
 S22.OsWL.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$OsWL),0.99)
 S22.GW.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$Groundwater),0.99)
 S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df[,-1],
-                   mqu =c(S22.Rainfall.Quantile,S22.OsWL.Quantile,S22.GW.Quantile))
+                   mth =c(S22.Rainfall.Quantile,S22.OsWL.Quantile,S22.GW.Quantile))
 }
 \seealso{
 \code{\link{Decluster}} \code{\link{Detrend}} \code{\link{Dataframe_Combine}}
