+++ R/Design_Event_2D.R
@@ -102,7 +102,7 @@ Design_Event_2D<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Thres2=
   }
 
   # Marginal distribution validation
-  valid_dists <- c("BS","Exp", "Gam(2)", "Gam(3)", "GamMix(2)", "GamMix(3)", "Gaus", "InvG", "Logis", "LNorm", "TNorm","Twe", "Weib")
+  valid_dists <- c("BS","Exp", "Gam(2)", "Gam(3)", "GamMix(2)", "GamMix(3)", "Gaus", "Gum", "InvG", "Lapl", "Logis", "LNorm", "RGum", "TNorm","Twe", "Weib")
   if (!Marginal_Dist1 %in% valid_dists) {
     stop("Invalid Marginal_Dist1. Must be one of: ", paste(valid_dists, collapse = ", "), ".")
   }
@@ -229,14 +229,14 @@ Design_Event_2D<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Thres2=
   ###Fit the 4 marginal distributions (2 GPD and 2 parametric non-extreme value distributions).
 
   #Fit the GPD to the conditioned variable con1 in Data_Con1.
-  if(is.na(GPD1[[1]][1])==T & is.na(Thres1)==T & is.na(Tab1[[1]][1])==T){
+  if(is.na(GPD1[[1]][1])==T & is.na(Thres1)==T & is.null(Tab1)==T){
     Thres1<-quantile(na.omit(Data[,con1]),u1)
   }
 
-  if(is.na(GPD1[[1]][1])==T & GPD_Bayes==T & is.na(Tab1[[1]][1])==T){
+  if(is.na(GPD1[[1]][1])==T & GPD_Bayes==T & is.null(Tab1)==T){
     GPD_con1<-evm(Data_Con1[,con1], th = Thres1,penalty = "gaussian",priorParameters = list(c(0, 0), matrix(c(100^2, 0, 0, 0.25), nrow = 2)))
   }
-  if(is.na(GPD1[[1]][1])==T & GPD_Bayes==F & is.na(Tab1[[1]][1])==T){
+  if(is.na(GPD1[[1]][1])==T & GPD_Bayes==F & is.null(Tab1)==T){
     GPD_con1<-evm(Data_Con1[,con1], th = Thres1)
   }
 
@@ -357,14 +357,14 @@ Design_Event_2D<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Thres2=
   }
 
   #Fit the GPD to the conditioned variable con2 in Data_Con2.
-  if(is.na(GPD2[[1]][1])==T & is.na(Thres2)==T & is.na(Tab2[[1]][1])==T){
+  if(is.na(GPD2[[1]][1])==T & is.na(Thres2)==T & is.null(Tab2)==T){
     Thres2<-quantile(na.omit(Data[,con2]),u2)
   }
 
-  if(is.na(GPD2[[1]][1])==T & GPD_Bayes==T & is.na(Tab2[[1]][1])==T){
+  if(is.na(GPD2[[1]][1])==T & GPD_Bayes==T & is.null(Tab2)==T){
     GPD_con2<-evm(Data_Con2[,con2], th=Thres2 ,penalty = "gaussian",priorParameters = list(c(0, 0), matrix(c(100^2, 0, 0, 0.25), nrow = 2)))
   }
-  if(is.na(GPD2[[1]][1])==T & GPD_Bayes==F & is.na(Tab2[[1]][1])==T){
[…]
