+++ R/Design_Event_2D.R
@@ -229,14 +229,14 @@ Design_Event_2D<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Thres2=
   ###Fit the 4 marginal distributions (2 GPD and 2 parametric non-extreme value distributions).
 
   #Fit the GPD to the conditioned variable con1 in Data_Con1.
-  if(is.na(GPD1[[1]][1])==T & is.na(Thres1)==T & is.null(Tab1)==T){
+  if(is.null(GPD1)==T & is.na(Thres1)==T & is.null(Tab1)==T){
     Thres1<-quantile(na.omit(Data[,con1]),u1)
   }
 
-  if(is.na(GPD1[[1]][1])==T & GPD_Bayes==T & is.null(Tab1)==T){
+  if(is.null(GPD1)==T & GPD_Bayes==T & is.null(Tab1)==T){
     GPD_con1<-evm(Data_Con1[,con1], th = Thres1,penalty = "gaussian",priorParameters = list(c(0, 0), matrix(c(100^2, 0, 0, 0.25), nrow = 2)))
   }
-  if(is.na(GPD1[[1]][1])==T & GPD_Bayes==F & is.null(Tab1)==T){
+  if(is.null(GPD1)==T & GPD_Bayes==F & is.null(Tab1)==T){
     GPD_con1<-evm(Data_Con1[,con1], th = Thres1)
   }
 
@@ -357,14 +357,14 @@ Design_Event_2D<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Thres2=
   }
 
   #Fit the GPD to the conditioned variable con2 in Data_Con2.
-  if(is.na(GPD2[[1]][1])==T & is.na(Thres2)==T & is.null(Tab2)==T){
+  if(is.null(GPD2)==T & is.na(Thres2)==T & is.null(Tab2)==T){
     Thres2<-quantile(na.omit(Data[,con2]),u2)
   }
 
-  if(is.na(GPD2[[1]][1])==T & GPD_Bayes==T & is.null(Tab2)==T){
+  if(is.null(GPD2)==T & GPD_Bayes==T & is.null(Tab2)==T){
     GPD_con2<-evm(Data_Con2[,con2], th=Thres2 ,penalty = "gaussian",priorParameters = list(c(0, 0), matrix(c(100^2, 0, 0, 0.25), nrow = 2)))
   }
-  if(is.na(GPD2[[1]][1])==T & GPD_Bayes==F & is.null(Tab2)==T){
+  if(is.null(GPD2)==T & GPD_Bayes==F & is.null(Tab2)==T){
     GPD_con2<-evm(Data_Con2[,con2], th= Thres2)
   }
 
@@ -478,10 +478,10 @@ Design_Event_2D<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Thres2=
   #Simulate a sample from the fitted copula. Out of the sample size 'N' the proportion of the sample from the copula associated with Data_Con1 is proportional to the size of Data_Con1 relative to Data_Con2.
   sample<-BiCopSim(round(N*nrow(Data_Con1)/(nrow(Data_Con1)+nrow(Data_Con2)),0),obj1)
   #Transform the realizations of the conditioned variable con1 to the original scale using inverse cumulative distribution a.k.a. quantile functions (inverse probability integral transform) of the GPD contained in the u2gpd function.
-  if(is.na(GPD1[[1]][1])==T & is.null(Tab1)==T){
+  if(is.null(GPD1)==T & is.null(Tab1)==T){
    cop.sample1.con<-u2gpd(sample[,con1], p = 1, th=Thres1 , sigma=exp(GPD_con1$coefficients[1]),xi= GPD_con1$coefficients[2])
   }
[…]
