+++ R/Design_Event_2D.R
@@ -291,7 +291,7 @@ Design_Event_2D<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Thres2=
   if(Marginal_Dist1 == "GamMix(2)"){
     if(is.na(Marginal_Dist1_Par)==T){
       data.gamlss<-data.frame(X=Data_Con1[,con2])
-      marginal_non_con1 <- tryCatch(gamlssMX(X~1, data=data.gamlss, family=GA, K=2),
+      marginal_non_con1 <- tryCatch(gamlssMX(X~1, data=data.gamlss, family=GA, K=2, trace=FALSE),
                                     error = function(e) "error")
     }else{
       marginal_non_con1<-Marginal_Dist1_Par
@@ -300,7 +300,7 @@ Design_Event_2D<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Thres2=
   if(Marginal_Dist1 == "GamMix(3)"){
     if(is.na(Marginal_Dist1_Par)==T){
       data.gamlss<-data.frame(X=Data_Con1[,con2])
-      marginal_non_con1 <- tryCatch(gamlssMX(X~1, data=data.gamlss, family=GA, K=3),
+      marginal_non_con1 <- tryCatch(gamlssMX(X~1, data=data.gamlss, family=GA, K=3, trace=FALSE),
                                     error = function(e) "error")
     }else{
       marginal_non_con1<-Marginal_Dist1_Par
@@ -313,9 +313,23 @@ Design_Event_2D<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Thres2=
       marginal_non_con1<-Marginal_Dist1_Par
     }
   }
+  if(Marginal_Dist1 == "Gum"){
+    if(is.na(Marginal_Dist1_Par)==T){
+      marginal_non_con1 <- gamlss(Data_Con1[,con2]  ~ 1, family= GU, trace=FALSE)
+    }else{
+      marginal_non_con1<-Marginal_Dist1_Par
+    }
+  }
   if(Marginal_Dist1 == "InvG"){
     if(is.na(Marginal_Dist1_Par)==T){
       marginal_non_con1<-fitdist(Data_Con1[,con2], "invgauss", start = list(mean = 5, shape = 1))
+    }else{
+      marginal_non_con1<-Marginal_Dist1_Par
+    }
+  }
+  if(any(Marginal_Dist1=="Lapl")){
+    if(is.na(Marginal_Dist1_Par)==T){
+      marginal_non_con1 <- fitdistr(Data_Con1[,con2], dlaplace, start=list(location=mean(Data_Con1[,con2]), scale=sd(Data_Con1[,con2])/sqrt(2)))
     }else{
       marginal_non_con1<-Marginal_Dist1_Par
     }
@@ -329,11 +343,20 @@ Design_Event_2D<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Thres2=
   }
   if(Marginal_Dist1 == "LNorm"){
     if(is.na(Marginal_Dist1_Par)==T){
-      marginal_non_con1<-fitdistr(Data_Con1[,con2],"LNormormal")
+      marginal_non_con1<-fitdistr(Data_Con1[,con2],"lognormal")
     }else{
       marginal_non_con1<-Marginal_Dist1_Par
     }
   }
+
+  if(Marginal_Dist1 == "RGum"){
+    if(is.na(Marginal_Dist1_Par)==T){
+      marginal_non_con1 <- gamlss(Data_Con1[,con2] ~ 1,family=RG, trace=FALSE)
+    }else{
[…]
