+++ R/Design_Event_2D_Multi_Pop.R
@@ -32,22 +32,22 @@
 #' @param Con2 Character vector of length one specifying the name of variable in the second column of \code{Data_Con2}.
 #' @param Con1 Character vector of length one specifying the name of variable in the first column of \code{Data_Con3}.
 #' @param Con2 Character vector of length one specifying the name of variable in the second column of \code{Data_Con4}.
-#' @param GPD1 Output of \code{GPD_Fit} applied to variable \code{con1} i.e., GPD fit \code{con1}. Default \code{NA}. Only one of \code{u1}, \code{Thres1}, \code{GPD1} and \code{Tab1} is required.
-#' @param GPD2 Output of \code{GPD_Fit} applied to variable \code{con2} i.e., GPD fit \code{con2}. Default \code{NA}. Only one of \code{u2}, \code{Thres2}, \code{GPD2} and \code{Tab2} is required.
-#' @param GPD3 Output of \code{GPD_Fit} applied to variable \code{con3} i.e., GPD fit \code{con3}. Default \code{NA}. Only one of \code{u3}, \code{Thres3}, \code{GPD3} and \code{Tab3} is required.
-#' @param GPD4 Output of \code{GPD_Fit} applied to variable \code{con4} i.e., GPD fit \code{con4}. Default \code{NA}. Only one of \code{u4}, \code{Thres4}, \code{GPD4} and \code{Tab4} is required.
+#' @param GPD1 Output of \code{GPD_Fit} applied to variable \code{con1} i.e., GPD fit \code{con1}. Default \code{NULL}. Only one of \code{u1}, \code{Thres1}, \code{GPD1} and \code{Tab1} is required.
+#' @param GPD2 Output of \code{GPD_Fit} applied to variable \code{con2} i.e., GPD fit \code{con2}. Default \code{NULL}. Only one of \code{u2}, \code{Thres2}, \code{GPD2} and \code{Tab2} is required.
+#' @param GPD3 Output of \code{GPD_Fit} applied to variable \code{con3} i.e., GPD fit \code{con3}. Default \code{NULL}. Only one of \code{u3}, \code{Thres3}, \code{GPD3} and \code{Tab3} is required.
+#' @param GPD4 Output of \code{GPD_Fit} applied to variable \code{con4} i.e., GPD fit \code{con4}. Default \code{NULL}. Only one of \code{u4}, \code{Thres4}, \code{GPD4} and \code{Tab4} is required.
 #' @param Rate_Con1 Numeric vector of length one specifying the occurrence rate of observations in \code{Data_Con1}. Default is \code{NA}.
 #' @param Rate_Con2 Numeric vector of length one specifying the occurrence rate of observations in \code{Data_Con2}. Default is \code{NA}.
 #' @param Rate_Con3 Numeric vector of length one specifying the occurrence rate of observations in \code{Data_Con3}. Default is \code{NA}.
 #' @param Rate_Con4 Numeric vector of length one specifying the occurrence rate of observations in \code{Data_Con4}. Default is \code{NA}.
-#' @param Tab1 Data frame specifying the return periods of variable \code{con1}, when conditioning on \code{con1}. First column specifies the return period and the second column gives the corresponding levels. First row must contain the return level of \code{con1} for the inter-arrival time (1/rate) of the sample.
[…]
