+++ R/HT04_Lag.R
@@ -12,7 +12,7 @@
 #' }
 #' @param Migpd An \code{Migpd} object, containing the parameterized Pareto models fitted (independently) to each of the variables.
 #' @param u_Dependence Dependence quantile. Specifies the (sub-sample of) data to which the dependence model is fitted, that for which the conditioning variable exceeds the threshold associated with the prescribed quantile. Default is \code{0.7}, thus the dependence parameters are estimated using the data with the highest \code{30\%} of values of the conditioning variables.
-#' @param Lag Matrix specifying the lags. The no lag i.e. \code{0} lag cases need to be specified. Row n denotes the lags applied to the variable in the nth column of \code{data_Detrend_Dependence_df}. Column n corresponds to the nth largest lag applied to any variable. \code{NA}. Default is \code{matrix(c(0,1,0,NA),nrow=2,byrow = T)}, which corresponds to a lag of 1 being applied to variable in the first column of \code{data_Detrend_Dependence_df} and no lag being applied to the variable in the second column of \code{data_Detrend_Dependence_df}.
+#' @param Lags Matrix specifying the lags. The no lag i.e. \code{0} lag cases need to be specified. Row n denotes the lags applied to the variable in the nth column of \code{data_Detrend_Dependence_df}. Column n corresponds to the nth largest lag applied to any variable. Default is \code{matrix(c(NA,0,1,NA,0,1,NA,0,NA),nrow=3,byrow = T)}, which corresponds to a lag of 1 being applied to variables in the first and second columns of \code{data_Detrend_Dependence_df} and no lag being applied to the variable in the third column of \code{data_Detrend_Dependence_df}.
 #' @param Margins Character vector specifying the form of margins to which the data are transformed for carrying out dependence estimation. Default is \code{"gumbel"}, alternative is \code{"laplace"}. Under Gumbel margins, the estimated parameters \code{a} and \code{b} describe only positive dependence, while \code{c} and \code{d} describe negative dependence in this case. For Laplace margins, only parameters \code{a} and \code{b} are estimated as these capture both positive and negative dependence.
 #' @param V See documentation for \code{mexDependence}.
 #' @param Maxit See documentation for \code{mexDependence}.
@@ -20,16 +20,131 @@
 #' @seealso \code{\link{Dataframe_Combine}} \code{\link{Decluster}} \code{\link{GPD_Fit}} \code{\link{Migpd_Fit}}
 #' @export
 #' @examples
-#' HT04(data_Detrend_Dependence_df = S22.Detrend.df,
-#' data_Detrend_Declustered_df = S22.Detrend.Declustered.df,
-#' Migpd = S22_GPD, u_Dependence=0.7,Margins = "gumbel")
+#' HT04_Lag(data_Detrend_Dependence_df = S22.Detrend.df,
+#'      data_Detrend_Declustered_df = S22.Detrend.Declustered.df,
+#'      Lags = matrix(c(NA,0,1,NA,0,1,NA,0,NA),nrow=3,byrow = T),
+#'      Migpd = S22_GPD, u_Dependence=0.7,Margins = "gumbel")
 HT04_Lag<-function (data_Detrend_Dependence_df, data_Detrend_Declustered_df, Lags, u_Dependence, Migpd, mu = 365.25, N = 100, Margins = "gumbel",V = 10, Maxit = 10000){
 
[…]
