+++ man/Design_Event_2D_Grid.Rd
@@ -21,12 +21,12 @@ Design_Event_2D_Grid(
   Marginal_Dist2_Par = NA,
   Con1 = "Rainfall",
   Con2 = "OsWL",
-  GPD1 = NA,
-  GPD2 = NA,
+  GPD1 = NULL,
+  GPD2 = NULL,
   Rate_Con1 = NA,
   Rate_Con2 = NA,
-  Tab1 = NA,
-  Tab2 = NA,
+  Tab1 = NULL,
+  Tab2 = NULL,
   mu = 365.25,
   GPD_Bayes = FALSE,
   Decimal_Place = 2,
@@ -82,21 +82,23 @@ Design_Event_2D_Grid(
 
 \item{Con2}{Character vector of length one specifying the name of variable in the second column of \code{Data}.}
 
-\item{GPD1}{Output of \code{GPD_Fit} applied to variable \code{con1} i.e., GPD fit \code{con1}. Default \code{NA}. Only one of \code{u1}, \code{Thres1}, \code{GPD1} and \code{Tab1} is required.}
+\item{GPD1}{Output of \code{GPD_Fit} applied to variable \code{con1} i.e., GPD fit \code{con1}. Default \code{NULL}. Only one of \code{u1}, \code{Thres1}, \code{GPD1} and \code{Tab1} is required.}
 
-\item{GPD2}{Output of \code{GPD_Fit} applied to variable \code{con2} i.e., GPD fit \code{con2}. Default \code{NA}. Only one of \code{u2}, \code{Thres2}, \code{GPD2} and \code{Tab2} is required.}
+\item{GPD2}{Output of \code{GPD_Fit} applied to variable \code{con2} i.e., GPD fit \code{con2}. Default \code{NULL}. Only one of \code{u2}, \code{Thres2}, \code{GPD2} and \code{Tab2} is required.}
 
 \item{Rate_Con1}{Numeric vector of length one specifying the occurrence rate of observations in \code{Data_Con1}. Default is \code{NA}.}
 
 \item{Rate_Con2}{Numeric vector of length one specifying the occurrence rate of observations in \code{Data_Con2}. Default is \code{NA}.}
 
-\item{Tab1}{Data frame specifying the return periods of variable \code{con1}, when conditioning on \code{con1}. First column specifies the return period and the second column gives the corresponding levels. First row must contain the return level of \code{con1} for the inter-arrival time (1/rate) of the sample.}
+\item{Tab1}{Data frame specifying the return periods of variable \code{con1}, when conditioning on \code{con1}. First column specifies the return period and the second column gives the corresponding levels. First row must contain the return level of \code{con1} for the inter-arrival time (1/rate) of the sample. Default is \code{NULL}.}
 
-\item{Tab2}{Data frame specifying the return periods of variable \code{con2}, when conditioning on \code{con2}. First column specifies the return period and the second column gives the corresponding levels. First row must contain the return level of \code{con2} for the inter-arrival time (1/rate) of the sample.}
+\item{Tab2}{Data frame specifying the return periods of variable \code{con2}, when conditioning on \code{con2}. First column specifies the return period and the second column gives the corresponding levels. First row must contain the return level of \code{con2} for the inter-arrival time (1/rate) of the sample. Default is \code{NULL}.}
 
 \item{mu}{Numeric vector of length one specifying the (average) occurrence frequency of events in \code{Data}. Default is \code{365.25}, daily data.}
 
 \item{GPD_Bayes}{Logical; indicating whether to use a Bayesian approach to estimate GPD parameters. This involves applying a penalty to the likelihood to aid in the stability of the optimization procedure. Default is \code{FALSE}.}
+
[…]
