+++ tests/testthat/test-Design_Event_2D_Multi_Pop.R
@@ -1,0 +1,294 @@
+#Append S-22 dataframe with HURDAT data
+S22.Detrend.df = HURDAT(Data=S22.Detrend.df,lat.loc=25.669610,lon.loc=80.283732,rad=500)
+
+#Inter event rainfall
+S22.Detrend.df$Rainfall[which(S22.Detrend.df$Rainfall<0.01)]<-0
+z<-which(is.na(S22.Detrend.df$Rainfall)==T)
+S22.Detrend.df$Rainfall[z]<-0
+no.rain<-rep(NA,length(S22.Detrend.df$Rainfall))
+for(i in 6:length(S22.Detrend.df$Rainfall)){
+  no.rain[i]<-ifelse(sum(S22.Detrend.df$Rainfall[(i-5):i])==0,i,NA)
+}
+no.rain<-na.omit(no.rain)
+
+#Find start and end of all rainfall events
+start = no.rain[which(diff(no.rain)>1)]
+end = no.rain[which(diff(no.rain)>1)+1]
+
+start = start + 1
+end = end - 6
+d = end - start + 1
+
+S22.Detrend.df$Rainfall[z]<-NA
+
+#Find volume
+v<-rep(NA,length(start))
+for(i in 1:length(start)){
+  v[i] = sum(S22.Detrend.df$Rainfall[(start[i]):(end[i])])
+}
+
+event.mid = start+ round(d/2,0)
+
+#Putting volume into a dataframe
+S22.Detrend.df.declust = data.frame(S22.Detrend.df[,1],rep(NA,nrow(S22.Detrend.df)),S22.Detrend.df[,c(3,5)])
+colnames(S22.Detrend.df.declust)<-c("Date","Rainfall_vol","OsWL","Name")
+S22.Detrend.df.declust$Rainfall_vol = rep(0, nrow(S22.Detrend.df.declust))
+S22.Detrend.df.declust$Rainfall_vol[event.mid] = v
+
+#Obtaining the final conditional sample
+rainfall.thres = 5
+events = which(v>rainfall.thres)
+wl = rep(NA,length(events))
+tc = rep(NA,length(events))
+for(j in 1:length(events)){
+  if(d[events[j]]<72){
+    wl[j] = max(S22.Detrend.df$OsWL[(max(1,event.mid[events[j]]-5)):(min(event.mid[events[j]]+5,nrow(S22.Detrend.df)))])
+  }else{
+    wl[j] = max(S22.Detrend.df$OsWL[start[events[j]]:end[events[j]]])
+  }
+  tc[j] = ifelse(any(!is.na(S22.Detrend.df$Name[(max(1,event.mid[events[j]]-5)):(min(event.mid[events[j]]+5,nrow(S22.Detrend.df)))])),"tc","nontc")
+}
+
+Sample_df = na.omit(data.frame(v[events],wl, tc, event.mid[events]))
+colnames(Sample_df) = c("Rainfall_vol","OsWL","type","x.con")
+
+con.sample.rainfall.tc = list(Threshold = rainfall.thres, Data = Sample_df[which(Sample_df$type=="tc"),1:2], Con_Variable = "rainfall_vol",  x.con = Sample_df$x.con[which(Sample_df$type=="tc")])
+con.sample.rainfall.nontc = list(Threshold = rainfall.thres, Data = Sample_df[which(Sample_df$type=="nontc"),1:2], Con_Variable = "rainfall_vol",  x.con = Sample_df$x.con[which(Sample_df$type=="nontc")])
+
+cop.rainfall.tc <- BiCopSelect(pobs(con.sample.rainfall.tc$Data$Rainfall_vol), pobs(con.sample.rainfall.tc$Data$OsWL), familyset = NA, selectioncrit = "AIC",
[…]
