### 2025-07-21 a53a75c
Delete tests/testthat/return_curve_diag.R
