### 2025-07-21 c34b86a
Update paper.md
