+++ joss_submission/paper.md
@@ -81,7 +81,7 @@ The package possesses the functionaity to compute the desired isoline by interpo
 
 ![The 100-year isoline at case study site derived using two methods. Two-sided conditional sample - copula theory method: (a) Partial isolines from the samples conditioned on rainfall (red line) and water level (blue line) whose outer envelope when overlaid gives (b) the full isoline. Heffernan and Tawn model method: (c) Sample uncertainties are calculated along 100 rays emanating from the point comprising the minimum observed values of both drivers. (d) The median isoline (red line) and associated 90% confidence interval (dashed black lines) obtained via a block bootstrap procedure . \label{fig:isolines}](Figure_1.png)
 
-Compound events are of increasing concern for entities responsible for managing flood risk. `MultiHazard` is designed as a comprehensive user-friendly tool for copula-based joint probability analysis in R. As such, the package provides functions for pre-processing data including imputing missing values, detrending and declustering time series, as well as exploratory data analysis, e.g., analyzing pairwise correlations over a range of time-lags between the two drivers/hazards. The package also contains two automated threshold selection method for the Generalized Pareto distribution [@Solari:2017; Murphy:2024] and approaches for robustly capturing the dependence structure when there are more than two relevant drivers/hazards,namely, standard (elliptic/Archimedean) copulas, Pair Copula Constructions (PCCs) and the HT04 model. For the analysis undertaken for the SFWMD, the higher dimensional approaches enabled groundwater level to be included in the analysis. More recently, an approach to generate time varying synthetic events, i.e., hyetographs and hydrographs, was added. Time varying conditions are a prerequisite for non-steady state hydrodynamic modeling. 
+Compound events are of increasing concern for entities responsible for managing flood risk. `MultiHazard` is designed as a comprehensive user-friendly tool for copula-based joint probability analysis in R. As such, the package provides functions for pre-processing data including imputing missing values, detrending and declustering time series, as well as exploratory data analysis, e.g., analyzing pairwise correlations over a range of time-lags between the two drivers/hazards. The package also contains two automated threshold selection method for the Generalized Pareto distribution [@Solari:2017; @Murphy:2024] and approaches for robustly capturing the dependence structure when there are more than two relevant drivers/hazards,namely, standard (elliptic/Archimedean) copulas, Pair Copula Constructions (PCCs) and the HT04 model. For the analysis undertaken for the SFWMD, the higher dimensional approaches enabled groundwater level to be included in the analysis. More recently, an approach to generate time varying synthetic events, i.e., hyetographs and hydrographs, was added. Time varying conditions are a prerequisite for non-steady state hydrodynamic modeling. 
 
 # Related packages
 
