+++ R/GPD_Threshold_Solari.R
@@ -56,7 +56,7 @@
 GPD_Threshold_Solari<-function(Event,Data,RPs=c(10,50,100,500,1000),RPs_PLOT=c(2,3,4),Min_Quantile=0.95,Alpha=0.1,mu=365.25,N_Sim=10){
 
     # Check for missing required parameters
-    if (is.null(Event)) {
+    if (is.null(Data)) {
       stop("Data is missing.")
     }
 
