+++ tests/testthat/test-GPD_Threshold_Solari_Sel.R
@@ -21,7 +21,7 @@ test_that("GPD_Threshold_Solari_Sel works", {
   expect_equal(length(result$Estimate),3)
   expect_equal(length(result$CI_Lower),32)
   expect_equal(length(result$CI_Upper),32)
-  expect_true(nrow(result$BOOT)=10000 & ncol(result$BOOT)=32)
+  expect_true(nrow(result$BOOT)==10000 & ncol(result$BOOT)==32)
 })
 
 test_that("Invalid inputs", {
