+++ tests/testthat/test-GPD_Threshold_Solari_Sel.R
@@ -4,7 +4,7 @@ Rainfall_Declust_SW<-Decluster_SW(Data=S22.Detrend.df[,c(1:2)],Window_Width=7)
 S22_OsWL_Solari<-GPD_Threshold_Solari(Event=Rainfall_Declust_SW$Declustered,
                                      Data=na.omit(S22.Detrend.df[,2]))
 
-                                              #'
+
 test_that("GPD_Threshold_Solari_Sel works", {
 
   result <- GPD_Threshold_Solari_Sel(Event=Rainfall_Declust_SW$Declustered,
@@ -18,7 +18,7 @@ test_that("GPD_Threshold_Solari_Sel works", {
 
 
   #Checking length of outputs
-  expect_equal(length(result$Estimate),3)
+  expect_equal(length(result$Estimate),32)
   expect_equal(length(result$CI_Lower),32)
   expect_equal(length(result$CI_Upper),32)
   expect_true(nrow(result$BOOT)==10000 & ncol(result$BOOT)==32)
@@ -29,6 +29,7 @@ test_that("Invalid inputs", {
   expect_error(
     GPD_Threshold_Solari_Sel(Event=Rainfall_Declust_SW$Declustered,
                              Data=S22.Detrend.df[,2],
+                             Solari_Output = NULL,
                              Thres=S22_OsWL_Solari$Candidate_Thres),
     "Solari_Output is missing.")
 
