+++ R/Design_Event_2D_Multi_Pop.R
@@ -68,31 +68,26 @@
 #' @return Plot of all the observations (grey circles) as well as the declustered excesses above Thres1 (blue circles) or Thres2 (blue circles), observations may belong to both conditional samples. Also shown is the isoline associated with \code{RP} contoured according to their relative probability of occurrence on the basis of the sample from the two joint distributions, the "most likely" design event (black diamond), and design event under the assumption of full dependence (black triangle) are also shown in the plot. The function also returns a list comprising the design events assuming full dependence \code{"FullDependence"}, as well as once the dependence between the variables is accounted for the "Most likley" \code{"MostLikelyEvent"} as well as an \code{"Ensemble"} of possible design events and relative probabilities of events on the isoline \code{Contour}. The quantile isolines with \code{Quantile_Isoline_1} and \code{Quantile_Isoline_2}, and GPD thresholds with \code{Threshold_1} and \code{Threshold_2}.
 #' @seealso \code{\link{Copula_Threshold_2D}} \code{\link{Diag_Non_Con}} \code{\link{Diag_Non_Con_Trunc}}
 #' @export
-Design_Event_2D_Multi_Pop<-function(Data, Data_Con1, Data_Con2, Data_Con3, Data_Con4, u1, u2, u3, u4, Thres1=NA, Thres2=NA, Thres3=NA, Thres4=NA, N_Both_1, N_Both_2, Copula_Family1, Copula_Family2, Copula_Family3, Copula_Family4, Marginal_Dist1, Marginal_Dist2, Marginal_Dist3, Marginal_Dist4, Marginal_Dist1_Par=NA, Marginal_Dist2_Par=NA, Marginal_Dist3_Par=NA, Marginal_Dist4_Par=NA, Con1="Rainfall",Con2="OsWL", Con3="Rainfall", Con4="OsWL", GPD1=NA, GPD2=NA, GPD3=NA, GPD4=NA, Rate_Con1=NA, Rate_Con2=NA, Rate_Con3=NA, Rate_Con4=NA, Tab1= NA, Tab2 = NA, Tab3 = NA, Tab4 = NA, mu=365.25, GPD_Bayes=FALSE, Decimal_Place=2, Grid_x_min = NA ,Grid_x_max = NA, Grid_y_min = NA, Grid_y_max = NA, Grid_x_interval=NA, Grid_y_interval=NA, RP, Interval=10000, End=F, Resolution="Low", x_lab="Rainfall (mm)",y_lab="O-sWL (mNGVD 29)",x_lim_min = NA,x_lim_max = NA,y_lim_min = NA,y_lim_max = NA,Isoline_Probs="Sample", N=10^6,N_Ensemble=0,Sim_Max=10,Plot_Quantile_Isoline=FALSE){
-
-  #Validation of inputs
-  if (!is.data.frame(Data) || ncol(Data) != 2) {
-    stop("Data must be a data frame with exactly 2 columns.")
-  }
[…]
