+++ tests/testthat/test-Design_Event_2D_Multi_Pop.R
@@ -212,20 +212,16 @@ test_that("Design_Event_2D_Multi_Pop works", {
   expect_true(nrow(result$FullDependence$`100`) == 1 & ncol(result$FullDependence$`100`)==2)
   expect_true(nrow(result$MostLikelyEvent$`100`) == 1 & ncol(result$MostLikelyEvent$`100`)==2)
   expect_equal(nrow(result$Ensemble$`100`),10)
-  expect_equal(nrow(result$Isoline$`100`),10000)
-  expect_equal(length(result$Contour$`100`),10000)
-  expect_equal(nrow(result$Quantile_Isoline_1$`100`),1429)
-  expect_equal(nrow(result$Quantile_Isoline_2$`100`),1212)
+  expect_equal(nrow(result$Isoline$`100`),1915)
+  expect_equal(length(result$Contour$`100`),1915)
   expect_equal(length(result$Threshold_1),1)
   expect_equal(length(result$Threshold_2),1)
 
   #Checking column names of outputs
-  expect_equal(colnames(S20.Detrend.df[,2:3]), colnames(result$FullDependence$`100`))
-  expect_equal(colnames(S20.Detrend.df[,2:3]), colnames(result$MostLikelyEvent$`100`))
-  expect_equal(colnames(S20.Detrend.df[,2:3]), colnames(result$Ensemble$`100`))
-  expect_equal(colnames(S20.Detrend.df[,2:3]), colnames(result$Isoline$`100`))
-  expect_equal(colnames(S20.Detrend.df[,2:3]), colnames(result$Quantile_Isoline_1$`100`))
-  expect_equal(colnames(S20.Detrend.df[,2:3]), colnames(result$Quantile_Isoline_2$`100`))
+  expect_equal(colnames(S22.Detrend.df[,2:3]), colnames(result$FullDependence$`100`))
+  expect_equal(colnames(S22.Detrend.df[,2:3]), colnames(result$MostLikelyEvent$`100`))
+  expect_equal(colnames(S22.Detrend.df[,2:3]), colnames(result$Ensemble$`100`))
+  expect_equal(colnames(S22.Detrend.df[,2:3]), colnames(result$Isoline$`100`))
 })
 
 test_that("Invalid inputs", {
