+++ README.md
@@ -1,1761 +1,0 @@
-
-# MultiHazard <img src="https://user-images.githubusercontent.com/15319503/195926656-9d3d37b5-86ab-4d4b-9e6d-3c70d5399c73.png" align="right" height="200"/>
-
-The `MultiHazard` package provides tools for stationary multivariate
-statistical modeling, for example, to estimate joint occurrence
-probabilities of MULTIple co-occurring HAZARDs. The package contains
-functions for pre-processing data, including imputing missing values,
-detrending and declustering time series as well as analyzing pairwise
-correlations over a range of lags. Functionality is built in to
-implement the conditional sampling - copula theory approach described in
-[Jane et al. (2020)](https://doi.org/10.5194/nhess-20-2681-2020)
-including the automated threshold selection method from [Solari et
-al. (2017)](https://doi.org/10.1002/2016WR019426). There is a function
-that calculates joint probability contours using the method of
-overlaying conditional contours given in [Bender et
-al. (2016)](https://doi.org/10.1080/02626667.2015.1052816) and extracts
-design events such as the “most likely” event or an ensemble of possible
-design events. The package also includes methods from [Murphy-Barltrop
-et al. (2023)](https://doi.org/10.1002/env.2797) and [Murphy-Barltrop et
-al. (2024)](https://doi.org/10.1007/s10687-024-00490-4) for deriving
-isolines using the [Heffernan and Tawn
-(2004)](https://doi.org/10.1111/j.1467-9868.2004.02050.x) [HT04](#ht04)
-and [Wadsworth and Tawn (2013)](https://doi.org/10.3150/12-BEJ471)
-\[WT13\] models, together with a novel bootstrap procedure for
-quantifying sampling uncertainty in the isolines. Three higher
-dimensional approaches — standard (elliptic/Archimedean) copulas, Pair
-Copula Constructions (PCCs) and a conditional threshold exceedance
-approach (HT04) — are coded. Finally, the package can be implemented to
-derive temporally coherent extreme events comprising a hyetograph and
-water level curve for simulated peak rainfall and peak sea level events,
-as outlined in (Report).
-
-### Citation:
-
-> Jane, R., Wahl, T., Peña, F., Obeysekera, J., Murphy-Barltrop, C.,
-> Ali, J., Maduwantha, P., Li, H., and Malagón Santos, V. (under review)
-> MultiHazard: Copula-based Joint Probability Analysis in R. Journal of
-> Open Source Software. \[under revision\]
-
-<br>
-
-### Applications of package:
-
-> Li, H., Jane, R. A., Eilander, D., Enríquez, A. R., Haer, T., and
-> Ward, P. J. (2025). Assessing the spatial correlation of potential
-> compound flooding in the United States, EGUsphere \[preprint\],
-> <https://doi.org/10.5194/egusphere-2025-2993>.
-
-> Amorim, R., Villarini, G., Kim, H., Jane, R., and Wahl, T. (2025). A
-> Practitioner’s approach to process-driven modeling of compound
-> rainfall and storm surge extremes for coastal Texas, J. Hydrol. Eng.,
-> 30(5), 04025025, <https://doi.org/10.1061/JHYEFF.HEENG-648>.
-
-> Maduwantha, P., Wahl, T., Santamaria-Aguilar, S., Jane, R., Booth, J.
-> F., Kim, H., and Villarini, G. (2024). A multivariate statistical
-> framework for mixed storm types in compound flood analysis, Nat.
[…]
