### 2025-07-30 #8350 @rjaneUCF
Hi again (@lrnv @mingzehuang, @EFAcar), I have updated the README to include a Community Guidelines section and fixed the issue with the Diag_Non_Con() function. The package also now contains automated test for each of the key functions. Thank you @mingzehuang for the suggestions.
