### 2025-08-04 #8350 @mingzehuang
Hi @rjaneUCF Thank you so much for fixing them! I've rerun the README.md and checked the Community Guidelines as well as your test folders. Everything is great now:)
I have only one small suggestion, in README.md line 1522: "SeaLevelRise.2022<-read.csv("F:/OneDrive - University of Central Florida/Documents/sl_taskforce_scenarios_psmsl_id_1106_Fort_Myers.csv")" You can either remove it if the original file is not suitable for public access, or put it into your "data-raw" folder if it's suitable for public access.
Thanks!
