+++ man/Design_Event_2D_Multi_Pop.Rd
@@ -44,10 +44,10 @@ Design_Event_2D_Multi_Pop(
   Rate_Con2 = NA,
   Rate_Con3 = NA,
   Rate_Con4 = NA,
-  Tab1 = NA,
-  Tab2 = NA,
-  Tab3 = NA,
-  Tab4 = NA,
+  Tab1 = NULL,
+  Tab2 = NULL,
+  Tab3 = NULL,
+  Tab4 = NULL,
   mu = 365.25,
   GPD_Bayes = FALSE,
   Decimal_Place = 2,
