+++ DESCRIPTION
@@ -49,6 +49,8 @@ Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
 Imports: 
+    actuar,
+    akima,
     copula,
     dplyr,
     fitdistrplus,
@@ -62,6 +64,7 @@ Imports:
     RColorBrewer,
     ReturnCurves,
     rmatio,
+    scales,
     statmod,
     texmex,
     truncnorm,
