+++ data/return_curve_est.Rd
@@ -1,136 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/return_curve_est.R
-\name{return_curve_est}
-\alias{return_curve_est}
-\title{Derives return curves that capture uncertainty}
-\usage{
-return_curve_est(
-  data,
-  q,
-  rp,
-  mu,
-  n_sim,
-  n_grad,
-  n_boot,
-  boot_method,
-  boot_replace,
-  block_length,
-  boot_prop,
-  decl_method_x,
-  decl_method_y,
-  window_length_x,
-  window_length_y,
-  u_x = NA,
-  u_y = NA,
-  sep_crit_x = NA,
-  sep_crit_y = NA,
-  alpha = 0.1,
-  most_likely = F,
-  n_interp = 1000,
-  n = 10^6,
-  n_ensemble = 0,
-  x_lab = colnames(data)[2],
-  y_lab = colnames(data)[2],
-  x_lim_min = min(data[, 2], na.rm = T),
-  x_lim_max = max(data[, 2], na.rm = T) + 0.3 * diff(range(data[, 2], na.rm = T)),
-  y_lim_min = min(data[, 3], na.rm = T),
-  y_lim_max = max(data[, 3], na.rm = T) + 0.3 * diff(range(data[, 3], na.rm = T)),
-  plot = T
-)
-}
-\arguments{
-\item{data}{Data frame of raw data detrended if necessary. First column should be of class \code{Date}.}
-
-\item{q}{Numeric vector of length one specifying quantile level for fitting GPDs and the HT04 and WT13 models.}
-
-\item{rp}{Numeric vector of length one specifying return period of interest.}
-
-\item{mu}{Numeric vector of length one specifying the (average) occurrence frequency of events in Data. Default is 365.25, daily data.}
-
-\item{n_sim}{Numeric vector of length one specifying the number of simulations for HT model. Default is \code{50}.}
-
-\item{n_grad}{Numeric vector of length one specifying number of number of rays along which to compute points on the curve. Default is \code{50}.}
-
-\item{n_boot}{Numeric vector of length one specifying number of bootstrap samples. Default is \code{100}.}
-
-\item{boot_method}{Character vector of length one specifying the bootstrap method. Options are \code{"basic"} (default), \code{"block"} or \code{"monthly"}.}
-
-\item{boot_replace}{Character vector of length one specifying whether bootstrapping is carried out with \code{"T"} or without \code{"F"} replacement. Only required if \code{boot_method = "basic"}. Default is \code{NA}.}
-
-\item{block_length}{Numeric vector of length one specifying block length. Only required if \code{boot_method = "block"}. Default is \code{NA}.}
-
-\item{boot_prop}{Numeric vector of length one specifying the minimum proportion of non-missing values of at least of the variables for a month to be included in the bootstrap. Only required if \code{boot_method = "monthly"}. Default is \code{0.8}.}
-
-\item{decl_method_x}{Character vector of length one specifying the declustering method to apply to the first variable. Options are the storm window approach \code{"window"} (default) and the runs method \code{"runs"}.}
-
-\item{decl_method_y}{Character vector of length one specifying the declustering method to apply to the second variable. Options are the storm window approach \code{"window"} (default) and the runs method \code{"runs"}.}
-
-\item{window_length_x}{Numeric vector of length one specifying the storm window length to apply during the declustering of the  first variable if \code{decl_method_x = "window"}.}
-
-\item{window_length_y}{Numeric vector of length one specifying the storm window length to apply during the declustering of the second variable if \code{decl_method_y = "window"}.}
-
[…]
