### 2025-09-02 42ae574
Update paper.md
