+++ joss_submission/paper.md
@@ -75,7 +75,7 @@ The `MultiHazard` R package is designed to enable practitioners to estimate the 
 
 # Statement of need
 
-The `MultiHazard` package was developed in collaberation with the South Florida Water Management District (SFWMD) to improve the level of service assessments for coastal infrastructure affected by both inland and coastal drivers [@Jane:2020]. Initially, the package was created to implement the conditional sampling - copula theory approach. In two-sided conditional sampling, a driver is conditioned to be extreme and paired with the maximum value of the other driver within a specified lag-time. The process is repeated with the drivers reversed yielding two conditional samples. The best fitting of 40 copulas are tested to model the dependence between drivers in each conditional sample. The isoline corresponding to a user specified return period is given by the outer envelope when overlaying the (conditional) contours from the copula model fit to each sample @Bender:2016, see \autoref{fig:isolines}a,b. To obtain a single design event, the probability of events on the isolines is calculated using an empirical density estimate, selecting the event with the highest likelihood. Some experts recommend sampling an ensemble of events from the isoline to account for uncertainty in design event selection which is also possible in the package. The conditional sampling – copula theory approach has continued to gain traction [@Kim:2023; @Maduwantha:2024] and in a review of the best-available, actionable science was highlighted as an approach that Federal agencies in the United States may wish to develop detailed technical guidance on how to use it meet their needs [@FFRMS:2023]. 
+The `MultiHazard` package was developed in collaboration with the South Florida Water Management District (SFWMD) to improve the level of service assessments for coastal infrastructure affected by both inland and coastal drivers [@Jane:2020]. Initially, the package was created to implement the conditional sampling - copula theory approach. In two-sided conditional sampling, a driver is conditioned to be extreme and paired with the maximum value of the other driver within a specified lag-time. The process is repeated with the drivers reversed yielding two conditional samples. The best fitting of 40 copulas are tested to model the dependence between drivers in each conditional sample. The isoline corresponding to a user specified return period is given by the outer envelope when overlaying the (conditional) contours from the copula model fit to each sample [@Bender:2016], see \autoref{fig:isolines}a,b. To obtain a single design event, the probability of events on the isolines is calculated using an empirical density estimate, selecting the event with the highest likelihood. Some experts recommend sampling an ensemble of events from the isoline to account for uncertainty in design event selection which is also possible in the package. The conditional sampling – copula theory approach has continued to gain traction [@Kim:2023; @Maduwantha:2024] and in a review of the best-available, actionable science was highlighted as an approach that Federal agencies in the United States may wish to develop detailed technical guidance on how to use it meet their needs [@FFRMS:2023]. 
 
 The package possesses the functionaity to compute the desired isoline by interpolating joint return periods calculated over a user-defined grid which can result in smoother curves than overlaying the partial isolines. Isoline generation for time series comprising events from multiple (independent) populations e.g. different types of storms is faciliated by multiplying independent annual-non exceedance probabilities associated with each population as outlined in @Maduwantha:2024. Also coded in the package are the methods proposed in @Barltrop:2023 and @Barltrop:2024 to derive isolines using the HT04 [@Heffernan:2004] and WT13 [@Wadsworth:2013] models, along with a novel bootstrap procedure for evaluating isoline uncertainty (see \autoref{fig:isolines}c,d), and a diagnotic tool proposed for evaluating the goodness of fit for a curve estimate. The HT04 and WT13 models allow for the analysis of compound events without assuming any copula form. Furthermore, the results from @Barltrop:2023 guarantee that the estimated isolines from the HT04 and WT13 approaches satisfy certain theoretical properties. 
 
[…]
