+++ vignettes/MultiHazard.Rmd
@@ -8,7 +8,8 @@ vignette: >
 ---
 
 ```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
+knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, collapse = TRUE,
+  comment = "#>")
 library(MultiHazard)
 library(dplyr)
 library(scales)
