+++ R/Annual_Max.R
@@ -10,6 +10,9 @@
 #' @return List comprising the index of the annual maximum \code{Event} and the annual maximum values \code{AM}.
 #' @export
 #' @examples
+#' #Loading data contained in the package
+#' data("S20_T_MAX_Daily_Completed_Detrend")
+#' #Finding annual maximum
 #' Annual_Max(Data_Detrend=S20_T_MAX_Daily_Completed_Detrend$Detrend)
 Annual_Max<-function(Data_Detrend, Complete_Prop=0.8){
 
