+++ man/Annual_Max.Rd
@@ -21,5 +21,8 @@ List comprising the index of the annual maximum \code{Event} and the annual maxi
 Extract annual maximum in years with over a user-defined proportion of non-missing values.
 }
 \examples{
+#Loading data contained in the package
+data("S20_T_MAX_Daily_Completed_Detrend")
+#Finding annual maximum
 Annual_Max(Data_Detrend=S20_T_MAX_Daily_Completed_Detrend$Detrend)
 }
