+++ man/Annual_Max.Rd
@@ -22,7 +22,7 @@ Extract annual maximum in years with over a user-defined proportion of non-missi
 }
 \examples{
 #Loading data contained in the package
-data("S20_T_MAX_Daily_Completed_Detrend")
+data("S20_T_MAX_Daily_Completed_Detrend_Declustered")
 #Finding annual maximum
-Annual_Max(Data_Detrend=S20_T_MAX_Daily_Completed_Detrend$Detrend)
+Annual_Max(Data_Detrend=S20_T_MAX_Daily_Completed_Detrend_Declustered$Detrend)
 }
