+++ R/Annual_Max.R
@@ -12,8 +12,8 @@
 #' @examples
 #' #Loading data contained in the package
 #' data("S20_T_MAX_Daily_Completed_Detrend_Declustered")
-#' #Finding annual maximum
-#' Annual_Max(Data_Detrend=S20_T_MAX_Daily_Completed_Detrend_Declustered$Detrend)
+#' #Finding annual maximum of detrended time series
+#' Annual_Max(Data_Detrend=S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(1,4)])
 Annual_Max<-function(Data_Detrend, Complete_Prop=0.8){
 
   # Input validation
