+++ man/Annual_Max.Rd
@@ -21,8 +21,8 @@ List comprising the index of the annual maximum \code{Event} and the annual maxi
 Extract annual maximum in years with over a user-defined proportion of non-missing values.
 }
 \examples{
-#Loading data contained in the package
-data("S20_T_MAX_Daily_Completed_Detrend_Declustered")
+#Converting
+S20_T_MAX_Daily_Completed_Detrend_Declustered[,1] <- as.Date(S20_T_MAX_Daily_Completed_Detrend_Declustered[,1], format = "\%m/\%d/\%Y")
 #Finding annual maximum of detrended time series
 Annual_Max(Data_Detrend=S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(1,4)])
 }
