+++ man/Conditional_RP_2D_Equal.Rd
@@ -129,6 +129,23 @@ A large number of realizations are simulated from the copulas fit to the conditi
 #Under a 10yr rainfall event condition, what is the joint probability that a 10yr surge (O-sWL)
 #event occurs simultaneously?  What is the cumulative probability of events with the frequency
 #equal to or less than a 10yr surge event?
+#' #Conditional samples
+con.sample.Rainfall<-Con_Sampling_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
+                                     Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
+                                     Con_Variable="Rainfall",u=0.98)
+con.sample.OsWL<-Con_Sampling_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
+                                 Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
+                                 Con_Variable="OsWL",u=0.98)
+#Add some noise to rainfall to aid distribution fitting in Conditional_RP_2D function
+con.sample.OsWL$Data$Rainfall<-con.sample.OsWL$Data$Rainfall+runif(length(con.sample.OsWL$Data$Rainfall),0.001,0.01)
+#Find the best fitting copula
+cop.Rainfall <- Copula_Threshold_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
+                                    Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
+                                     u1=0.98, u2=NA, PLOT=FALSE)$Copula_Family_Var1
+cop.OsWL<- Copula_Threshold_2D(Data_Detrend=S22.Detrend.df[,-c(1,4)],
+                               Data_Declust=S22.Detrend.Declustered.df[,-c(1,4)],
+                               u1=NA, u2=0.98, PLOT=FALSE)$Copula_Family_Var2
+#Calculate conditional probabilities
 Conditional_RP_2D_Equal(Data=S22.Detrend.df,
                         Data_Con1=con.sample.Rainfall$Data, Data_Con2=con.sample.OsWL$Data,
                         u1=0.98, u2=0.98,
@@ -140,7 +157,7 @@ Conditional_RP_2D_Equal(Data=S22.Detrend.df,
                         RP_Con=10, RP_Non_Con=10,
                         x_lab = "Rainfall (Inches)", y_lab = "O-sWL (ft NGVD 29)",
                         y_lim_max = 10,
-                        N=10^8)
+                        N=10^5)
 }
 \seealso{
 \code{\link{Design_Event_2D}} \code{\link{Conditional_RP_2D}}
