+++ R/Annual_Max.R
@@ -11,7 +11,8 @@
 #' @export
 #' @examples
 #' #Converting
-#' S20_T_MAX_Daily_Completed_Detrend_Declustered[,1] <- as.Date(S20_T_MAX_Daily_Completed_Detrend_Declustered[,1], format = "%m/%d/%Y")
+#' S20_T_MAX_Daily_Completed_Detrend_Declustered[,1] <-
+#'                   as.Date(S20_T_MAX_Daily_Completed_Detrend_Declustered[,1], format = "%m/%d/%Y")
 #' #Finding annual maximum of detrended time series
 #' Annual_Max(Data_Detrend=S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(1,4)])
 Annual_Max<-function(Data_Detrend, Complete_Prop=0.8){
