+++ man/Annual_Max.Rd
@@ -22,7 +22,8 @@ Extract annual maximum in years with over a user-defined proportion of non-missi
 }
 \examples{
 #Converting
-S20_T_MAX_Daily_Completed_Detrend_Declustered[,1] <- as.Date(S20_T_MAX_Daily_Completed_Detrend_Declustered[,1], format = "\%m/\%d/\%Y")
+S20_T_MAX_Daily_Completed_Detrend_Declustered[,1] <-
+                  as.Date(S20_T_MAX_Daily_Completed_Detrend_Declustered[,1], format = "\%m/\%d/\%Y")
 #Finding annual maximum of detrended time series
 Annual_Max(Data_Detrend=S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(1,4)])
 }
