+++ R/Time_Series_Plot.R
@@ -15,7 +15,6 @@
 #' @return Figure containing a (Row * Col) matrix of subplots each displaying the hyetogaph (grey bars) and water level curve (blue lines) comprising an event.
 #' @seealso \code{\link{U_Sample}} \code{\link{WL_Curve}}
 #' @export
-#' @examples
 Time_Series_Plot<-function(Rainfall_Series,Oswl_Time_Series,Sample,Con_Variable,Buffer=6,Intensity=NA,Event_ID=1:16,Row=4,Col=4,Mar=c(4.2,4.5,1.5,3.5)){
  #Figure layout
  par(mfrow=c(Row,Col))
