+++ R/Design_Event_2D_Grid.R
@@ -232,7 +232,7 @@ Design_Event_2D_Grid<-function(Data, Data_Con1, Data_Con2, u1, u2, Thres1=NA, Th
   names(FullDependence)<-RP
 
   #Remove 1st column of Data if it is a Date or factor object.
-  if(class(Data[,1])=="Date" | class(Data[,1])=="factor" | class(Data[,1])[1]=="POSIXct"){
+  if(inherits(Data[,1], c("Date", "factor","POSIXct"))){
     Data<-Data[,-1]
   }
 
