+++ man/HT04.Rd
@@ -48,7 +48,7 @@ List comprising the fitted HT04 models \code{Models}, proportion of the time eac
 Fits the conditional multivariate approach of Heffernan and Tawn (2004) to a dataset and simulates realizations from the fitted model. Function utilizes the \code{mexDependence} and \code{predict.mex.conditioned} functions from the \code{texmex} package.
 }
 \examples{
-Fit GPD marginal distributions above the threshold
+#Fit GPD marginal distributions above the threshold
 S20.Migpd<-Migpd_Fit(Data=S20.Detrend.Declustered.df[,-1],
                      Data_Full=S20.Detrend.Declustered.df[,-1],
                      mqu =c(0.99,0.99,0.99))
