### 2025-09-05 b6d66dc
Update Cooley19.Rd
