+++ R/Cooley19.R
@@ -20,9 +20,9 @@
 #' @seealso \code{\link{Dataframe_Combine}} \code{\link{Decluster}} \code{\link{GPD_Fit}} \code{\link{Migpd_Fit}}
 #' @export
 #' @examples
-#' S20.GPD<-Migpd_Fit(Data=S20.Detrend.Declustered.df[,-1],
-#'                    Data_Full=S20.Detrend.Declustered.df[,-1],
-#'                    mqu =c(0.99,0.99,0.99))
+#' S20.GPD<-Migpd_Fit(Data=S20.Detrend.Declustered.df[,c(3,4)],
+#'                    Data_Full=S20.Detrend.Declustered.df[,c(3,4)],
+#'                    mqu =c(0.99,0.99))
 #' Cooley19(Data=na.omit(S20.Detrend.df[,3:4]),Migpd=S20.GPD,
 #' p.base=0.01,p.proj=0.001,PLOT=TRUE,x_lim_max_T=500,y_lim_max_T=500)
 Cooley19<-function(Data,Migpd,p.base=0.01,p.proj=0.001,u=0.95,PLOT=FALSE,x_lim_min_T=NA, x_lim_max_T=NA,y_lim_min_T=NA,y_lim_max_T=NA,x_lim_min=NA,x_lim_max=NA,y_lim_min=NA,y_lim_max=NA){
