+++ R/datasets.R
@@ -1,0 +1,27 @@
+#' Groundwater Well G-3355
+#'
+#' Time series of daily water elevation at Well G-3355.
+#'
+#' @format A data frame with 12,340 rows and 1 variable:
+#' \describe{
+#'   \item{Date}{Date of observation (YYYY-MM-DD)}
+#'   \item{Value}{Groudwater elevation}
+#' }
+#' @source South Florida Water Management District, via [https://www.sfwmd.gov/science-data/dbhydro/](https://www.sfwmd.gov/science-data/dbhydro/)
+#'
+#' @usage data(G_3355)
+"G_3355"
+
+#' Groundwater Well G-3356
+#'
+#' Time series of daily water elevation at Well G-3356.
+#'
+#' @format A data frame with 12,272 rows and 1 variable:
+#' \describe{
+#'   \item{Date}{Date of observation (YYYY-MM-DD)}
+#'   \item{Value}{Groudwater elevation}
+#' }
+#' @source South Florida Water Management District, via [https://www.sfwmd.gov/science-data/dbhydro/](https://www.sfwmd.gov/science-data/dbhydro/)
+#'
+#' @usage data(G_3356)
+"G_3356"
