### 2025-09-05 45c840b
Delete R/HT04_Lag.R
