### 2025-09-05 fe4872e
Delete man/HT04_Lag.Rd
