+++ man/HT04_Lag.Rd
@@ -1,66 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/HT04_Lag.R
-\name{HT04_Lag}
-\alias{HT04_Lag}
-\title{Implements the version of the conditional multivariate approach of Heffernan and Tawn (2004) proposed in Keef et al. (2013) which incorporates lags between the variables.}
-\usage{
-HT04_Lag(
-  data_Detrend_Dependence_df,
-  data_Detrend_Declustered_df,
-  Lags,
-  u_Dependence,
-  Migpd,
-  mu = 365.25,
-  N = 100,
-  Margins = "gumbel",
-  V = 10,
-  Maxit = 10000
-)
-}
-\arguments{
-\item{data_Detrend_Dependence_df}{A data frame with (n+1) columns, containing in column \itemize{
-\item 1 - Continuous sequence of dates spanning the first to the final time of any of the variables are recorded.
-\item 2:(n+1) - Values, detrended where necessary, of the variables to be modelled.
-}}
-
-\item{data_Detrend_Declustered_df}{A data frame with (n+1) columns, containing in column \itemize{
-\item 1 - Continuous sequence of dates spanning the first to the final time of any of the variables are recorded.
-\item 2:(n+1) - Declustered and if necessary detrended values of the variables to be modelled.
-}}
-
-\item{Lags}{Matrix specifying the lags. The no lag i.e. \code{0} lag cases need to be specified. Row n denotes the lags applied to the variable in the nth column of \code{data_Detrend_Dependence_df}. Column n corresponds to the nth largest lag applied to any variable. Default is \code{matrix(c(NA,0,1,NA,0,1,NA,0,NA),nrow=3,byrow = T)}, which corresponds to a lag of 1 being applied to variables in the first and second columns of \code{data_Detrend_Dependence_df} and no lag being applied to the variable in the third column of \code{data_Detrend_Dependence_df}.}
-
-\item{u_Dependence}{Dependence quantile. Specifies the (sub-sample of) data to which the dependence model is fitted, that for which the conditioning variable exceeds the threshold associated with the prescribed quantile. Default is \code{0.7}, thus the dependence parameters are estimated using the data with the highest \code{30\%} of values of the conditioning variables.}
-
-\item{Migpd}{An \code{Migpd} object, containing the parameterized Pareto models fitted (independently) to each of the variables.}
-
-\item{mu}{Numeric vector of length one specifying the (average) occurrence frequency of events in \code{data_Detrend_Dependence_df}. Default is \code{365.25}, daily data.}
-
-\item{N}{Numeric vector of length one specifying the number of years worth of extremes to simulate. Default is \code{100} years.}
-
-\item{Margins}{Character vector specifying the form of margins to which the data are transformed for carrying out dependence estimation. Default is \code{"gumbel"}, alternative is \code{"laplace"}. Under Gumbel margins, the estimated parameters \code{a} and \code{b} describe only positive dependence, while \code{c} and \code{d} describe negative dependence in this case. For Laplace margins, only parameters \code{a} and \code{b} are estimated as these capture both positive and negative dependence.}
-
-\item{V}{See documentation for \code{mexDependence}.}
-
-\item{Maxit}{See documentation for \code{mexDependence}.}
-}
-\value{
[…]
