+++ NAMESPACE
@@ -53,7 +53,6 @@ export(inverse_pit_gpd)
 import(GeneralizedHyperbolic)
 import(MASS)
 import(RColorBrewer)
-import(ReturnCurves)
 import(VGAM)
 import(VineCopula)
 import(actuar)
