### 2025-09-05 271cfd3
Delete man/Joint_RP_2D.Rd
