+++ man/Joint_RP_2D.Rd
@@ -1,127 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/Joint_RP_2D.R
-\name{Joint_RP_2D}
-\alias{Joint_RP_2D}
-\title{Calculates joint return periods}
-\usage{
-Joint_RP_2D(
-  Data,
-  Data_Con1,
-  Data_Con2,
-  u1,
-  u2,
-  Thres1 = NA,
-  Thres2 = NA,
-  Copula_Family1,
-  Copula_Family2,
-  Marginal_Dist1,
-  Marginal_Dist2,
-  Con1 = "Rainfall",
-  Con2 = "OsWL",
-  mu = 365.25,
-  RP_Var1,
-  RP_Var2,
-  Var1 = NA,
-  Var2 = NA,
-  x_lab = "Rainfall (mm)",
-  y_lab = "O-sWL (mNGVD 29)",
-  x_lim_min = NA,
-  x_lim_max = NA,
-  y_lim_min = NA,
-  y_lim_max = NA,
-  DecP = 2,
-  N = 10^6
-)
-}
-\arguments{
-\item{Data}{Data frame containing two co-occurring time series.}
-
-\item{Data_Con1}{Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the first column.}
-
-\item{Data_Con2}{Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the second column. Can be obtained using the \code{Con_Sampling_2D} function.}
-
-\item{u1}{Numeric vector of length one specifying the (quantile) threshold above which the variable in the first column was sampled in Data_Con1.}
-
-\item{u2}{Numeric vector of length one specifying the (quantile) threshold above which the variable in the second column was sampled in Data_Con2.}
-
-\item{Thres1}{Numeric vector of length one specifying the threshold above which the variable in the first column was sampled in Data_Con1. Only one of \code{u1} and \code{Thres1} should be supplied. Default is \code{NA}.}
-
-\item{Thres2}{Numeric vector of length one specifying the threshold above which the variable in the second column was sampled in Data_Con2. Only one of \code{u2} and \code{Thres2} should be supplied. Default is \code{NA}.}
-
-\item{Copula_Family1}{Numeric vector of length one specifying the copula family used to model the \code{Data_Con1} dataset.}
-
-\item{Copula_Family2}{Numeric vector of length one specifying the copula family used to model the \code{Data_Con2} dataset. Best fitting of 40 copulas can be found using the \code{Copula_Threshold_2D} function.}
-
-\item{Marginal_Dist1}{Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in \code{Data_Con1}.}
-
-\item{Marginal_Dist2}{Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in \code{Data_Con2}.}
-
-\item{Con1}{Character vector of length one specifying the name of variable in the first column of \code{Data}.}
-
-\item{Con2}{Character vector of length one specifying the name of variable in the second column of \code{Data}.}
-
-\item{mu}{Numeric vector of length one specifying the (average) occurrence frequency of events in \code{Data}. Default is \code{365.25}, daily data.}
-
-\item{RP_Var1}{Numeric vector of length one specifying the univariate return period of variable \code{Con1}.}
-
-\item{RP_Var2}{Numeric vector of length one specifying the univariate return period of variable \code{Con2}.}
-
[…]
