### 2025-09-06 bc63f36
Delete man/Migpd_Fit.Rd
