+++ man/G_580A.Rd
@@ -1,11 +1,11 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/datasets.R
 \docType{data}
-\name{G_3356}
-\alias{G_3356}
-\title{Groundwater Levels at Well G-3356}
+\name{G_580A}
+\alias{G_580A}
+\title{Groundwater Levels at Well G-580A}
 \format{
-A data frame with 12,272 rows and 2 variables:
+A data frame with XX rows and 2 variables:
 \describe{
 \item{Date}{Date of observation (YYYY-MM-DD)}
 \item{Value}{Groundwater elevation (ft NGVD29)}
@@ -16,9 +16,9 @@ South Florida Water Management District, via
 \url{https://www.sfwmd.gov/science-data/dbhydro/}
 }
 \usage{
-data(G_3356)
+data(G_580A)
 }
 \description{
-Time series of daily water elevation at Well G-3356.
+Time series of daily water elevation at Well G-580A.
 }
 \keyword{datasets}
