+++ man/Miami_Airport_df.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/datasets.R
+\docType{data}
+\name{Miami_Airport_df}
+\alias{Miami_Airport_df}
+\title{Rainfall Totals at Miami International Airport}
+\format{
+A data frame with XX rows and 2 variables:
+\describe{
+\item{Date}{Date of observation (YYYY-MM-DD)}
+\item{Value}{Rainfall totals (inches)}
+}
+}
+\source{
+NOAA National Centers for Environmental Information, via
+\url{https://www.ncdc.noaa.gov/}
+}
+\usage{
+data(Miami_Airport_df)
+}
+\description{
+Time series of daily rainfall totals from the gauge at Miami International Airport, FL \link{Network:ID	GHCND:USW00012839}.
+}
+\keyword{datasets}
