+++ R/datasets.R
@@ -1,27 +1,396 @@
-#' Groundwater Well G-3355
+#' Groundwater Levels at Well G-3355
 #'
 #' Time series of daily water elevation at Well G-3355.
 #'
-#' @format A data frame with 12,340 rows and 1 variable:
+#' @format A data frame with 12,340 rows and 2 variables:
 #' \describe{
 #'   \item{Date}{Date of observation (YYYY-MM-DD)}
-#'   \item{Value}{Groudwater elevation}
+#'   \item{Value}{Groundwater elevation (ft NGVD29)}
 #' }
-#' @source South Florida Water Management District, via [https://www.sfwmd.gov/science-data/dbhydro/](https://www.sfwmd.gov/science-data/dbhydro/)
+#' @source South Florida Water Management District, via
+#' [https://www.sfwmd.gov/science-data/dbhydro/](https://www.sfwmd.gov/science-data/dbhydro/)
 #'
 #' @usage data(G_3355)
 "G_3355"
 
-#' Groundwater Well G-3356
+#' Groundwater Levels at Well G-3356
 #'
 #' Time series of daily water elevation at Well G-3356.
 #'
-#' @format A data frame with 12,272 rows and 1 variable:
+#' @format A data frame with 12,272 rows and 2 variables:
 #' \describe{
 #'   \item{Date}{Date of observation (YYYY-MM-DD)}
-#'   \item{Value}{Groudwater elevation}
+#'   \item{Value}{Groundwater elevation (ft NGVD29)}
 #' }
-#' @source South Florida Water Management District, via [https://www.sfwmd.gov/science-data/dbhydro/](https://www.sfwmd.gov/science-data/dbhydro/)
+#' @source South Florida Water Management District, via
+#' [https://www.sfwmd.gov/science-data/dbhydro/](https://www.sfwmd.gov/science-data/dbhydro/)
 #'
 #' @usage data(G_3356)
 "G_3356"
+
+#' Groundwater Levels at Well G-580A
+#'
+#' Time series of daily water elevation at Well G-580A.
+#'
+#' @format A data frame with XX rows and 2 variables:
+#' \describe{
+#'   \item{Date}{Date of observation (YYYY-MM-DD)}
+#'   \item{Value}{Groundwater elevation (ft NGVD29)}
+#' }
+#' @source South Florida Water Management District, via
+#' [https://www.sfwmd.gov/science-data/dbhydro/](https://www.sfwmd.gov/science-data/dbhydro/)
+#'
+#' @usage data(G_580A)
+"G_580A"
+
+#' Atlantic hurricane database (HURricane DATa 2nd generation) - HURDAT2)
+#'
+#' "Best-track" data at a six-hourly resolution for all known tropical cyclones and subtropical cyclones available from the b-decks in the Automated Tropical Cyclone Forecast (ATCF – Sampson and Schrader 2000) system database.
+#'
+#' @format A data frame with columns parsed from the fixed-width HURDAT2 format:
+#' \describe{
+#'    \item{Spaces 1-4}{Year}
+#'    \item{Spaces 5-6}{Month}
+#'    \item{Spaces 7-8, before 1st comma}{Day}
+#'    \item{Spaces 11-12}{Hours in UTC (Universal Time Coordinate)}
+#'    \item{Spaces 13-14 (before 2nd comma)}{Minutes}
+#'    \item{Space 17 (before 3rd comma)} {Record identifier: \itemize{
+#'                           \item C – Closest approach to a coast, not followed by a landfall
+#'                           \item G – Genesis
+#'                           \item vI – An intensity peak in terms of both pressure and wind
+#'                           \item L – Landfall (center of system crossing a coastline)
+#'                           \item P – Minimum in central pressure
[…]
