+++ R/Migpd_Fit.R
@@ -1,0 +1,103 @@
+#' Fits Multiple independent generalized Pareto models - Fit
+#'
+#' Fit multiple independent generalized Pareto models to each column of a data frame. Edited version of the \code{migpd} function in \code{texmex}, to allow for \code{NA}s in a time series.
+#'
+#' @param Data A data frame with \code{n} columns, each comprising a declustered and if necessary detrended time series to be modelled.
+#' @param Data_Full A data frame with \code{n} columns, each comprising the original (detrended if necessary) time series to be modelled. Only required if threshold is specified using \code{mqu}.
+#' @param mth Marginal thresholds, above which generalized Pareto models are fitted. Numeric vector of length \code{n}.
+#' @param mqu Marginal quantiles, above which generalized Pareto models are fitted. \strong{Only one of \code{mth} and \code{mqu} should be supplied.} Numeric vector of length \code{n}.
+#' @param penalty How the likelihood should be penalized. Defaults to \code{"gaussian"}.
+#' @param maxit Numeric vector of length specifying the maximum number of iterations used in the optimization. Default is \code{10000}.
+#' @param trace Default is \code{0}.
+#' @param verbose Default is \code{FALSE}.
+#' @param priorParameters Prior parameters. Only use if \code{penalty = "gaussian"}.
+#' @return An object of class \code{"migpd"}. There are \code{coef}, \code{print}, \code{plot}, and \code{summary} functions available.
+#' @seealso \code{\link{Decluster}} \code{\link{Detrend}} \code{\link{Dataframe_Combine}}
+#' @export
+#' @examples
+#' #With date as first column
+#' S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df, Data_Full=S22.Detrend.df, mqu =c(0.99,0.99,0.99))
+#' #Same GPDs fit as above but thresholds given on the original scale
+#' S22.Rainfall.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$Rainfall),0.99)
+#' S22.OsWL.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$OsWL),0.99)
+#' S22.GW.Quantile<-quantile(na.omit(S22.Detrend.Declustered.df$Groundwater),0.99)
+#' S22.GPD<-Migpd_Fit(Data=S22.Detrend.Declustered.df[,-1],
+#'                    Data_Full=S22.Detrend.df[,-1],
+#'                    mth =c(S22.Rainfall.Quantile,S22.OsWL.Quantile,S22.GW.Quantile))
+Migpd_Fit<-function (Data, Data_Full=NA, mth, mqu, penalty = "gaussian", maxit = 10000,
+                     trace = 0, verbose = FALSE, priorParameters = NULL){
+
+  if(class(Data[,1])[1]=="Date" | class(Data[,1])[1]=="factor" | class(Data[,1])[1]=="POSIXct" | class(Data[,1])[1]=="character"){
+    data <- Data[,-1]
+  } else {
+    data <- Data
+  }
+
+  if(missing(mth)){
+    if(class(Data_Full[,1])[1]=="Date" | class(Data_Full[,1])[1]=="factor" | class(Data_Full[,1])[1]=="POSIXct" | class(Data_Full[,1])[1]=="character"){
+      data_full <- Data_Full[,-1]
+    } else {
+      data_full <- Data_Full
+    }
+  }
+
+  theCall <- match.call()
+  if (is.null(colnames(data))) {
+    colnames(data) <- paste(rep("Column", ncol(data)), 1:ncol(data),
+                            sep = "")
+  }
+  d <- dim(data)[2]
[…]
