+++ data-raw/PVAL_AU2_LMOM_1.R
@@ -1,0 +1,2 @@
+PVAL_AU2_LMOM_1<-read.mat("PVAL_AU2_LMOM_1.mat")
+devtools::use_data(PVAL_AU2_LMOM_1)
