+++ data-raw/HURDAT2.R
@@ -1,0 +1,2 @@
+HURDAT2<-read.table("HURDAT2.txt",header=TRUE)
+devtools::use_data(HURDAT2)
