+++ NAMESPACE
@@ -1,8 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
 export(Annual_Max)
-export(bootstrap_block)
-export(bootstrap_month)
 export(Con_Sampling_2D)
 export(Con_Sampling_2D_Lag)
 export(Conditional_RP_2D)
@@ -37,9 +35,6 @@ export(Mean_Excess_Plot)
 export(Migpd_Fit)
 export(NOAA_SLR)
 export(OsWL_Intensity)
-export(return_curve_diag)
-export(return_curve_est)
-export(SLR_Scenarios)
 export(Standard_Copula_Fit)
 export(Standard_Copula_Sel)
 export(Standard_Copula_Sim)
@@ -49,7 +44,11 @@ export(U_Sample)
 export(Vine_Copula_Fit)
 export(Vine_Copula_Sim)
 export(WL_Curve)
+export(bootstrap_block)
+export(bootstrap_month)
 export(inverse_pit_gpd)
+export(return_curve_diag)
+export(return_curve_est)
 import(GeneralizedHyperbolic)
 import(MASS)
 import(RColorBrewer)
