+++ man/S22_T_MAX_Daily_Completed.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/datasets.R
 \docType{data}
-\name{S20_T_MAX_Daily_Completed_Detrend_Declustered}
-\alias{S20_T_MAX_Daily_Completed_Detrend_Declustered}
+\name{S22_T_MAX_Daily_Completed}
+\alias{S22_T_MAX_Daily_Completed}
 \title{Declustered Time Series of the Ocean-side Water Level at control structure S-20}
 \format{
 A data frame with XX rows and 4 variables:
@@ -19,7 +19,7 @@ A data frame with XX rows and 4 variables:
 South Florida Water Management District: \url{https://www.sfwmd.gov/science-data/dbhydro/}
 }
 \usage{
-data(S20_T_MAX_Daily_Completed_Detrend_Declustered)
+data(S22_T_MAX_Daily_Completed)
 }
 \description{
 Declustered Time Series of the Ocean-side Water Level at control structure S-20.
