+++ man/G_860.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/datasets.R
+\docType{data}
+\name{G_860}
+\alias{G_860}
+\title{Groundwater Levels at Well G-860}
+\format{
+A data frame with 12,272 rows and 2 variables:
+\describe{
+\item{Date}{Date of observation (YYYY-MM-DD)}
+\item{Value}{Groundwater elevation (ft NGVD29)}
+}
+}
+\source{
+South Florida Water Management District, via
+\url{https://www.sfwmd.gov/science-data/dbhydro/}
+}
+\usage{
+data(G_860)
+}
+\description{
+Time series of daily water elevation at Well G-860.
+}
+\keyword{datasets}
