### 2025-09-08 10778d5
Update S22_T_MAX_Daily_Completed.R
