+++ man/S22_T_MAX_Daily_Completed.Rd
@@ -9,10 +9,8 @@ A data frame with XX rows and 4 variables:
 \describe{
 \item{X}{Index}
 \item{Date}{Date of observation (YYYY-MM-DD)}
-\item{Value}{Raw time series (ft NGVD29)}
-\item{Value_Filled}{Raw times series with missing values in-filled using record at structure S-21A_T (ft NGVD29)}
+\item{Value_Filled}{Raw times series record at structure S-20 (ft NGVD29)}
 \item{Detrend}{Detrended time series (ft NGVD29)}
-\item{Declustered}{Declustered time series (ft NGVD29)}
 }
 }
 \source{
