+++ man/Detrend.Rd
@@ -39,9 +39,11 @@ Numeric vector of the detrended time series.
 Detrends a time series using either a linear fit covering the entire dataset or moving average trend correction with a user-specified window width.
 }
 \examples{
+\donttest{
 #Detrending ocean-side water level at site S22 using a 3 month moving average window and the last
 #five years of observations to calculate the present day average.
 data("S22_T_MAX_Daily_Completed")
 Detrend(S22_T_MAX_Daily_Completed[,1:2],Method = "window",Window_Width= 89,
         End_Length = 1826, PLOT=FALSE,x_lab="Data",y_lab="Data")
+}
 }
