### 2025-09-08 123fd50
Delete man/S22_T_MAX_Daily_Completed.Rd
