+++ man/S22_T_MAX_Daily_Completed.Rd
@@ -1,26 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/datasets.R
-\docType{data}
-\name{S22_T_MAX_Daily_Completed}
-\alias{S22_T_MAX_Daily_Completed}
-\title{Declustered Time Series of the Ocean-side Water Level at control structure S-20}
-\format{
-A data frame with XX rows and 4 variables:
-\describe{
-\item{X}{Index}
-\item{Date}{Date of observation (YYYY-MM-DD)}
-\item{Value_Filled}{Raw times series record at structure S-20 (ft NGVD29)}
-\item{Detrend}{Detrended time series (ft NGVD29)}
-}
-}
-\source{
-South Florida Water Management District: \url{https://www.sfwmd.gov/science-data/dbhydro/}
-}
-\usage{
-data(S22_T_MAX_Daily_Completed)
-}
-\description{
-Declustered Time Series of the Ocean-side Water Level at control structure S-20.
-Declustering is implemented using a peaks-over-threshold approach (runs method with a 0.98 quantile threshold and a 3-day separation criterion).
-}
-\keyword{datasets}
