+++ data-raw/S22_T_MAX_Daily_Completed.R
@@ -1,0 +1,2 @@
+S22_T_MAX_Daily_Completed<-read.csv("S22_T_MAX_Daily_Completed.csv")
+devtools::use_data(S22_T_MAX_Daily_Completed)
