+++ R/datasets.R
@@ -307,6 +307,21 @@
 #' @usage data(S20_T_MAX_Daily_Completed_Detrend_Declustered)
 "S20_T_MAX_Daily_Completed_Detrend_Declustered"
 
+#' Time Series of the Ocean-side Water Level at control structure S-22
+#'
+#' Time Series of the Ocean-side Water Level at control structure S-22.
+#'
+#' @format A data frame with XX rows and 4 variables:
+#' \describe{
+#'   \item{Date}{Date of observation (YYYY-MM-DD)}
+#'   \item{ValueFilled}{Raw times series with missing values in-filled using record at structure S-21A_T (ft NGVD29)}
+#' }
+#'
+#' @source South Florida Water Management District: \url{https://www.sfwmd.gov/science-data/dbhydro/}
+#'
+#' @usage data(S22_T_MAX_Daily_Completed_Detrend)
+"S22_T_MAX_Daily_Completed_Detrend"
+
 #' Declustered Time Series for Case Study Site S-22 in Jane et al. (2020)
 #'
 #' Time series of rainfall totals from the Miami International Airport gauge, ocean-side water levels at control structure S-22, and groundwater at Well G-580A (with missing values interpolated using groundwater levels at Well G-860).
