+++ man/S22_T_MAX_Daily_Completed_Detrend.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/datasets.R
+\docType{data}
+\name{S22_T_MAX_Daily_Completed_Detrend}
+\alias{S22_T_MAX_Daily_Completed_Detrend}
+\title{Time Series of the Ocean-side Water Level at control structure S-22}
+\format{
+A data frame with XX rows and 4 variables:
+\describe{
+\item{Date}{Date of observation (YYYY-MM-DD)}
+\item{ValueFilled}{Raw times series with missing values in-filled using record at structure S-21A_T (ft NGVD29)}
+}
+}
+\source{
+South Florida Water Management District: \url{https://www.sfwmd.gov/science-data/dbhydro/}
+}
+\usage{
+data(S22_T_MAX_Daily_Completed_Detrend)
+}
+\description{
+Time Series of the Ocean-side Water Level at control structure S-22.
+}
+\keyword{datasets}
