### 2025-09-09 e8a9e02
Update Detrend.R
