### 2025-09-10 6c4f2cf
Delete data-raw/S22_T_MAX_Daily_Completed.R
