+++ R/datasets.R
@@ -32,7 +32,7 @@
 #'
 #' Time series of daily water elevation at Well G-860.
 #'
-#' @format A data frame with 12,272 rows and 2 variables:
+#' @format A data frame with 16,507 rows and 2 variables:
 #' \describe{
 #'   \item{Date}{Date of observation (YYYY-MM-DD)}
 #'   \item{Value}{Groundwater elevation (ft NGVD29)}
@@ -47,7 +47,7 @@
 #'
 #' Time series of daily water elevation at Well G-580A.
 #'
-#' @format A data frame with XX rows and 2 variables:
+#' @format A data frame with 12,301 rows and 2 variables:
 #' \describe{
 #'   \item{Date}{Date of observation (YYYY-MM-DD)}
 #'   \item{Value}{Groundwater elevation (ft NGVD29)}
@@ -76,7 +76,7 @@
 #'
 #' Time series of daily rainfall totals from the gauge at Miami International Airport, FL (Network:ID	GHCND:USW00012839).
 #'
-#' @format A data frame with XX rows and 2 variables:
+#' @format A data frame with 25,959 rows and 2 variables:
 #' \describe{
 #'   \item{X}{Index}
 #'   \item{Date}{Date of observation (YYYY-MM-DD)}
@@ -88,9 +88,9 @@
 #' @usage data(Miami_Airport_df)
 "Miami_Airport_df"
 
-#' 2012 NOAA Sea Level Rise Projections for Virginia Key
+#' 2012 NOAA Sea Level Rise Projections for Miami Beach
 #'
-#' Sea level rise projections for Virginia Key under four emission scenarios from NOAA Technical Report OAR CPO-1 (2012).
+#' Sea level rise projections for Miami Beach under four emission scenarios from NOAA Technical Report OAR CPO-1 (2012).
 #'
 #' @format A data frame containing the projections with columns as follows:
 #' \describe{
@@ -106,9 +106,9 @@
 #' @usage data(NOAAetal2012)
 "NOAAetal2012"
 
-#' 2017 NOAA Sea Level Rise Projections for Virginia Key
+#' 2017 NOAA Sea Level Rise Projections for Miami Beach
 #'
-#' Sea level rise projections for Virginia Key derived from NOAA Technical Report NOS CO-OPS 083 (2017). The report defines six global mean sea level rise scenarios, which are regionally downscaled on a 1-degree grid covering U.S. coastlines.
+#' Sea level rise projections for Miami Beach derived from NOAA Technical Report NOS CO-OPS 083 (2017). The report defines six global mean sea level rise scenarios, which are regionally downscaled on a 1-degree grid covering U.S. coastlines.
 #'
 #' @format A data frame containing the projections with columns as follows:
 #' \describe{
@@ -128,7 +128,7 @@
 
 #' 2022 NOAA Global and Regional Sea Level Rise Scenarios for the United States
 #'
-#' Probabilistic sea level rise scenarios for Virginia Key from given in NOAA Technical Report NOS 01. In the report, five global mean sea level rise scenarios are used to derive probabilistic regional RSL responses on a 1-degree grid covering the coastlines of the U.S. mainland.
+#' Probabilistic sea level rise scenarios for Miami Beach from given in NOAA Technical Report NOS 01. In the report, five global mean sea level rise scenarios are used to derive probabilistic regional RSL responses on a 1-degree grid covering the coastlines of the U.S. mainland.
 #'
 #' @format A data frame containing the projections with columns as follows:
 #' \describe{
@@ -201,7 +201,7 @@
 #'
 #' Time series of daily rainfall totals from the gauge at Perrine 4W, FL (Network:ID GHCND:USC00087020).
 #'
-#' @format A data frame with XX rows and 2 variables:
[…]
