+++ README.md
@@ -1,9 +1,29 @@
 
 # MultiHazard <img src="https://user-images.githubusercontent.com/15319503/195926656-9d3d37b5-86ab-4d4b-9e6d-3c70d5399c73.png" align="right" height="200"/>
 
+## Overview
+
 The `MultiHazard` package provides tools for stationary multivariate
 statistical modeling, for example, to estimate joint occurrence
-probabilities of MULTIple co-occurring HAZARDs. The package contains
+probabilities of MULTIple co-occurring HAZARDs. 
+
+<br>
+
+## Installation
+
+Install the latest version of this package by entering the following in
+R:
+
+``` r
+install.packages("remotes")
+remotes::install_github("rjaneUCF/MultiHazard")
+```
+
+<br>
+
+### Features:
+
+The package contains
 functions for pre-processing data, including imputing missing values,
 detrending and declustering time series as well as analyzing pairwise
 correlations over a range of lags. Functionality is built in to
@@ -32,6 +52,8 @@ as outlined in (Report).
 
 ### Citation:
 
+If you use this package, please cite:
+
 > Jane, R., Wahl, T., Peña, F., Obeysekera, J., Murphy-Barltrop, C.,
 > Ali, J., Maduwantha, P., Li, H., and Malagón Santos, V. (under review)
 > MultiHazard: Copula-based Joint Probability Analysis in R. Journal of
@@ -39,7 +61,25 @@ as outlined in (Report).
 
 <br>
 
+## Community guidelines
+
+Contributions to the `MultiHazard` package are welcome! Please feel free
+to submit issues or pull requests on GitHub.
+
+<br>
+
 ### Applications of package:
+
+This package has been used in several peer-reviewed publications:
+
+Li et al. (2025) - Assessing spatial correlation of compound flooding in the US. <https://doi.org/10.5194/egusphere-2025-2993>
+Amorim et al. (2025) - Process-driven modeling of compound rainfall and storm surge in coastal Texas. Link
+Maduwantha et al. (2024) - Multivariate statistical framework for mixed storm types. Link
+Nasr et al. (2023) - Temporal changes in compound coastal and inland flooding drivers. Link
+Kim et al. (2023a) - High-resolution probabilistic design events for rainfall and storm surge. Link
+Kim et al. (2023b) - Compounding of nitrate loads and discharge. Link
+Peña et al. (2023) - Compound flooding in low elevation coastal karst environments. Link
+Jane et al. (2020) - Multivariate statistical modeling of compound flood events in South Florida. Link
 
 > Li, H., Jane, R. A., Eilander, D., Enríquez, A. R., Haer, T., and
 > Ward, P. J. (2025). Assessing the spatial correlation of potential
@@ -83,1794 +123,3 @@ as outlined in (Report).
 > Multivariate statistical modelling of the drivers of compound flood
 > events in South Florida, Nat. Hazards Earth Syst. Sci., 20, 2681–2699,
 > <https://doi.org/10.5194/nhess-20-2681-2020>.
-
-<br>
-
-## Community guidelines
-
-Contributions to the `MultiHazard` package are welcome! Please feel free
-to submit issues or pull requests on GitHub.
-
-## Installation
-
-Install the latest version of this package by entering the following in
-R:
-
-``` r
-install.packages("remotes")
-remotes::install_github("rjaneUCF/MultiHazard")
-```
-
-## 1. Introduction
-
-The `MultiHazard` package provides tools for stationary multivariate
[…]
