+++ NAMESPACE
@@ -35,6 +35,7 @@ export(Mean_Excess_Plot)
 export(Migpd_Fit)
 export(NOAA_SLR)
 export(OsWL_Intensity)
+export(SLR_Scenarios)
 export(Standard_Copula_Fit)
 export(Standard_Copula_Sel)
 export(Standard_Copula_Sim)
@@ -68,3 +69,59 @@ import(statmod)
 import(texmex)
 import(truncnorm)
 import(tweedie)
+importFrom(VGAM,dbisa)
+importFrom(VGAM,pbisa)
+importFrom(VGAM,vglm)
+importFrom(VineCopula,BiCopEst)
+importFrom(VineCopula,BiCopSim)
+importFrom(VineCopula,RVineMatrix)
+importFrom(VineCopula,RVineSim)
+importFrom(grDevices,contourLines)
+importFrom(grDevices,heat.colors)
+importFrom(graphics,abline)
+importFrom(graphics,axis)
+importFrom(graphics,hist)
+importFrom(graphics,layout)
+importFrom(graphics,legend)
+importFrom(graphics,lines)
+importFrom(graphics,mtext)
+importFrom(graphics,par)
+importFrom(graphics,points)
+importFrom(graphics,polygon)
+importFrom(graphics,rect)
+importFrom(graphics,segments)
+importFrom(graphics,text)
+importFrom(lubridate,year)
+importFrom(stats,approx)
+importFrom(stats,bw.nrd)
+importFrom(stats,cor)
+importFrom(stats,cor.test)
+importFrom(stats,dexp)
+importFrom(stats,dgamma)
+importFrom(stats,dlnorm)
+importFrom(stats,dlogis)
+importFrom(stats,dnorm)
+importFrom(stats,dweibull)
+importFrom(stats,ecdf)
+importFrom(stats,lm)
+importFrom(stats,na.omit)
+importFrom(stats,optim)
+importFrom(stats,pexp)
+importFrom(stats,pgamma)
+importFrom(stats,plnorm)
+importFrom(stats,plogis)
+importFrom(stats,pnorm)
+importFrom(stats,pweibull)
+importFrom(stats,qexp)
+importFrom(stats,qgamma)
+importFrom(stats,qlnorm)
+importFrom(stats,qlogis)
+importFrom(stats,qnorm)
+importFrom(stats,quantile)
+importFrom(stats,qweibull)
+importFrom(stats,rexp)
+importFrom(stats,rnorm)
+importFrom(stats,runif)
+importFrom(stats,spline)
+importFrom(stats,uniroot)
+importFrom(utils,capture.output)
