+++ R/imports.R
@@ -1,0 +1,9 @@
+#' @keywords internal
+#' @importFrom grDevices contourLines heat.colors
+#' @importFrom graphics abline axis hist layout legend lines mtext par points polygon rect segments text
+#' @importFrom stats approx bw.nrd cor cor.test dexp dgamma dlnorm dlogis dnorm dweibull ecdf lm na.omit optim pexp pgamma plnorm plogis pnorm pweibull qexp qgamma qlnorm qlogis qnorm quantile qweibull rexp rnorm runif spline uniroot
+#' @importFrom utils capture.output
+#' @importFrom VGAM vglm dbisa pbisa
+#' @importFrom lubridate year
+#' @importFrom VineCopula RVineMatrix BiCopEst BiCopSim RVineSim
+"_PACKAGE"
