+++ NAMESPACE
@@ -53,16 +53,12 @@ export(return_curve_est)
 import(GeneralizedHyperbolic)
 import(MASS)
 import(RColorBrewer)
-import(VGAM)
-import(VineCopula)
-import(actuar)
 import(copula)
 import(dplyr)
 import(fitdistrplus)
 import(gamlss)
 import(gamlss.mx)
 import(ks)
-import(lubridate)
 import(rmatio)
 import(scales)
 import(statmod)
