### 2025-09-29 7dcc925
Delete R/actuar.R
