+++ R/actuar.R
@@ -1,2 +1,0 @@
-#' @import actuar
-NULL
