+++ R/lubridate.R
@@ -1,2 +1,0 @@
-#' @import lubridate
-NULL
