### 2025-09-29 2494067
Delete R/VGAM.R
