+++ R/VGAM.R
@@ -1,2 +1,0 @@
-#' @import VGAM
-NULL
