+++ R/imports.R
@@ -6,4 +6,7 @@
 #' @importFrom VGAM vglm dbisa pbisa
 #' @importFrom lubridate year
 #' @importFrom VineCopula RVineMatrix BiCopEst BiCopSim RVineSim
+#' @importFrom gamlss gamlss
+#' @importFrom gamlss.mx gamlssMX dMX pMX
+#' @importFrom gamlss.dist GG GA dGG pGG dGA pGA
 "_PACKAGE"
