+++ NAMESPACE
@@ -56,8 +56,6 @@ import(RColorBrewer)
 import(copula)
 import(dplyr)
 import(fitdistrplus)
-import(gamlss)
-import(gamlss.mx)
 import(ks)
 import(rmatio)
 import(scales)
@@ -72,6 +70,16 @@ importFrom(VineCopula,BiCopEst)
 importFrom(VineCopula,BiCopSim)
 importFrom(VineCopula,RVineMatrix)
 importFrom(VineCopula,RVineSim)
+importFrom(gamlss,gamlss)
+importFrom(gamlss.dist,GA)
+importFrom(gamlss.dist,GG)
+importFrom(gamlss.dist,dGA)
+importFrom(gamlss.dist,dGG)
+importFrom(gamlss.dist,pGA)
+importFrom(gamlss.dist,pGG)
+importFrom(gamlss.mx,dMX)
+importFrom(gamlss.mx,gamlssMX)
+importFrom(gamlss.mx,pMX)
 importFrom(grDevices,contourLines)
 importFrom(grDevices,heat.colors)
 importFrom(graphics,abline)
