+++ R/gamlss.R
@@ -1,2 +1,0 @@
-#' @import gamlss
-NULL
