### 2025-09-29 1fe9f91
Delete R/gamlss.mx.R
