+++ R/grDevices.R
@@ -1,0 +1,2 @@
+#' @import grDevices
+NULL
