+++ R/graphics.R
@@ -1,0 +1,2 @@
+#' @import graphics
+NULL
