+++ R/rmatio.R
@@ -1,2 +1,0 @@
-#' @import rmatio
-NULL
