### 2025-09-29 6bb22df
Delete R/stats.R
