+++ R/stats.R
@@ -1,2 +1,0 @@
-#' @import stats
-NULL
