### 2025-09-29 863d540
Delete R/texmex.R
