### 2025-09-29 a0d60b1
Delete R/scales.R
