+++ R/scales.R
@@ -1,2 +1,0 @@
-#' @import scales
-NULL
