### 2025-09-29 1acf801
Delete R/graphics.R
