+++ R/graphics.R
@@ -1,2 +1,0 @@
-#' @import graphics
-NULL
