+++ R/MASS.R
@@ -1,2 +1,0 @@
-#' @import MASS
-NULL
