### 2025-09-29 da96f45
Delete R/copula.R
