+++ R/copula.R
@@ -1,2 +1,0 @@
-#' @import copula
-NULL
