+++ R/grDevices.R
@@ -1,2 +1,0 @@
-#' @import grDevices
-NULL
