### 2025-09-29 8890240
Delete R/akima.R
