+++ R/akima.R
@@ -1,2 +1,0 @@
-#' @import akima
-NULL
