+++ R/RColorBrewer.R
@@ -1,2 +1,0 @@
-#' @import RColorBrewer
-NULL
