+++ R/imports.R
@@ -18,8 +18,8 @@
 #' @importFrom utils capture.output
 #' @importFrom VGAM vglm dbisa pbisa
 #' @importFrom lubridate year
-#' @importFrom VineCopula RVineMatrix BiCopEst BiCopSim RVineSim
+#' @importFrom VineCopula RVineMatrix BiCopEst BiCopSim RVineSim BiCopSelect
 #' @importFrom gamlss gamlss
-#' @importFrom gamlss.mx gamlssMX dMX pMX
-#' @importFrom gamlss.dist GG GA dGG pGG dGA pGA
+#' @importFrom gamlss.mx gamlssMX dMX pMX qMX
+#' @importFrom gamlss.dist GU RU GG GA dGG pGG dGA pGA
 "_PACKAGE"
