+++ NAMESPACE
@@ -71,17 +71,25 @@ importFrom(VGAM,pbisa)
 importFrom(VGAM,vglm)
 importFrom(VineCopula,BiCopEst)
 importFrom(VineCopula,BiCopSim)
+importFrom(VineCopula,BiCopSelect)
 importFrom(VineCopula,RVineMatrix)
 importFrom(VineCopula,RVineSim)
 importFrom(gamlss,gamlss)
+importFrom(gamlss.dist,GU)
+importFrom(gamlss.dist,RU)
 importFrom(gamlss.dist,GA)
 importFrom(gamlss.dist,GG)
+importFrom(gamlss.dist,dRU)
+importFrom(gamlss.dist,dGU)
 importFrom(gamlss.dist,dGA)
 importFrom(gamlss.dist,dGG)
+importFrom(gamlss.dist,pRU)
+importFrom(gamlss.dist,pGU)
 importFrom(gamlss.dist,pGA)
 importFrom(gamlss.dist,pGG)
 importFrom(gamlss.mx,dMX)
 importFrom(gamlss.mx,gamlssMX)
 importFrom(gamlss.mx,pMX)
+importFrom(gamlss.mx,qMX)
 importFrom(lubridate,year)
 importFrom(utils,capture.output)
