+++ R/imports.R
@@ -16,10 +16,10 @@
 #' @import truncnorm
 #' @import tweedie
 #' @importFrom utils capture.output
-#' @importFrom VGAM vglm dbisa pbisa
+#' @importFrom VGAM vglm dbisa pbisa bisa
 #' @importFrom lubridate year
 #' @importFrom VineCopula RVineMatrix BiCopEst BiCopSim RVineSim BiCopSelect
 #' @importFrom gamlss gamlss
 #' @importFrom gamlss.mx gamlssMX dMX pMX qMX
-#' @importFrom gamlss.dist GU RU GG GA dGG pGG dGA pGA
+#' @importFrom gamlss.dist GU RU GG GA dGG pGG dGA pGA dGU pGU dRU pRU
 "_PACKAGE"
