+++ NAMESPACE
@@ -68,6 +68,7 @@ import(truncnorm)
 import(tweedie)
 importFrom(VGAM,dbisa)
 importFrom(VGAM,pbisa)
+importFrom(VGAM,bisa)
 importFrom(VGAM,vglm)
 importFrom(VineCopula,BiCopEst)
 importFrom(VineCopula,BiCopSim)
