+++ NAMESPACE
@@ -77,14 +77,11 @@ importFrom(VineCopula,RVineMatrix)
 importFrom(VineCopula,RVineSim)
 importFrom(gamlss,gamlss)
 importFrom(gamlss.dist,GU)
-importFrom(gamlss.dist,RU)
 importFrom(gamlss.dist,GA)
 importFrom(gamlss.dist,GG)
-importFrom(gamlss.dist,dRU)
 importFrom(gamlss.dist,dGU)
 importFrom(gamlss.dist,dGA)
 importFrom(gamlss.dist,dGG)
-importFrom(gamlss.dist,pRU)
 importFrom(gamlss.dist,pGU)
 importFrom(gamlss.dist,pGA)
 importFrom(gamlss.dist,pGG)
