+++ R/imports.R
@@ -20,6 +20,6 @@
 #' @importFrom lubridate year
 #' @importFrom VineCopula RVineMatrix BiCopEst BiCopSim RVineSim BiCopSelect
 #' @importFrom gamlss gamlss
-#' @importFrom gamlss.mx gamlssMX dMX pMX qMX
+#' @importFrom gamlss.mx gamlssMX dMX pMX
 #' @importFrom gamlss.dist GU GG GA dGG pGG dGA pGA dGU pGU
 "_PACKAGE"
