+++ NAMESPACE
@@ -88,6 +88,5 @@ importFrom(gamlss.dist,pGG)
 importFrom(gamlss.mx,dMX)
 importFrom(gamlss.mx,gamlssMX)
 importFrom(gamlss.mx,pMX)
-importFrom(gamlss.mx,qMX)
 importFrom(lubridate,year)
 importFrom(utils,capture.output)
