+++ NAMESPACE
@@ -68,13 +68,20 @@ import(truncnorm)
 import(tweedie)
 importFrom(VGAM,dbisa)
 importFrom(VGAM,pbisa)
+importFrom(VGAM,qbisa)
 importFrom(VGAM,bisa)
 importFrom(VGAM,vglm)
+importFrom(VGAM,dlaplace)
+importFrom(VGAM,plaplace)
+importFrom(VGAM,qlaplace)
 importFrom(VineCopula,BiCopEst)
+importFrom(VineCopula,BiCopCDF)
 importFrom(VineCopula,BiCopSim)
 importFrom(VineCopula,BiCopSelect)
 importFrom(VineCopula,RVineMatrix)
 importFrom(VineCopula,RVineSim)
+importFrom(VineCopula,RVineCopSelect)
+importFrom(VineCopula,RVineStructureSelect)
 importFrom(gamlss,gamlss)
 importFrom(gamlss.dist,GU)
 importFrom(gamlss.dist,GA)
@@ -85,8 +92,15 @@ importFrom(gamlss.dist,dGG)
 importFrom(gamlss.dist,pGU)
 importFrom(gamlss.dist,pGA)
 importFrom(gamlss.dist,pGG)
+importFrom(gamlss.dist,dRG)
+importFrom(gamlss.dist,pRG)
+importFrom(gamlss.dist,qRG)
+importFrom(gamlss.dist,qGA)
+importFrom(gamlss.dist,qGU)
+importFrom(gamlss.dist,qGG)
 importFrom(gamlss.mx,dMX)
 importFrom(gamlss.mx,gamlssMX)
 importFrom(gamlss.mx,pMX)
 importFrom(lubridate,year)
+importFrom(lubridate,month)
 importFrom(utils,capture.output)
