+++ NAMESPACE
@@ -66,6 +66,7 @@ import(stats)
 import(texmex)
 import(truncnorm)
 import(tweedie)
+importFrom(actuar,qMX)
 importFrom(VGAM,dbisa)
 importFrom(VGAM,pbisa)
 importFrom(VGAM,qbisa)
@@ -74,6 +75,7 @@ importFrom(VGAM,vglm)
 importFrom(VGAM,dlaplace)
 importFrom(VGAM,plaplace)
 importFrom(VGAM,qlaplace)
+importFrom(VGAM,Coef)
 importFrom(VineCopula,BiCopEst)
 importFrom(VineCopula,BiCopCDF)
 importFrom(VineCopula,BiCopSim)
@@ -86,6 +88,7 @@ importFrom(gamlss,gamlss)
 importFrom(gamlss.dist,GU)
 importFrom(gamlss.dist,GA)
 importFrom(gamlss.dist,GG)
+importFrom(gamlss.dist,RG)
 importFrom(gamlss.dist,dGU)
 importFrom(gamlss.dist,dGA)
 importFrom(gamlss.dist,dGG)
