+++ R/imports.R
@@ -16,11 +16,10 @@
 #' @import truncnorm
 #' @import tweedie
 #' @importFrom utils capture.output
-#' @importFrom actuar qMx
 #' @importFrom VGAM Coef vglm dbisa pbisa qbisa bisa dlaplace plaplace qlaplace
 #' @importFrom lubridate year month
 #' @importFrom VineCopula RVineMatrix BiCopEst BiCopSim RVineSim BiCopSelect BiCopCDF
 #' @importFrom gamlss gamlss
-#' @importFrom gamlss.mx gamlssMX dMX pMX
+#' @importFrom gamlss.mx gamlssMX dMX pMX qMx
 #' @importFrom gamlss.dist GU GG GA RG dGG pGG dGA pGA dGU pGU dRG qRG pRG qGU qGG qGA
 "_PACKAGE"
