+++ NAMESPACE
@@ -66,7 +66,6 @@ import(stats)
 import(texmex)
 import(truncnorm)
 import(tweedie)
-importFrom(actuar,qMX)
 importFrom(VGAM,dbisa)
 importFrom(VGAM,pbisa)
 importFrom(VGAM,qbisa)
@@ -102,6 +101,7 @@ importFrom(gamlss.dist,qGA)
 importFrom(gamlss.dist,qGU)
 importFrom(gamlss.dist,qGG)
 importFrom(gamlss.mx,dMX)
+importFrom(gamlss.mx,qMX)
 importFrom(gamlss.mx,gamlssMX)
 importFrom(gamlss.mx,pMX)
 importFrom(lubridate,year)
