+++ R/imports.R
@@ -20,6 +20,6 @@
 #' @importFrom lubridate year month
 #' @importFrom VineCopula RVineMatrix BiCopEst BiCopSim RVineSim BiCopSelect BiCopCDF
 #' @importFrom gamlss gamlss
-#' @importFrom gamlss.mx gamlssMX dMX pMX qMx
+#' @importFrom gamlss.mx gamlssMX dMX pMX qMX
 #' @importFrom gamlss.dist GU GG GA RG dGG pGG dGA pGA dGU pGU dRG qRG pRG qGU qGG qGA
 "_PACKAGE"
