+++ NAMESPACE
@@ -101,7 +101,6 @@ importFrom(gamlss.dist,qGA)
 importFrom(gamlss.dist,qGU)
 importFrom(gamlss.dist,qGG)
 importFrom(gamlss.mx,dMX)
-importFrom(gamlss.mx,qMX)
 importFrom(gamlss.mx,gamlssMX)
 importFrom(gamlss.mx,pMX)
 importFrom(lubridate,year)
