+++ R/Joint_RP_2D.R
@@ -182,20 +182,24 @@ Joint_RP_2D<-function (Data, Data_Con1, Data_Con2, u1, u2,
     if(Marginal_Dist1 == "Gam(3)"){
       RP_Var2_con1<-1/(1-pGG(Var2, mu=exp(marginal_non_con1$mu.coefficients), sigma=exp(marginal_non_con1$sigma.coefficients), nu=marginal_non_con1$nu.coefficients))
     }
-    if(Marginal_Dist1 == "GamMix(2)"){
+    if(Marginal_Dist1=="GamMix(2)"){
+      xx <- seq(0, max(Data_Con2[,2])*10, 0.001)
       prob.MX1 <- round(marginal_non_con1$prob[1],3)
       prob.MX2 <- 1 - prob.MX1
-      RP_Var2_con1<-1/(1-pMX(Var2, mu=list(mu1=exp(marginal_non_con1$models[[1]]$mu.coefficients), mu2=exp(marginal_non_con1$models[[2]]$mu.coefficients)),
-                             sigma=list(sigma1=exp(marginal_non_con1$models[[1]]$sigma.coefficients), sigma2=exp(marginal_non_con1$models[[2]]$sigma.coefficients)),
-                             pi = list(pi1=prob.MX1, pi2=prob.MX2), family=list(fam1="GA", fam2="GA")))
+      cdf.MX<-pMX(xx, mu=list(mu1=exp(marginal_non_con1$models[[1]]$mu.coefficients), mu2=exp(marginal_non_con1$models[[2]]$mu.coefficients)),
+                  sigma=list(sigma1=exp(marginal_non_con1$models[[1]]$sigma.coefficients), sigma2=exp(marginal_non_con1$models[[2]]$sigma.coefficients)),
+                  pi = list(pi1=prob.MX1, pi2=prob.MX2), family=list(fam1="GA", fam2="GA"))
+      RP_Var2_con1 <- approx(cdf.MX, xx, Var2)$y
     }
-    if(Marginal_Dist1 == "GamMix(3)"){
+    if(Marginal_Dist1=="GamMix(3)"){
+      xx <- seq(0, max(Data_Con2[,2])*10, 0.001)
       prob.MX1 <- round(marginal_non_con1$prob[1],3)
       prob.MX2 <- round(marginal_non_con1$prob[2],3)
       prob.MX3 <- 1 - prob.MX1 - prob.MX2
-      RP_Var2_con1<-1/(1-pMX(Var2, mu=list(mu1=exp(marginal_non_con1$models[[1]]$mu.coefficients), mu2=exp(marginal_non_con1$models[[2]]$mu.coefficients), mu3=exp(marginal_non_con1$models[[3]]$mu.coefficients)),
-                        sigma=list(sigma1=exp(marginal_non_con1$models[[1]]$sigma.coefficients), sigma2=exp(marginal_non_con1$models[[2]]$sigma.coefficients), sigma3=exp(marginal_non_con1$models[[3]]$sigma.coefficients)),
-                        pi = list(pi1=prob.MX1, pi2=prob.MX2, pi3=prob.MX3), family=list(fam1="GA", fam2="GA", fam3="GA")))
+      cdf.MX<-pMX(xx, mu=list(mu1=exp(marginal_non_con1$models[[1]]$mu.coefficients), mu2=exp(marginal_non_con1$models[[2]]$mu.coefficients), mu3=exp(marginal_non_con1$models[[3]]$mu.coefficients)),
[…]

[…]
