+++ NAMESPACE
@@ -66,43 +66,41 @@ import(stats)
 import(texmex)
 import(truncnorm)
 import(tweedie)
-importFrom(VGAM,dbisa)
-importFrom(VGAM,pbisa)
-importFrom(VGAM,qbisa)
-importFrom(VGAM,bisa)
-importFrom(VGAM,vglm)
-importFrom(VGAM,dlaplace)
-importFrom(VGAM,plaplace)
-importFrom(VGAM,qlaplace)
 importFrom(VGAM,Coef)
-importFrom(VineCopula,BiCopEst)
+importFrom(VGAM,bisa)
+importFrom(VGAM,dbisa)
+importFrom(VGAM,dlaplace)
+importFrom(VGAM,pbisa)
+importFrom(VGAM,plaplace)
+importFrom(VGAM,qbisa)
+importFrom(VGAM,qlaplace)
+importFrom(VGAM,vglm)
 importFrom(VineCopula,BiCopCDF)
-importFrom(VineCopula,BiCopSim)
+importFrom(VineCopula,BiCopEst)
 importFrom(VineCopula,BiCopSelect)
+importFrom(VineCopula,BiCopSim)
 importFrom(VineCopula,RVineMatrix)
 importFrom(VineCopula,RVineSim)
-importFrom(VineCopula,RVineCopSelect)
-importFrom(VineCopula,RVineStructureSelect)
 importFrom(gamlss,gamlss)
-importFrom(gamlss.dist,GU)
 importFrom(gamlss.dist,GA)
 importFrom(gamlss.dist,GG)
+importFrom(gamlss.dist,GU)
 importFrom(gamlss.dist,RG)
-importFrom(gamlss.dist,dGU)
 importFrom(gamlss.dist,dGA)
 importFrom(gamlss.dist,dGG)
-importFrom(gamlss.dist,pGU)
+importFrom(gamlss.dist,dGU)
+importFrom(gamlss.dist,dRG)
 importFrom(gamlss.dist,pGA)
 importFrom(gamlss.dist,pGG)
-importFrom(gamlss.dist,dRG)
+importFrom(gamlss.dist,pGU)
 importFrom(gamlss.dist,pRG)
-importFrom(gamlss.dist,qRG)
 importFrom(gamlss.dist,qGA)
-importFrom(gamlss.dist,qGU)
 importFrom(gamlss.dist,qGG)
+importFrom(gamlss.dist,qGU)
+importFrom(gamlss.dist,qRG)
 importFrom(gamlss.mx,dMX)
 importFrom(gamlss.mx,gamlssMX)
 importFrom(gamlss.mx,pMX)
-importFrom(lubridate,year)
 importFrom(lubridate,month)
+importFrom(lubridate,year)
 importFrom(utils,capture.output)
