+++ man/MultiHazard-package.Rd
@@ -1,0 +1,34 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/imports.R
+\docType{package}
+\name{MultiHazard-package}
+\alias{MultiHazard}
+\alias{MultiHazard-package}
+\title{MultiHazard: Tools for modeling compound events}
+\description{
+The `MultiHazard` package provides tools for stationary multivariate statistical modeling, for example, to estimate the joint distribution of MULTIple co-occurring HAZARDs. The package contains functions for pre-processing data including imputing missing values, detrending and declustering time series as well as analyzing pairwise correlations over a range of lags. Functionality is also built in to impliment the conditional sampling - copula theory approach in Jane et al. (2020) including the automated threshold selection approach in Solari et al. (2017). Tools are provided for selecting the best fitting amongst an array of (non-extreme, truncated and non-truncated) parametric marginal distributions, and, copulas to model the dependence structure. The package contains a function that calculates joint probability contours using the method of overlaying (conditional) contours given in Bender et al. (2016), and extracting design events such as the 'most likely' event or an ensemble of possible design events. The package also provides the capability of fitting and simulating synthetic records from three higher dimensional approaches - standard (elliptic/Archimedean) copulas, Pair Copula Constructions (PCCs) and the conditional threshold exceedance approach of Heffernan and Tawn (2004). Finally, a function that calculates the time for a user-specified height of sea level rise to occur under various scenarios is supplied.
+}
+\seealso{
+Useful links:
+\itemize{
+  \item \url{https://rdrr.io/github/rjaneUCF/MultiHazard-R-Package/}
+}
+
+}
+\author{
+\strong{Maintainer}: Robert Jane \email{r.jane@ucf.edu} (\href{https://orcid.org/0000-0003-4194-852X}{ORCID})
+
+Other contributors:
+\itemize{
+  \item Thomas Wahl \email{t.wahl@ucf.edu} (\href{https://orcid.org/0000-0003-3643-5463}{ORCID}) [contributor]
+  \item Francisco Pena \email{fpena@sfwmd.gov} (\href{https://orcid.org/0000-0002-1779-9926}{ORCID}) [contributor]
+  \item Jayantha Obeysekera \email{jobeysek@fiu.edu} (\href{https://orcid.org/0000-0002-6203-1190}{ORCID}) [contributor]
+  \item Callum Murphy-Barltrop \email{callum.murphy-barltrop@tu-dresden.de} (\href{https://orcid.org/0000-0002-3479-2902}{ORCID}) [contributor]
+  \item Javed Ali \email{javed.ali@ucf.edu} (\href{https://orcid.org/0000-0002-6203-1190}{ORCID}) [contributor]
+  \item Pravin Maduwantha \email{pravin@ucf.edu} (\href{https://orcid.org/0000-0002-9819-5170}{ORCID}) [contributor]
+  \item Huazhi Li \email{huazhi.li@vu.nl} (\href{https://orcid.org/0000-0001-9589-2918}{ORCID}) [contributor]
+  \item Victor Malagon Santos \email{victor.malagon.santos@nioz.nlu} (\href{https://orcid.org/0000-0002-4458-0890}{ORCID}) [contributor]
+}
+
+}
+\keyword{internal}
