+++ DESCRIPTION
@@ -48,6 +48,30 @@ URL: https://rdrr.io/github/rjaneUCF/MultiHazard-R-Package/
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.3
+Imports: 
+    copula,
+    dplyr,
+    fitdistrplus,
+    gamlss,
+    gamlss.dist,
+    gamlss.mx,
+    GeneralizedHyperbolic,
+    graphics,
+    grDevices,
+    ks,
+    lubridate,
+    MASS,
+    RColorBrewer,
+    rmatio,
+    scales,
+    statmod,
+    stats,
+    texmex,
+    truncnorm,
+    tweedie,
+    utils,
+    VGAM,
+    VineCopula
 Suggests: 
     knitr,
     rmarkdown,
