+++ NAMESPACE
@@ -79,8 +79,10 @@ importFrom(VineCopula,BiCopCDF)
 importFrom(VineCopula,BiCopEst)
 importFrom(VineCopula,BiCopSelect)
 importFrom(VineCopula,BiCopSim)
+importFrom(VineCopula,RVineCopSelect)
 importFrom(VineCopula,RVineMatrix)
 importFrom(VineCopula,RVineSim)
+importFrom(VineCopula,RVineStructureSelect)
 importFrom(gamlss,gamlss)
 importFrom(gamlss.dist,GA)
 importFrom(gamlss.dist,GG)
