+++ R/Detrend.R
@@ -15,9 +15,9 @@
 #' @return Numeric vector of the detrended time series.
 #' @export
 #' @examples
-#' #Detrending ocean-side water level at site S22 using a 3 month moving average window and the last
+#' #Detrending ground water level at Well G-3355 using a 3 month moving average window and the last
 #' #five years of observations to calculate the present day average.
-#' Detrend(S22_T_MAX_Daily_Completed[,1:2],Method = "window",Window_Width= 89,
+#' Detrend(G_3355[,1:2],Method = "window",Window_Width= 89,
 #'         End_Length = 1826, PLOT=FALSE,x_lab="Data",y_lab="Data")
 Detrend<-function(Data, Method = "window",Window_Width= 89, End_Length = 1826, PLOT=FALSE,x_lab="Date",y_lab="Data"){
   data_Detrend<-Data[,2]
