+++ NAMESPACE
@@ -83,6 +83,16 @@ importFrom(VineCopula,RVineCopSelect)
 importFrom(VineCopula,RVineMatrix)
 importFrom(VineCopula,RVineSim)
 importFrom(VineCopula,RVineStructureSelect)
+importFrom(copula,claytonCopula)
+importFrom(copula,dCopula)
+importFrom(copula,fitCopula)
+importFrom(copula,frankCopula)
+importFrom(copula,gumbelCopula)
+importFrom(copula,normalCopula)
+importFrom(copula,pCopula)
+importFrom(copula,pobs)
+importFrom(copula,rCopula)
+importFrom(copula,tCopula)
 importFrom(gamlss,gamlss)
 importFrom(gamlss.dist,GA)
 importFrom(gamlss.dist,GG)
