+++ NAMESPACE
@@ -93,6 +93,7 @@ importFrom(copula,pCopula)
 importFrom(copula,pobs)
 importFrom(copula,rCopula)
 importFrom(copula,tCopula)
+importFrom(dplyr,left_join)
 importFrom(gamlss,gamlss)
 importFrom(gamlss.dist,GA)
 importFrom(gamlss.dist,GG)
