+++ NAMESPACE
@@ -93,6 +93,7 @@ importFrom(copula,pCopula)
 importFrom(copula,pobs)
 importFrom(copula,rCopula)
 importFrom(copula,tCopula)
+importFrom(dplyr,full_join)
 importFrom(dplyr,left_join)
 importFrom(gamlss,gamlss)
 importFrom(gamlss.dist,GA)
