+++ R/imports.R
@@ -13,7 +13,7 @@
 #' @import texmex
 #' @import truncnorm
 #' @import tweedie
-#' @importFrom dplyr left_join
+#' @importFrom dplyr left_join full_join
 #' @importFrom copula pobs fitCopula normalCopula tCopula gumbelCopula claytonCopula frankCopula dCopula pCopula rCopula
 #' @importFrom utils capture.output
 #' @importFrom VGAM Coef vglm dbisa pbisa qbisa bisa dlaplace plaplace qlaplace
