+++ vignettes/MultiHazard.Rmd
@@ -26,30 +26,6 @@ The `MultiHazard` package provides tools for stationary multivariate statistical
 
 <br>
 
-### Applications of package: 
-
->Li, H., Jane, R. A., Eilander, D., Enríquez, A. R., Haer, T., and Ward, P. J. (2025). Assessing the spatial correlation of potential compound flooding in the United States, EGUsphere [preprint], https://doi.org/10.5194/egusphere-2025-2993.
-
->Amorim, R., Villarini, G., Kim, H., Jane, R., and Wahl, T. (2025). A Practitioner’s approach to process-driven modeling of compound rainfall and storm surge extremes for coastal Texas, J. Hydrol. Eng., 30(5), 04025025, https://doi.org/10.1061/JHYEFF.HEENG-648.
-
->Maduwantha, P., Wahl, T., Santamaria-Aguilar, S., Jane, R., Booth, J. F., Kim, H., and Villarini, G. (2024). A multivariate statistical framework for mixed storm types in compound flood analysis, Nat. Hazards Earth Syst. Sci., 24, 4091–4107, https://doi.org/10.5194/nhess-24-4091-2024.
-
->Nasr, A. A., Wahl, T., Rashid, M. M., Jane, R., Camus, P. and Haigh, I. D. (2023). Temporal changes in dependence between compound coastal and inland flooding drivers around the contiguous United States coastline, Weather Clim. Extrem., 41, 100594, https://doi.org/10.1016/j.wace.2023.100594.
-
->Kim, H., Villarini, G., Jane, R., Wahl, T., Misra, S., and Michalek, A. (2023). On the generation of high‐resolution probabilistic design events capturing the joint occurrence of rainfall and storm surge in coastal basins, Int. J. Climatol, 43(2), 761-771, https://doi.org/10.1002/joc.7825.
-
->Kim, T., Villarini, G., Kim, H., Jane, R., and Wahl, T. (2023). On the compounding of nitrate loads and discharge, J. Environ. Qual., 52, 706–717. https://doi.org/10.1002/jeq2.20458.
-
->Peña, F., Obeysekera, J., Jane R., Nardi, F., Maran, C., Cadogan, A., de Groen, F., and Melesse, A. (2023). Investigating compound flooding in a low elevation coastal karst environment using multivariate statistical and 2D hydrodynamic modeling, Weather Clim. Extrem., 39, 100534. https://doi.org/10.1016/j.wace.2022.100534.
-
->Jane, R., Cadavid, L., Obeysekera, J., and Wahl, T. (2020). Multivariate statistical modelling of the drivers of compound flood events in South Florida, Nat. Hazards Earth Syst. Sci., 20, 2681–2699, https://doi.org/10.5194/nhess-20-2681-2020.
-
-<br>
-
-## Community guidelines
-
-Contributions to the `MultiHazard` package are welcome! Please feel free to submit issues or pull requests on GitHub.
-
 ## Installation
 Install the latest version of this package by entering the following in R:
 
@@ -565,7 +541,7 @@ curve$ensemble_wt13
 
[…]
