+++ doc/MultiHazard.R
@@ -1,0 +1,639 @@
+## ----setup, include=FALSE-----------------------------------------------------
+knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, collapse = TRUE,
+  comment = "#>")
+library(MultiHazard)
+library(dplyr)
+library(scales)
+library(copula)
+
+## ----eval=FALSE---------------------------------------------------------------
+# install.packages("remotes")
+# remotes::install_github("rjaneUCF/MultiHazard")
+
+## -----------------------------------------------------------------------------
+#Viewing first few rows of in the groundwater level records
+head(G_3356)
+
+## -----------------------------------------------------------------------------
+head(G_3355)
+
+## -----------------------------------------------------------------------------
+#Converting Date column to "Date"" object
+G_3356$Date<-seq(as.Date("1985-10-23"), as.Date("2019-05-29"), by="day")
+G_3355$Date<-seq(as.Date("1985-08-20"), as.Date("2019-06-02"), by="day")
+#Converting column containing the readings to a "numeric"" object
+G_3356$Value<-as.numeric(as.character(G_3356$Value))
+G_3355$Value<-as.numeric(as.character(G_3355$Value))
+
+## -----------------------------------------------------------------------------
+#Merge the two dataframes by date
+GW_S20<-left_join(G_3356,G_3355,by="Date")
+colnames(GW_S20)<-c("Date","G3356","G3355")
+#Carrying out imputation
+Imp<-Imputation(Data=GW_S20,Variable="G3356",
+                x_lab="G-3355 (ft NGVD 29)", y_lab="G-3356 (ft NGVD 29)")
+
+## -----------------------------------------------------------------------------
+head(Imp$Data)
+
+## -----------------------------------------------------------------------------
+Imp$Model
+
+## -----------------------------------------------------------------------------
+G_3356_ValueFilled_NA<-which(is.na(Imp$Data$ValuesFilled)==TRUE)
+length(G_3356_ValueFilled_NA) 
+
+## -----------------------------------------------------------------------------
+G3356_approx<-approx(seq(1,length(Imp$Data$ValuesFilled),1),Imp$Data$ValuesFilled,
+                     xout=seq(1,length(Imp$Data$ValuesFilled),1))
+Imp$Data$ValuesFilled[which(is.na(Imp$Data$ValuesFilled)==TRUE)]<-
+  G3356_approx$y[which(is.na(Imp$Data$ValuesFilled)==TRUE)]
+
+## -----------------------------------------------------------------------------
+#Creating a data frame with the imputed series alongside the corresponding dates 
+G_3356_Imp<-data.frame(Imp$Data$Date,Imp$Data$ValuesFilled)
+colnames(G_3356_Imp)<-c("Date","ValuesFilled")
+#Detrending
+G_3356_Detrend<-Detrend(Data=G_3356_Imp,PLOT=TRUE,x_lab="Date",
+                        y_lab="Groundwater level (ft NGVD 29)")
+
+## -----------------------------------------------------------------------------
+head(G_3356_Detrend)
+
+## -----------------------------------------------------------------------------
+S20.Groundwater.Detrend.df<-data.frame(as.Date(GW_S20$Date),G_3356_Detrend)
+colnames(S20.Groundwater.Detrend.df)<-c("Date","Groundwater")
+
+## -----------------------------------------------------------------------------
+G_3356.Declustered<-Decluster(Data=G_3356_Detrend,u=0.95,SepCrit=3,mu=365.25)
+
[…]
