+++ man/Con_Sampling_2D.Rd
@@ -7,9 +7,9 @@
 Con_Sampling_2D(Data_Detrend, Data_Declust, Con_Variable, u = 0.97, Thres = NA)
 }
 \arguments{
-\item{Data_Detrend}{Data frame containing two at least partially concurrent time series, detrended if necessary. Time steps must be equally spaced, with missing values assigned \code{NA}. First object may be a \code{"Date"} object. Can be \code{Dataframe_Combine} output.}
+\item{Data_Detrend}{Data frame containing two at least partially concurrent time series, detrended if necessary. Time steps must be equally spaced, with missing values assigned \code{NA}. First column may be a \code{"Date"} object. Can be \code{Dataframe_Combine} output.}
 
-\item{Data_Declust}{Data frame containing two (independently) declustered at least partially concurrent time series. Time steps must be equally spaced, with missing values assigned \code{NA}. Columns must be in the same order as in \code{Data_Detrend}. First object may be a \code{"Date"} object. Can be \code{Dataframe_Combine} output.}
+\item{Data_Declust}{Data frame containing two (independently) declustered at least partially concurrent time series. Time steps must be equally spaced, with missing values assigned \code{NA}. Columns must be in the same order as in \code{Data_Detrend}. First column may be a \code{"Date"} object. Can be \code{Dataframe_Combine} output.}
 
 \item{Con_Variable}{Column number (1 or 2) or the column name of the conditioning variable. Default is \code{1}.}
 
@@ -26,5 +26,5 @@ Creates a data frame where the declustered excesses of a (conditioning) variable
 \examples{
 S20.Rainfall<-Con_Sampling_2D(Data_Detrend=S20.Detrend.df[,-c(1,4)],
                               Data_Declust=S20.Detrend.Declustered.df[,-c(1,4)],
-                              Con_Variable="Rainfall",Thres=0.97)
+                              Con_Variable="Rainfall",u=0.97)
 }
