### 2025-12-17 #8350 @lrnv
@rjaneUCF 3 hours is still very very large for such testing routines. I suggest you try to decipher what part of this timings is really necessary and reduce it drastically -- if possible. If you want me to take a look at the tests themselves to help you refactor them, please ping me :)
