### 2026-01-07 #8350 @rjaneUCF
Hey @lrnv, Thank you very much for the suggestions. These changes along with a few others have reduced the total test time down to around 1 hr and 22 mins, I've been reading that there is an option to skip some of the tests on CRAN, perhaps we can do that for a the 4 or 5 functions that take the majority of the time, Best, Rob.

✔ | F W  S  OK | Context
✔ |         15 | Annual_Max [3.7s]
✔ |         15 | Con_Sampling_2D
✔ |         20 | Con_Sampling_2D_Lag
✔ |         22 | Cooley19 [12.2s]
✔ |         19 | Copula_Threshold_2D [35.8s]
✔ |         24 | Copula_Threshold_2D_Lag [32.7s]
✔ |         21 | Decluster
✔ |         20 | Decluster_S_SW [59.2s]
✔ |         19 | Decluster_SW
✔ |         19 | Design_Event_2D [25.4s]
✔ |         15 | Design_Event_2D_Grid [21.6s]
✔ |         15 | Design_Event_2D_Multi_Pop [85.2s]
✔ |         13 | Detrend [6.2s]
✔ |         18 | Diag_Non_Con [1.1s]
✔ |         14 | Diag_Non_Con_Sel [1.3s]
✔ |         17 | Diag_Non_Con_Trunc [820.7s]
✔ |         15 | Diag_Non_Con_Trunc_Sel [2185.4s]
✔ |         13 | GPD_Fit
✔ |         17 | GPD_Threshold_Solari [3.4s]
✔ |          8 | GPD_Threshold_Solari_Sel [11.6s]
✔ |         12 | HT04 [22.7s]
✔ |         11 | HT04_Lag [39.9s]
✔ |         17 | Imputation [2.2s]
✔ |         14 | Intensity [7.2s]
✔ |         11 | return_curve_diag [694.6s]
✔ |         13 | return_curve_est [472.2s]
✔ |          8 | SLR_Scenarios
✔ |         26 | Standard_Copula_Fit [34.8s]
✔ |          8 | Standard_Copula_Sel [5.1s]
✔ |          9 | Standard_Copula_Sim [30.3s]
✔ |         13 | U_Sample [12.8s]
✔ |         13 | Vine_Copula_Fit [185.8s]
✔ |         12 | Vine_Copula_Sim [88.4s]
✔ |         12 | WL_Curve [3.7s]
