### 2026-01-12 rjaneUCF/MultiHazard#2 @lrnv
Reduce size of tests data (1/2)
In the test that takes the most time, there :

https://github.com/rjaneUCF/MultiHazard/blob/e46939fdc55e16d0810be3abd64f3345840c122a/tests/testthat/test-Diag_Non_Con_Trunc_Sel.R#L1-L17

you are using a dataset `S20.OsWL`, and at a first glance i was not able to see what length this dataset has. It looks like its the same dataset in this file too:

https://github.com/rjaneUCF/MultiHazard/blob/e46939fdc55e16d0810be3abd64f3345840c122a/tests/testthat/test-Diag_Non_Con_Trunc.R#L3-L13

Where does this dataset come from // what is really its size ? I am sure we can do something to reduce it drastically while still validating what your tests need to validate.

-------

ref to https://github.com/openjournals/joss-reviews/issues/8350
