### 2026-01-12 rjaneUCF/MultiHazard#3 @lrnv
Reduce size of test data (2/2)
In this file:

https://github.com/rjaneUCF/MultiHazard/blob/e46939fdc55e16d0810be3abd64f3345840c122a/tests/testthat/test-return_curve_diag.R#L4-L14

the dataset size is explicitely about 26000, which is very large. Can you draft a mockup dataset that would still validate the internal structure of your code but with say 100 points and not 26000 ? I think its the same one in `return_curve_est `.

ref to https://github.com/openjournals/joss-reviews/issues/8350
