### 2026-01-13 #8350 @lrnv
That is much much better for you and for future leveraging these tests. We could go further by looking at :

✔ | 13 | Vine_Copula_Fit [191.7s]
✔ | 13 | return_curve_est [165.1s]
✔ | 11 | return_curve_diag [149.8s]
✔ | 12 | Vine_Copula_Sim [91.5s]
✔ | 15 | Design_Event_2D_Multi_Pop [85.2s]
✔ | 20 | Decluster_S_SW [61.0s]
....

but this is up to you.
