### 2026-01-14 #8350 @rjaneUCF
I notice the other papers do not contain the following lines, should I remove them?

aas-doi: "10.3847/xxxxx <- update this with the DOI from AAS once you know it."
aas-journal: "Journal of Open Source Software"
