### 2026-01-14 #8350 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
